ws: Write size as big-endian always
I was just re-reviewing this code, and while I didn't actually hit
this problem, the way we write the size as host endianness would break
having the mgmt node be x86_64 and a target be ppc.

The simple fix (and this is the same as ext{2,3,4} do) is always use
big endian.

Stef: Part of this was committed during work on #153, and Cockpit broke
because the other side in dbus-server.c had not yet been changed. This
commit brings in the remaining bits.

Closes #100

Signed-off-by: Stef Walter <stefw@redhat.com>