doc: Document updated protocol with channels and commands
Note that this changed from earlier iterations. The channel number
is not an MSB integer, but rather a decimal followed by a newline.
This was because binary data cannot be transferred over some older
websocket standards.
doc: Document updated protocol with channels and commands

Note that this changed from earlier iterations. The channel number
is not an MSB integer, but rather a decimal followed by a newline.
This was because binary data cannot be transferred over some older
websocket standards.

Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>