ws: Split out transport framing and io into CockpitTransport
An interface called CockpitTransport handles the high level
concept of sending messages to the various hosts/sessions that
we support.

CockpitFdTransport is a concrete implementation of this which
transfers data over file descriptors.

This simplifies CockpitWebsocket significantly, so that we
can expand it to handle multiple hosts/transports/channels
per WebSocket.

Although there's the concept of channels in the CockpitTransport
API, these are not implemented in this commit. All channels must
be set to zero at this point.

Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>