tap-compiler: Don't skip tests if tests crash
We got a 'TAP plan' error if a test would crash or abort without
running all the tests. Account for this in our tap-compiler script.

So here's how this works.

 1. The tap-compiler now runs its test cases with the '--tap -l --verbose'
    arguments. In the case of GTest tests, this causes it to output a list
    of the available test cases. Non-GTest tests (such as test-server) are
    expected to ignore the -l and just start testing, and producing TAP
    output on their own.

 2. If we see that it's a GTest list of test cases, then we rerun the
    test without the -l argument for actual testing. We keep track of
    which tests have run.

 3. If the tests crash half-way, we can run the remaining tests with the
    GTest -p argument, individually each in their own process.

We try to run everything in one process initially for speed, falling back
to slower -p behavior once something errors out.

Reviewd-by: Marius Vollmer <redhat.com>
Closes: #228