channel.js: Implement cockpit transport code in cockpit-web
This is a new file called channel.js the API is documented in the file
itself. The API is purposefully sparse with predictable behavior to
build off of. In particular most of the transport stuff is hidden in a
private class.

Tested in test-channel.html, which runs from 'make check' or
in a web browser.

Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>