dbus.js: Include dbus.js in cockpit.js
It will soon depend on channel.js, and we also need to include it in
all the lint goodness. Side effect of this is that it speeds up loading.

dbus.js will no longer be a generic thing, but very cockpit specific
using the channels for transport.

Fix all the various coding lint issues in dbus.js that jsl found.

Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>