dbus-server: Make dbus_server_serve_dbus not block
Run the mainloop outside of dbus_server_serve_dbus() and handle
closing of the transport from there as well.

This is a step in the direction of having the agent handle channel
commands, and eventually multiple channels.

Some of the code duplication will be removed when we use cockpit-agent
instead of test-agent.

Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>