web: Don't display Unhandled exceptions in a dialog
These go to the console, and are not useful for users to see
in a dialog. Show an 'Ooops!' in the status bar instead, and
suggest that the user restarts Cockpit, when they click on the
oops.

This method catches all of the unhandled exceptions, not just
those we were wrapping in a besafe() function.

Closes #305
Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>