web: Add .filter-unimportant and .unimportant CSS classes
These are used when you have a bunch of things (like rows in a table)
some of which are unimportant, and there's an optional filter
to hide them. .unimportant gets set on the thing (like the row)
and .filter-unimportant gets set on an ancestor, when you want
to hide the unimportant things.

Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>