web: Add a div.slider control
This will be used for allocating memory and CPU limits.

See src/web/STYLE.html for an example, or src/web/controls.js
for a bit of documentation on how to use it.

Slider can be disabled with .disabled or disabled='true' as
well and the value accessed with .value or value='xxx'

Work towards fixing issue #295
Closes #322
Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>