web: As containers are starting/stopping on overview page show waiting
Show a throbber as containers on the overview page are starting
or stopping.

We cannot use /events or container state to track the waiting state,
as docker doesn't seem to give us anything useful here. So we do this
ourselves.

The small-spinner.gif looks *really* ugly, so we use an alternate
throbber here, at least for now.

Along with previous commits this, fixes #301, closes #317

Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>