web: Consistent debugging and waiting state for docker operations
Expand the functions in DockerClient for each of the level docker
REST actions. Add create(), rm(), rmi().

Add better debugging to all the operations and pages. Enable one of
these in the javascript console:

$cockpit.debugging="docker"
document.cookie="debugging=docker"

Update the waiting state consistently for all actions on images
and containers.

Closes #361
Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>