Add support for building with travis-ci.org
This means putting in a .travis.yml file.

We use packages from the latest trusty version of ubuntu in order to
get the build to succeed, as well building jsl as part of our build.
libgsystem and storaged are also not available, so we bring in enough
for cockpit to get past 'make check'

Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>