daemon: Make Machines object testable and add basic test
Simplify the periphery and around Machines and Machine objects
and decouple them from Daemon. This allows us to start writing
tests for this stuff. In addition construct properties don't
need to be getable if that's not necessary functionality.

Add a single test of a basic 'Add' call to Machines. This is
not meant to be comprehensive, but just get us started.

Reviewed-by: Marius Vollmer <mvollmer@redhat.com>