web: Remove cockpit_dbus_client.
Instead, each page gets the clients it wants.  The pages have been
fixed to survive with a client that is not yet ready, but they might
not do that in the prettiest way possible.

When the connection is lost spontaneously, the UI does nothing special
now.  It mostly freezes and button clicks result in immediate errors.
This will be improved in a later commit.

Reviewed-by: Stef Walter <stefw@redhat.com>