ws: Serve the / HTTP path with its own handler
Since the response to / cannot be aggressively cached, unlike almost
everything else ... include dynamic content right in the response to
minimize server round trips.

Removed the unused avatar data url. If we do want an avatar on
the login page, it should be retrieved as a normal image and have
its own handler.

Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>