ws: Expose cockpit module resources via HTTP
This adds handlers under /res/ and /cache/ for loading cockpit
resources. These are served from data provided by cockpit-agent
and the resource1 channel payload.

In addition add code to load the full module listing including
manifests from a given cockpit-agent. We use this in the tests,
and in a later commit this is used to build the environment.

Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>