shell: Disable the "Next" button appropriately
Disable the "Next" button when the address or host name is empty.
When the address or host name is incorrect, the button is disabled,
but still freezed, this will be fixed together with the returned
error issue.

When the button is freezed, it will change back to the enabled
state whenever the address is altered.

The address text input will get the focus when the "address" tab
shows up.

Fixes #861
Closes #892
Reviewed-by: Stef Walter <stefw@redhat.com>