shell: Remove unused code from journal page
The search code is no longer in use and is rather complicated to
maintain going forward. The serivce filtering is not used by the UI
but is a vital part of tests: simplify it.

Add tag based filtering (ie: logger --tag) which is also used
from integration tests, and not exposed in the UI.

Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>