static: jQuery becomes part of the the packages
No jQuery on the login page is the upshot here.

We'll probably unfortunately get to the point where different versions
of component have different jQuery requirements and so on. So jQuery
gets loaded by the packages.

See later commits for a solution that prevents needless copying of
jQuery all over the place.

Closes #1315
Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>