bridge: Implement a new dbus-json3 protocol payload
This protocol is much simpler than the others. There will be some
additions to it in the future, but at it's core it's about relaying
messages and signals.

We don't have innate object manager support in the bridge, the
javascript code will respond to events and update an object model
there.

Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>