kubernetes: Delegated events for delete buttons and prevent bubbling
Because the delete buttons are nested inside other clickable elements
we have to prevent bubbling, so we use a "click" handler that returns
false.

In order to do this we have to use delegated jQuery events.

In addition stop having multiple elements with the same 'id' attribute.
We use 'data-kind' instead for this value.

Closes #1828
Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>