ws: Send ETags and respect If-None-Match header
We send ETags that match the server checksum, if one exists.

If we get a request with If-None-Match, then we check that it
matches the checksum in the URL, and return '304 Not Modified' if
that's the case.

In addition we can use the ETag header in XMLHttpRequset responses
to determine the checksum of a server, since XMLHttpRequest do
redirects transparently.

Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>