pkg: Use a requirejs loader plugin for translated resources
Finishing up the negotiation work that happened in earlier,
commits: We now do the selection of a translated resource on
the client side, rather than relying on the flakey Vary header
that would totally break caching.

So we have to choose a translated url on the client side for
modules, and we do this with a very simple requirejs loader
plugin.

Fixes #1852
Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>