common: Fix two test races that are failing in Travis
Well, at least try to fix them, that's how these things go.

In one case writev is failing with -1. I'm guessing that this
is EAGAIN or EINTR and we are now handling those.

The other case is where the close of file descriptores raced
with the closing of the stream. Put these in the right order.

Fingers crossed.

Closes #2280
Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>