bridge: Support for HTTP client certificates
This adds support for HTTP root and client certificates to the
http-stream1 payload, and rounds out the TLS support.

Certificates are either specified as files or PEM data.

PEM is the worst crypto format, except for all the others.

The PEM format for certificates and keys is non-standard, and
rather ugly. However all the plain DER based formats are even
nastier. In particular:

 * Plain DER formats are not self-describing. You have to carry
   around metadata about what kind of file format is present.
 * Not readable or debuggable at all.
 * Cannot combine multiple certificates together.

At least PEM is debuggable, readable, self-describing, and
combinable.

Closes #2297
Reviewed-by: Peter <petervo@redhat.com>