test: Extend meaning of "--clean" option.
* The "cockpit" flavored image is always created from scratch (but not
  saved).

* The development version of the selinux policy is not installed.

The idea is that we test master with "--clean" to get better
information about how the development version of Cockpit would do if
released to a distribution.

Otherwise, testing master would just repeat the tests that have been
done for the most recent pull request already.  Some races might have
differently winners, and we would catch commits on master that have
not been fully tested, but right now testing master is largely
uninteresting.

Closes #2445
Reviewed-by: Stef Walter <stefw@redhat.com>