Use symlinks and hashes to version control images
Images are simple symlinks in the test/images directories, and
are committed to git and versioned. These symlinks are relative
and point to qcow2 files in the same directory.

The actual image qcow2 data files are named after a hash of
their contents. These are the files that are not committed to
git (too big) but are uploaded to our image store.

After someone changes the setup script, they must:

 * Recreate the images with ./vm-create
 * Commit the generated symlinks
 * Use vm-upload to push the image data to our image store
 * Push the resulting commits

Closes #3037

Reviewed-by: Dominik Perpeet <dperpeet@redhat.com>