test: Manually choose mac addresses
We absolutely have to reuse mac addresses. At the rate we
run through VMs the DHCP server just gives up if each VM
shows up with a new mac address.

We select mac addresses predictably and then lock the use
of that mac address until the test process exits. This
means that mac addresses can be reused within a test run
with multiple jobs, but never within a single job.

Closes #3003
Reviewed-by: Marius Vollmer <marius.vollmer@redhat.com>