bridge: Include content-src for WebSocket in Content-Security-Policy
Note that browsers need to be explicitly told they can connect
to a WebSocket. This is non-obvious, but it stems from the fact
that some browsers treat 'https' and 'wss' as different protocols.

Since each component could establish a WebSocket connection back to
cockpit-ws, we need to insert that into the policy.

Reviewed-by: Peter <petervo@redhat.com>
 * Fixed doc typo