registry: Default activate elements by moving to their page
This behavior can be changed in the menu, but is this way for
simplicity.

When only dealing with admins, we were able to build a tool that
required them to learn how the UI worked and get used to it. When
dealing with users, for the self-service image registry UI, we
don't have that luxury. So keep things simple.

Reviewed-by: Dominik Perpeet <dperpeet@redhat.com>