test: Skip 'make check' when installing for avocado/selenium tasks
The idea of running 'make check' during our integration tests is
to have that run on various operating systems. The goal of the
avocado and selenium tests, when run during integration, is not
to do a spread of operating systems. So we can skip this step.

This change only takes effect when running under github-task

Closes #3521

Reviewed-by: Dominik Perpeet <dperpeet@redhat.com>