Test tarballs that are similar to release tarballs
Call `make dist` outside of the test VMs instead of calling autogen.sh
inside. Adjust the debian and rpm packaging to reflect that.

Tests are now running on tarballs that are more similar to release
tarballs and test packages are more similar to packages that
distributions ship.

The system running the tests must now also have build dependencies
installed.

make-source, the script that prepares the tarball, tries not to
interfere too much with existing builds. It configures the package in
tmp-dist/ without setting up the redirect Makefile. However, due to
limitations in autotools, it must use an existing in-tree configuration.

Closes #4178
Reviewed-by: Stef Walter <stefw@redhat.com>