base1: Allow require.js and bundle.js to work with data-main
A data-main attribute is a decent way to require that require.js
loads a specific module to initialize.

By defining var require before require.js we cause it to use
the configuration when parsing the data-main.

In addition we make data-main take effect after the page has
been loaded.

This all allows Content-Security-Policy and data-main to work
together.  And helps us to work well with preloaded bundles.

Reviewed-by: Peter <petervo@redhat.com>