kubernetes: Move D3 into the kubernetes package for now
D3 is only used by the kubernetes package.

While this is something we want ot have in the base1 package
in the future and use more broadly, that hasn't been the case
yet.

So as part of stabilizing the base1 package and cleaning it up
lets move this out for now. It's easy to add it later once the
work to share it is actually done.

Closes #4450
Reviewed-by: Peter <petervo@redhat.com>