shell: Add support for session machines
Session machines are machines that stored in sessionStorage.
These machines will be treated by cockpit as normal
machines but changes made to them will not be saved to
disk.

When adding a session machine a normal machine must not
already exist for it. If it does that machine will be
made visible but otherwise left untouched.

If a machine is added to the list of machines on disk that
matches an existing session machine, the session designation
will be removed any further changes will also be persisted
to disk.

Closes #4452
Reviewed-by: Stef Walter <stefw@redhat.com>