test: Test kerberos auth from test virtual machine
Previously we would test the kerberos auth from the verify
machine or at least the container. Unfortunately this led to
hard to debug races in certain overloaded scenarios.

It's simpler to act as the client on the test virtual machine itself.
Theoretically this would catch a slightly diverging set of integration
issues, such as when an operating system is broken as a kerberos
client.

If we do ever get to the point where we want to test browser
(rather than curl) client behavior here, the resulting test would need
to be a selenium test, rather than testing phantomjs.

Closes #4517
Reviewed-by: Peter <petervo@redhat.com>