bridge: Unregister more cleanly from the polkit agent
Because a GSimpleAsyncResult holds a ref to the source object
the CockpitPolkitAgentListener wasn't being cleaned up when
expected, but waiting until all pending operations were done.

This caused issues with Polkit itself already having cleaned
up certain structures by the time the listener was finishing
pending operations.

We now force the pending operations to be cancelled before
we unregister the listener.

Closes #4579
Reviewed-by: Dominik Perpeet <dperpeet@redhat.com>