test: Ignore unexpected SELinux denial of /proc/net/unix
This is a failure that seems to happen for no reason in
Fedora 24 and later. Neither Cockpit or libssh accesses this
file explicitly. As far as I can tell, neither do any systemd
libraries that we load.

I'd like to figure this out. Once we have a reply from SELinux
upstream for how to get a backtrace on this denial we can start
to diagnose the issue further.

https://bugzilla.redhat.com/show_bug.cgi?id=1242656

Closes #4598
Reviewed-by: Dominik Perpeet <dperpeet@redhat.com>