ws: Make the ETag depend on the language requested
This means that when changing languages all resources should
be refreshed and retrieved from the server again. In combination
with the window.location.reload(true) parameter which forces a
reload.

This is a much better than the alternative ... which is serving
the 'Vary: Cookie' header on everything ... or anything translated.

Reviewed-by: Peter <petervo@redhat.com>