common: Convert http Accept style codes to locale codes
Http Accept language codes are case insensitive
and use a dash as the language / country separator.

Locale codes, such as the ones used by zanata in our
po files the language is always lower case and the
country in upper. Underscore is used as a separator.

Use http style codes in the manifest and convert to
locale style codes when looking for a matching file.

Closes #4817
Reviewed-by: Dominik Perpeet <dperpeet@redhat.com>