Bump RPM requirement for storaged packages
It loads /manifests.js and requires a newer version of Cockpit.
Unfortunately earlier versions of the /manifests.js, assumed
that AMD and require.js were in play. This is no longer the case.

While we have code that toggles functionality in cockpit-storaged,
we should tie the packaged version more closely to those of
-bridge and -base.

More related work in the following pull request:

Issue #4976

Closes #4979
Reviewed-by: Dominik Perpeet <dperpeet@redhat.com>