ws: Don't constantly be in endless loop in cockpit-ssh
We don't need to remain in an endless loop in cockpit-ssh.

This happened because we were polling on G_IO_OUT, but then
in turn just doing looping blocking writes when writing to
the output channel.

In addition this should fix a number of test failures where
we were seeing the cockpit/ws container timing out under
heavy load.

Closes #5207
Reviewed-by: Peter <petervo@redhat.com>