Lock down versions of javascript dependencies
For package.json and devel time dependencies, we have problems
with newer versions not being supported by the versions of
Node that are included in distros. So we have to be careful
for updates. We don't mind micro updates. That is the
third digit of the semver number.

However for Bower packages, and particularly those that are
included in Cockpit itself, each point update needs explicit
testing through our integration tests. So we lock these ones
down tightly.

In addition prune the node_modules directory during autogen.sh

Closes #5266