storaged: The --with-storaged-iscsi-sessions option is a hack
Use the term 'hack' more consistently in the override.json and
in the javascript code, so that it's clear what's going on here.

This is a hack because it works around an old version of storaged
that was missing an API option ... and missing *any* way to tell
whether that API option was missing.

Reviewed-by: Marius Vollmer <mvollmer@redhat.com>