bridge: Finish making CockpitDBusJson handle multiple names
This makes a CockpitDBusJson channel capable of talking to multiple
DBus names. We do introspection, watching and rules per DBusi name.

"name" fields are specified along with messages like "call" in order
to direct that call to a specific name. If no "name" is specified,
then the default "name" in the "open" message is used.

It is possible to specify no "name" in the open message, in which
case a "name" must be specified along with messagen like "call".

Closes #5464