networking: Fix regressions re devices without connection settings
This was broken by

   770f2b42dc5162829b8adf4b5c7f4bc7cb9a96a1
   Don't assume unknown devices are ethernet

   767353cba18d19bed88d5d48a854fba532ae11a8
   Don't change settings structure in place

For a device without connection settings, we create "ghost" connection
settings that are added to NM once the device is activated or the
settings are applied via one of the many dialogs.

These settings dialogs didn't get the "ghost" settings anymore and
would thus crash upon being shown.

Also, the "ghost" settings do no longer include the "id" and "type"
fields, which NM usually requires.  We work around this by using the
AddAndActivateConnection method when activating a device or turning a
device into a slave, but the many settings dialogs didn't use it yet.

This is fixed by explicitly providing the ghost settings to the
dialogs, and by routing all settings changes through a common method
that can take care of all the special cases.

Closes #5511
Reviewed-by: Stef Walter <stefw@redhat.com>