ws: Proccess static css files as a template
Provide the OS release data to static css files
and process them as a template.

When the application refers to a different machine,
get the os-release data from the bridges init message.
This allows us to brand based on the actual
machine we are connected to rather than the machine
that cockpit-ws is running.

If no os data is present in the init message,
serve the default css.

Closes #5406
Reviewed-by: Stef Walter <stefw@redhat.com>