common: Add helper function for retrieving process start time
Since process ids are racy, several system components (like polkit)
use start-time of a process as a way to uniquely identify a process.

We're adding this helper function so that information is available
to our polkit agent later. It's by design that we can only retrieve
the start time of the current process.

Reviewed-by: Peter <petervo@redhat.com>