Fail when variable is used before declaration
Javascript allows variable hoisting, where a variable definition
anywhere in a function is as if it is declared at the top.

This leads to hard to find bugs in certain cases. We've been
catching this in review, but lets use jshint to catch it for us.

This commit enables detection of variable use before declaration,
and fixes all cases where that happened.

Closes #5618