bridge: Refactor CockpitChannel frozen support
When a channel is not yet ready all its messages are queued.

Add support for queueing control messages and replaying them in
the right order. In addition we make it possible for callers to
start a channel in a frozen state and not yet progress to ready.
Callers then need to call cockpit_channel_thaw() in order to
unfreeze.