docker: Add the security tab to the image details view
The image details view needs to be more consistent with the data
displayed in the expandable view. In addition it needs to be
more consistent with the kubernetes view of an image.

This does work to make both of these happen. It's not perfect
but it's far better than the status quo.

Closes #5686