test: Ensure cockpit.socket is not running for tests
Several of our tests expect that cockpit.socket is not running when the
VM starts, as they change the unit files. However, even when fixing
those there are still some hard to debug hangs when running the tests in
the production infra, so continue to disable cockpit.socket in the
Debian/Ubuntu images even though it would normally start by default.

Closes #5714
Reviewed-by: Stef Walter <stefw@redhat.com>