cockpit-atomic-storage: Remove unused variables and imports
Spotted by pyflakes. Cover this file in tools/static-code-tests.

cmd_monitor() goes through some effort to read uevent notifications from
udevadm, but then was not actually using that information. The intention
is that we need to call get_info() every 5 seconds even without uevents
ot update the usage information. However, we might want it in the
future, so pass got_event to get_info() for now to quiesce the
"'got_event' is assigned to but never used" error.

Reviewed-by: Dominik Perpeet <dperpeet@redhat.com>