machines: Support for external providers
With this patch, data and actions for the 'machines' package can be
provided by pluggable external providers.

Their use is optional. If none is installed, fails to load or to
initialize, the Libvirt is used by default.

This patch allows reuse of common UI for different virtualization
providers while keeping upstream independent on them. An external
provider brings more realistic view of a virtualization host if
the server is used in single-purpose installation.

The providers can be chained - i.e. certain operation can be
still performed via Libvirt instead of external system.

Provider-specific code is moved to the Libvirt provider (i.e.
VM state manipulation).

The external providers are supposed to be implemented separatly
from the cockpit-project code base.

A provider is installed by copy of it's index.js file under
/usr/share/cockpit/machines/provider and loaded dynamically.

For provider's API and further details, please refer the
machines/README.md.

Closes #5759
Reviewed-by: Stef Walter <stefw@redhat.com>