bridge: Add globbing to loading resources from packages
In order to combine resources in packages we should be able to
concatenate files at runtime. We do this by providing a '*'
package name, which loads the same path from all packages (if present).

This can be used for translations, as well as for having additional
runtime discovered and added functionality combined into a single
component.

Reviewed-by: Marius Vollmer <mvollmer@redhat.com>