bridge: read /etc/cockpit/machines.d/*.json machine definitions
Cockpit plugins, other packages, admins, VM management software, or
config management systems like Ansible/puppet/cloud-init might want to
pre-configure machines for cockpit. This is currently racy as each of
those and cockpit itself have to open and rewrite
/var/lib/cockpit/machines.json. Also, this should be treated as
configuration, not state.

So move the remote machines configuration to /etc/cockpit/machines.d/
and read all *.json files in that directory. Later files override
settings from earlier ones, i. e. they can add new hosts, add new
properties to existing hosts, or change existing properties. The D-Bus
Update() method now expects a file name to update, and the dashboard
writes into "99-webui.json".

Pre-create /etc/cockpit/machines.d in our install and thus our packages,
as otherwise we can't watch it for changes; and the bridge that does the
watching is running as user, so it does not have the privileges to
create the directory by itself.

Adjust the integration tests to the changed paths, and add a new one
that verifies that the old path still gets read (that will be extended
later to also cover the migration).

Closes #5880
Reviewed-by: Stef Walter <stefw@redhat.com>