bridge: Allow use of file descriptors passed over DBus
This works by replacing the file descriptor with a channel
identifier and then allowing that channel to be opened
as an internal stream channel.

There is a maximum number of file descriptors that will
be held in this state by a dbus client until those file
descriptors are used.

Fixes #4380