test: Fix race condition in machines.json migration test (#6007)
The migration happens asynchronously after login in the privileged
bridge; we can't assume it's already done after logging in and the
dashboard shows the expected machines, as these may still have been read
from the old file in /var. Thus wait until the /var file is gone, or for
getting the error about the failed migration in the "nonwritable" case.