ssh: Move to /etc/ssh/ssh_known_hosts
There is no real reason for maintaining our own
/var/lib/cockpit/known_hosts file, as ssh itself already has a global
one in /etc/ssh/ssh_known_hosts. Use that by default, but fallback to
the legacy file for (1) lookups if a host is not already known in the
former but known in the latter; and (2) for writing if the ws we talk to
is still an old version (by checking if ws still has the "ssh"
capability).

Move the determination and setting of the known hosts file into a new
set_knownhosts_file() function, as it is now reasonably complex, will
be extended further in the future with more sources of known hosts, and
avoids handling SSH_OPTIONS_KNOWNHOSTS in multiple different places.

Adjust the integration tests to the new path and add new tests for
covering the fallback to the legacy file.

Reviewed-by: Peter Volpe <pvolpe@redhat.com>