test: Fix TestDocker.testFailure on Ubuntu
In the distant past, Ubuntu used the "admin" group for sudoers, before
moving to the "sudo" group like Debian. Its /etc/sudoers still has an
entry for that for backwards compatibility. As our "admin" test image
user also has a primary group "admin" this matches that rule and thus
"usermod -G '' admin" is not sufficient there. Just remove it, as we
don't ever depend on this old group anywhere.

Closes #6109
Reviewed-by: Stef Walter <stefw@redhat.com>