common: Tell apart internal and opaque user fds in cockpitpipe
When encountering bad file descriptor operations on pipe fds (EBADF),
only generate a warning if it's a Cockpit-internal one. For opaque user
fds which get passed through pipes (i. e via D-Bus), generate a message
instead, to be consistent with "messages for externally wrong input,
warnings for internally wrong stuff".

Provide a new constructor cockpit_pipe_new_user_fd() for this and use it
in the bridge's pipe channel.

This fixes an abort on warning in test-dbus{,-address} when testing
wrong operations on passed fds.