ws: The "user" is not required in CockpitCreds
The user name is not a unique or required part of the session.
The session can have multiple users. We simply record the user
name in CockpitCreds so that we can cross check it when reusing
the password from the session.

This also means that the /login JSON response no longer has a
user.

Reviewed-by: Peter <petervo@redhat.com>