bridge: Don't produce a warning if dbus-daemon is not installed
If dbus-daemon is not installed (like in our Debian test/build images'
pbuilder), read() will return with EOF and we can't read an address.
This is not a Cockpit-internal inconsistency, as we don't depend on dbus
(and don't want to). Thus tone it down to a message, and point out the most
plausible explanation that dbus simply isn't installed.