ws: Add support for HEAD requests
Add CockpitWebResponse "set_method()" and pass the HTTP method in the
"http-stream" signal. When encountering a HEAD HTTP request, ignore all
body blocks in CockpitWebResponse's queueing.

This doesn't avoid the actual computation of the body, but ensures that
we get correct HTTP status codes and Content-Length: without introducing
parallel code paths everywhere. Also, the RFC [1] specifies that
Content-Length: should be identical to a GET request, so in most cases
we cannot optimize this much anyway.

[1] https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.13

Fixes #6202
Closes #6235
Reviewed-by: Peter <petervo@redhat.com>