bridge: Add the option for the host field to be expanded
This introduces + as a separator to allow us to include
an additional key/value parameter along with
the host parameter.

If present the key/value pair is added to the json open
command. This allows us to use those additional
parameters to route incoming requests for available
in alternate bridges correctly.

Closes #6237
Reviewed-by: Stef Walter <stefw@redhat.com>