systemd: Put [Install] section into foo.timer
* systemd: Put [Install] section into foo.timer

and not into foo.service.  This allows us to enable and disable the
actual timer, and not the service that runs when the timer expires.

Also, avoid partial lines in the unit files.

Put timers into WantedBy=timers.target. This is the recommended thing to do.  Also, WantedBy=default.target does not seem to work with certain versions of systemd.

Closes #6324