tools: Fail Debian builds on uninstalled files
Use dh_install's --fail-missing to abort the package build if there are
any uninstalled files left. This avoids bugs like the missing
translations fixed in the previous commit, and forgetting to update the
packaging when adding new modules.

Some modules are not applicable for Debian/Ubuntu or just not packaged
yet (kdump, selinux). Explicitly remove them so that --fail-missing does
not trip over them.

Closes #6376
Reviewed-by: Stef Walter <stefw@redhat.com>