test: Don't use predictable interface names
The interface names that systemd tries to make predictable are
totally not predictable for a given test image. This is because
various versions of qemu and libvirt are used to run the image
resulting in PCI slot ids varying from run to run.

This problem is compounded when we use more than one interface
name defined in the libvirt XML, such as when we use socket
based networking instead of bridge based networking. This refers
to work in later commits.

Instead we turn off the predictable interface names, in order
to make them (strangely enough) more predictable.

In some cases with virt-builder we need to enable DHCP on a number
of different possible interfaces (such as on RHEL) before we
get a chance to turn off the predictable names.

Closes #6403
Reviewed-by: Peter <petervo@redhat.com>