test: Openshift 1.5 does not deliver docker manifests any more
Adjust "check-openshift TestRegistry.testImages" for not expecting the
the docker manifest commands of an image tag any more, as Openshift 1.5
does not deliver that any more as part of "get images". Now just verify
that it shows the size of the layers.

This feature is rather brittle with the current Openshift (1.4) image
even and almost impossible to see manually, so it's not a big loss of
functionality.

Closes #6398
Reviewed-by: Stef Walter <stefw@redhat.com>