kubernetes: Fix incompatibility with certain Firefox versions
I get this error message when using Firefox 53 in the kubernetes
container view and open a shell:

MessageEvent.initMessageEvent can't be converted to a sequence.

I tried to substitute the last argument with null, undefined,
leave it out, all to no effect. This argument is not documented
here:

https://www.webplatform.org/docs/dom/MessageEvent/initMessageEvent/

In addition initMessageEvent is deprecated, so stop using
it when an alternative is available.