bots: Add a new bots directory
These are automated bots and tools that work on Cockpit. This
includes updating operating system images, testing changes,
releasing Cockpit and more.

Bots are being migrated into this directory and documentation
will be updated as they are.

We need to add this directory to git master before following
pull reequests because they rely on it being here.