test: Add ubuntu-stable image
Now that Cockpit is in Ubuntu 17.04 and in backports down to 16.04 LTS,
let's start testing on the latest Ubuntu release.

Use the Launchpad API to query for the latest release, as there is no
"stable" symlink like in Debian. This requires upgrading the 16.04
virt-builder image, there is no newer one.

Copy the naughty overrides from ubuntu-1604 which still apply.
Particularly, 4181 should be fixed in Ubuntu 17.04, so don't copy that
one.

Add workaround for FreeIPA regression (https://launchpad.net/bugs/1693154).

Closes #6570
Reviewed-by: Stef Walter <stefw@redhat.com>