bots: Fix linking of images into place for older branches
For older branches that we don't want to modify, the images need
to treat them correctly. So we create them all as such, even the
ones that we prepared.

In addition the targets for the symlinks need to be files, and
we need to unset TEST_DATA so that the files are no longer

Lastly, the naughty files for the images need to placed so that
older branches will find them.

Closes #6699
Reviewed-by: Peter <petervo@redhat.com>