tools: Fix Files: paths in generated Debian copyright
The change from Bower to NPM in commit 25cb221 caused the Files: fields
in the generated debian/copyright to look like
"Files: cockpit-143/node_modules/redux/*". This is not only wrong (the
paths are supposed to be relative to the source root), but also creates
pointless diff noise between releases.

Adjust build-debian-copyright to only show the path starting with
node_modules/.

Closes #7001
Reviewed-by: Stef Walter <stefw@redhat.com>