packagekit: Automatically refresh if PackageKit has no local data
The very first time when PackageKit gets used it has no cached package
information, and GetTimeSinceAction(REFRESH_CACHE) == -1 (aka. "large
unsigned number"). Force a refresh in this case to get a proper progress
bar and the "Refreshing.." state, instead of letting PK silently do that
in the background and just show an indefinite spinner for several minutes.

For that, separate the "time since action" check from the GetUpdates()
call, and do that in a new initialLoadOrRefresh() function.

Also fix the displaying of "Last checked:": Update the state after
calling RefreshCache(), and also show it if the number of seconds is 0.

Closes #7055
Reviewed-by: Stef Walter <stefw@redhat.com>