packagekit: Always recommend rebooting after applying updates
We currently don't have a reliable way of determining whether updates
need a reboot to fully take effect:

 * PackageKit's API for this is only implemented for Debian-based OSes
   (based on the /run/reboot-required flag), and is only reliable on
   Ubuntu.

 * RHEL/Fedora based OSes don't even restart services on package
   updates, so special-casing the kernel would be a very unsafe
   heuristics.

 * tracer on RHEL/Fedora is not installed by default and can take
   unacceptably long (observed 90 seconds).

During user testing this was one of the top uncertainties, so for now
err on the safe side and always recommend to reboot after applying
updates. This can be refined later (particularly on Ubuntu).

These two new states don't have a header bar (as per design decision),
thus update the tests to wait for the header bar to appear again after
applying the updates and ignoring the reboot.

Fixes #7065
Closes #7101