machines: Generalize updateVmDisksStats() to updateVm()
updateVmDisksStats() only covers a very special case, but the "only
update existing VM" pattern applies to almost every other case where we
currently call updateOrAddVm(). So generalize it to updateVm() and
support a lot more properties, use it where it makes sense, and remove
the now unused properties from updateOrAddVm().

Simplify the interface of updateOrAddVm() and updateVm() in actions.es6
to just pass the properties object as a whole, instead of dissecting and
rebuilding essentially the same object.

Note: This breaks the provider API, as updating e. g. "state" with
updateOrAddVm() no longer works. But this was recently made internal
anyway, so this is no external API change.