packagekit: Add (disabled) mock updates
Testing/changing the "available updates" layout is much more practical
on the local machine rather than through updating/running integration
test VMs. Add a mock-updates.es6 helper which can inject updates with
various interesting properties, and show in a comment how to update
updates.jsx to use it.

This will not get included in the webpack by default, as nothing
require()s it.