packagekit: Truncate long package lists/changelogs
Updates with an overly long package list or changelog make the entire
list hard to browse. So truncate it to 15 packages or 7 changelog lines
by default. On overflow, show a "more..." link and expand. The expansion
will also be done when clicking anywhere else in the table cell.

This requires turning UpdateItem from a functional into a class
component, to keep track of the "expanded" state.

Fixes #7171
Closes #7264