kubernetes: Handle url encoded selfLinks
Parts of a selfLink may be url encoded. We need
links to match even when they were created without knowing
the selfLink. Because not all versions of kube do this
the same way url encode the name portion when building
a resourcePath and url decode the path that is used as the
key in the object.

Closes #7321 
Reviewed-by: Stef Walter <stefw@redhat.com>