machines: Fix security and broken packaging in noVNC
Using unsafe-inline Content-Security-Policy opens up a security
in something as wild as noVNC. We simply shouldn't have unsafe-inline
or unsafe-eval in our packages.

In addition the noVNC component should be installed efficiently, in
a single bundle, not having to load 15 javascript files. The bundle
should be compressed, with unnecessary dependencies removed.

I've added a small custom Webpack loader to perform the task of
combining these files.

This also allows development on this component to work with pure
webpack without automake and friends. We also properly track the
dependencies and changes here.

Closes #7388