ws: Do not cache responses that default to shell html
When a user is logged in and they request a path that
isn't for a specific file we serve the shell html. When
they are not logged in we serve the login page.

Currently the shell html is served with a short cache
that varies on cookie. However the cookie can invalidated
by cockpit-ws, through an idleness or logout.
If the user visits that page again before the browser expires
the cookie they will see the cached response (shell html)
when really what they would expect is the login page.

In short we always want these requests that could reply with
either the login page or shell html to reach cockpit-ws
so that it can decide what to serve.

Closes #7486
Reviewed-by: Stef Walter <stefw@redhat.com>