test: Invoke a tests-policy script for figuring out retries
When figuring out known issues or retries, pass the entire output
of the test to the retry script.

This helps us be able to have the bots logic use other criteria
for whether to retry a test, such as learning about flakes.

In order to do this we need to capture the entire output of the test
in more cases, not just for buffering multiple concurrent jobs.
We do this unless the --verbose or --thorough options are passed
to the test invocation.

Closes #7808