tools/cockpit.spec: Adjust storaged/udisks2 dependency
Adjust the target OS specific dependencies of cockpit-storaged:

 * RHEL 7.4 and current CentOS 7 ship storaged and do not support
   Recommends:.
 * RHEL = 7.5 will only ship udisks2, which does (rightfully) not
   provide storaged. Still no support for Recommends:.
 * RHEL 8 will only ship udisks2, but does support Recommends:
 * Fedora 27 also moved to udisks2, earlier versions have storaged. All
   supported Fedora releases support Recommends:.

Standard RPM macros do not allow us to differentiate between RHEL 7.4
and 7.5 (%rhel == 7), so check /etc/os-release for that. This works for
"real-life" builds, but not for building on our test images as these use
a CentOS 7 mock chroot for all RHEL builds. Thus do the os-release
parsing in fedora.install and pass on the value as define.

The udisks2 modules for lvm2 and iscsi are not yet available on RHEL
7.5, so skip these dependencies for now.

Closes #7886
Reviewed-by: Stef Walter <stefw@redhat.com>