tools: Fix broken dist paths with out of tree builds
Running `make dist` in an out-of-tree build that is not underneath
srcdir previously resulted in a broken tarball, containing pkg/* and
node_modules/* under a path to the srcdir.

E. g., diffing file lists between an in-tree and out-of-tree dist
tarball looked like this:

  -cockpit-154.x/pkg/lib/journal.js
  +cockpit-154.x/upstream/cockpit/pkg/lib/journal.js

Avoid this by putting pkg and node_modules symlinks into the build dir
and chopping off absolute build dir prefixes in webpack-make, instead of
creating path prefixes to the srcdir.

Change the invocation of webpack-make to always pass a canonical
absolute $SRCDIR, so that maybePushInput() does not have to handle
different cases of absolute vs. relative srcdir.