tools: Drop Debian package cockpit-pcp package more thoroughly
We haven't built cockpit-pcp for a while as pcp hasn't been in stable
and testing for a long time. But keeping the "Package:" binary record in
control confuses the Debian archive scripts as the package regularly
lands in binNEW (even through they are no actual new packages). This
causes unnecessary release delays and manual ftpmaster actions.

So drop the binary package record together with the build dependencies
when releasing cockpit. This requires removing the installed pcp files
as well, otherwise `dh_install --fail-missing` will break the build.

Also remove the "Suggests: cockpit-pcp".

Closes #8106