test: Fix broken nslookup wait loop in check-realms
Commit 3c93a681 introduced a wrong assumption that sssd.service failure
would be unexpected at this point. But /etc/sssd/sssd.conf is only being
generated by `realm join` later on, so if sssd.service is enabled, it
would necessarily be in a failed state after booting due to the missing
configuration file.

This caused the nslookup waiting loop to be aborted after the first
iteration, and thus was a huge race condition. Put back the entire loop,
and instead move the sssd.service debugging until after a failed realm
join.

Closes #8136