test: Port unit tests to chrome debug protocol
Use the same code to launch a chromium instance as the integration
tests, which is why the new script lives under test/common instead of
tools/.

Also move get_js_log() to the cdp class. It is heavily tied to the
driver, because it accesses one of its global variables. Move this
access to the cdp class and always access logs via this function, so
that we minimize the impact when we change the variable.

Also parameterize the injection of `sizzle.js` and `test_functions.js`
in the cdp class. Those are only needed for the integration tests.

Closes #8314