packagekit: Remember update severity from PackageKit
In the `GetUpdates()` `Package` signal handler, don't translate the
`PK_INFO_ENUM_SECURITY` severity to a `security` boolean, but remember
it as-is. This allows us later on to also display other severity levels
(like "enhancement") in a different way.

Consider that some backends (like dnf) give a wrong severity and thus
clamp the value between `LOW` and `SECURITY`.