doc: Stop advertising cockpit bundled jQuery
We currently ship a really old jQuery 2.2.4 which has several
vulnerabilities. At some point we must kick this out of our API, and
move packages to include their own (or stop using jQuery altogether).

As a first step, stop documenting it as part of our API. This also gets
rid of the confusing reference to a "latest" package which doesn't
exist.

Closes #8504