bots: Drop obsolete unsetting of TEST_DATA (#8598)
This was introduced in commit 3a0980 and remedied in commit ea5dfbc6d to
support the old rhel-7.4 branch. But that does not use the old
test/images any more since
https://github.com/cockpit-project/cockpit/commit/f2ca300c51 (and they
would not work any more anyway, as the images are long gone). Thus this
hack is obsolete now.

This makes the image cache actually being used for testing third-party
modules, where this particular way of checking for bots/images failed.