bots: Build new openshift-prerelease image
We will use this for porting cockpit-kubernetes and checking for
regressions. Both of this should happen before the final release of a
new OpenShift version.

Check the image name in openshift.setup, and if it contains
"prerelease", search for prereleases on the GitHub release page.

The output format of `oc get services` changed with 3.9, it gets another
column. Avoid assuming which column contains the IP address, and just
search for it with a pattern.

Explicitly set `apiVersion` in the `RoleBinding` records, as this is
now required with 3.9. This is backwards compatible with earlier
OpenShift releases.

Patching admin rolebinding fails in 3.9. It says "not patched", and
scruffy is already an admin, so ignore failure.

Closes #8740