bots: Fix kubeconfig download path
With commit 3e118a46574 the default destination directory (for relative
paths) changed to bots/, where as before it was depending on whether the
module was included through the symlink in bots/ or through tests/. This
broke image-create's downloading of the openshift image's
openshift.kubeconfig.

All other usages of `Machine.download()` use absolute destination paths,
so do the same here to avoid relying on unintuitive relative paths.