packagekit, apps: Factorize PackageKit basic transaction into common library
Add functions for running and watching a PackageKit transaction to
pkg/lib/packagekit.es6. Replace the `failHandler` argument (as the
version in pkg/packagekit/updates.jsx had) with a promise rejection,
which is much more common JS style.

The apps' `transaction()` is significantly more complex than this basic
transaction, so write a shim for it for now.

Save the extra cockpit D-Bus proxy (which is rather expensive, and
occasionally causes hiccups with PackageKit) by remembering the progress
data throughout the entire transaction.