bots: Don't spam logs with image-download curl progress report
When the output goes to a log file and a test has to download an image,
this results in a ~ 500 kB (!) output of the progress meter. This blows
up the logs, causes tests-policy GitHub errors (as that tries to put the
log output into a GitHub comment, which is limited to 64 KiB), and
confuses the log learning pattern matching.

Avoid this by only showing the progress bar when stdout is a terminal,
i. e. the command is being run interactively.

Also show errors during actual download by adding `--show-error`, which
are normally hidden with `--silent`.

Closes ##8743