storage: Add Network keys demo
The idea is that we add the necessary APIs to UDisks2 (and libblockdev),
but this here is a quick and dirty skeleton that should inform the API
design.

There is a hidden switch that needs to be explicitly flipped before this
appears in the UI:

    window.localStorage["clevis-feature"] = "on"

If you want to play with this, you need to set up a `tang` server:

```
[Socket]
ListenStream=
ListenStream=88
EOF
```

This puts the tangs server on port 88 instead of the default 80.  Any
other port is fine as well.  (If port 80 is fine, you don't even need to
put the file into with /etc/systemd.)

Then install the `clevis-luks` package on the system where you want to
use this.

Closes #8785