test: Initial aXe integration
Start running accessibility checks in our tests with
[aXe](https://www.axe-core.org). Add a `MachineCase.check_axe()` method
which generates an aXe report for the current frame and writes the JSON
reports into an attachment (names similar to screenshots and logs). This
gets heavily trimmed and gzipped to reduce the size from ~ 70 to 2 kB.

Call this for the top-level navigation frame (in check-menu) and for the
initial /system page (in check-system-info) for now. Once we get some
experience how to process, compare, and fix these, these calls will be
added to more tests.

To avoid injecting the axe check code into every test (which might skew
the results), introduce an `@enableAxe` decorator which does the
injection. Every test that uses `check_axe()` must use that decorator.
Also, only do this on the default OS test for now (fedora-27 at the
moment), as the results are going to be mostly identical between OSes
anyway.

Fixes #8641
Closes #8780