test: Adjust OpenShift testImages test for different tag label sorting
OpenShift 3.9 now sorts the "latest" tag last, which makes it disappear
behind the "and 1 other" overflow label. Check for "2.11" instead which
is visible with all versions.

The intention in registry-image-widgets is to have the tags sorted by
name, so "latest" was always supposed to be last. This just didn't work
with 3.7 and earlier.

Fixes #8745
Closes #8791