bots: Stop testing branches
When they have nothing else to do, this caused our bots to perpetually
run tests on master and rhel-7.4 branches (notably not rhel-7.5, which
we never added). Nobody looks at them, there is no expected change (both
the code and the test VMs stay the same), and the meaningful CI happens
on pull requests.  We never push directly to master, other than very rare
reverts of a recent PR that cased trouble; so this is just a waste of
resources.

Moreover, this breaks badly with new external repositories: If they
don't yet have tests, then the bots will be on a tight loop of "start
tests on master, fail on missing test/run, start over" - in this case
there is no place in GitHub to attach a result to, so these can never be
marked as "failed" properly.

Closes #8795