realmd: Fix tooltip on disabled link
Bootstrap tooltips don't work on disabled elements (by design:
https://getbootstrap.com/docs/4.0/components/tooltips/#disabled-elements)

So nest the actual "Join Domain" link into a parent `<span>` that
carries the tooltip instead.

Add a new `tooltip_element` argument to `update_privileged()` for this
case, that will make it attach the tool tip to that element instead of
the `this` one. All other usages of `update_privileged()` are on
buttons, where tooltips work fine even if disabled, so no other
invocation needs to be changed.

Fixes #8827