lib: Always use setState
Setting this.state directly might not stick when a previous setState
is still in the queue.

Also, calling the "handler" function might do all kinds of things that
cause the setState queue to be processed.  Thus, this code might
actually print two different values:

    var foo = this.state.foo = handler();
    console.log(foo, this.state.foo);

This might seem impossible, but it might be executed like this:

    var state = this.state;
    var foo = state.foo = handler();
    console.log(foo, this.state.foo);

Thus, if handler causes this.state to change, we have a bug.