bots: Drop docker resolv.conf hack
https://bugzilla.redhat.com/show_bug.cgi?id=1448331 apparently got fixed
a while ago, and the hack is not necessary on any of our images any
more. Also, the resolv.conf creation hack was duplicated in a lot of
places.

Should this still happen in some special case, let's only do this in the
tests, not on the images, as they are much harder to change and clean
up.

Remove the hack from the Atomic images for now. It is still present in
{debian,fedora,openshift,rhel,selenium}.setup, these will be rebuilt and
cleaned up in separate pull requests.

Closes #9040