tools: Support only or not building optional RPMs
Move the subpackages that are currently in RHEL Extras into one block
and conditionalize it on a new `build_optional` define. Likewise,
conditionalize the basic packages on a new `build_basic` define.

This paves the way for easily doing a downstream package split into two
source packages (for BaseOS and AppStream).

Closes #9057