test: Adjust ostree rollback test to new rpm-ostree behaviour
Previously, when rolling back from base.2 to base.1, base.2 appeared
again as normal available update. Now rpm-ostree flips between the two
with "rollback" and reverses the order.

Both behaviours make sense and got ack'ed by upstream, so expect the new
behaviour for current fedora-atomic.

Also verify that the order is as expected for the old behaviour.