lib: Set attribute flag after `update_permissions()` is done
That makes it possible to write a test that waits until permission check
on an element finished, and asserts that the action is allowed. Without
this, a test can only cover the "not allowed" case as then it can wait
for the element to get disabled. But they are enabled by default,
so there is nothing to wait for in the "allowed" case.

Use the same `data-stable` attribute as the systemd/host.js page, let's
standardize on that.