bridge: Use SHA256 for package checksums
SHA1 got [broken](https://shattered.it/) a while ago. Our package
checksums are not security sensitive, but for the sake of easier
auditing, change them to SHA256.

Avoid the repeated inline checksums in tests and factorize them into
constants. Also update the command how to recompute them to use the
bridge from the build tree instead of the system.

See issue #5940 
Closes #9170