users: Notify that a login/logout is necessary to get role changes
The Roles checkboxes add and remove users from key groups which give
them access to do certain things.

However these changes do not fully apply until a user has logged out and
logged back in. This is true both outside of Cockpit in a terminal and
inside of Cockpit.

There is an exception: One can run 'newgrp' to start a shell with updated
groups. However, this does not 'fully apply' the group changes and the
notification is still correct despite this exception.

Closes #9477