build: Run tools with available python version
tap-{gtester,driver} can be run with Python 2 or 3. Call them with
whichever is available, as this differs between Fedora and RHEL.
This avoids having to build-depend on both python 2 and 3.

Make the remaining tools bilingual and switch them to Python 3 by
default.

Drop the python3 runtime hack from test-bots, as it is now always
invoked with Python 3.

Closes #9525