Respond to "ping" messages with "pong"
The protocol has been extended to respond to "ping" messages
with "pong" messages. "ping" messages with channels are forwarded
all the way through to their destination, before being responded
to. "ping" messages without channels are responded to immediately
and not forwarded.

The contents of a "pong" are identical to the ping except for the
fact that the command has changed.

Closes #9552