build: Skip test-httpstream unit test with GnuTLS 3.6.3
GnuTLS 3.6.3 fails to build up TLS connections to other GnuTLS servers
(https://gitlab.com/gnutls/gnutls/issues/530) when not providing
certificates. test-httpstream detects this.

This is outside of our control, and causes package build failures in
Fedora and RHEL. This is a cowboy hack, but as the regression is already
fixed upstream it won't last very long.

Closes #9730