bots: Fix debian-stable image build
Add the security updates apt source to the bootstrap stage, so that it
can already pick up the latest kernel. Otherwise the .setup might run an
older kernel but dist-upgrade to a newer one, which causes the aufs DKMS
module to be built only for the newer one; after that, docker
initializes its storage backend to devmapper as aufs is not available
for the running kernel.

Fixes #9873
Closes #9940