ssh: Query known hosts from sssd
When no local known_hosts file knows about the host to connect to, try
and ask `sss_ssh_knownhostsproxy` about it. This makes `cockpit-ssh`
work seamlessly with centrally managed known hosts list from FreeIPA.

This utilizes the new `--pubkey` option of sss_ssh_knownhostsproxy. If
it isn't available (on older releases), it will just fail fast and the
behaviour is as before.

On ubuntu-1604, the new test triggers the already known nsupdate crash
(issue #9487), add an alternative naughty pattern for it.

https://bugzilla.redhat.com/show_bug.cgi?id=1622835

Closes #9951