common: Ensure initialization of length return value in dump_value()
This Should Not Happen™ and a critical will be logged when dump_value()
encountered an unknown data type. But nevertheless
the returned length was uninitialized in the caller and passed on to
g_bytes_new_take(). This is a potential uninitialized memory access, so
robustify this to return 0.

Spotted by coverity:

    warning: 2nd function call argument is an uninitialized value

Cherry-picked from master commit 70b195faac0be

https://bugzilla.redhat.com/show_bug.cgi?id=1644209