machines: CREATE_AND_ATTACH_DISK action should not handle errors
The dialog that is calling this method should catch any errors
and display them to the user.

Heavy nesting of promises was removed with this commit, by chaining
them with the then operator, which will allow any error in the chain
to be propagated to the error hander in the calling side.

Closes #10271