users: Fix creation of locked accounts
Running "chpasswd" last will undo the earlier "usermod --lock", so we
do it the other way around.

The added checks in the test also remove a race condition. We need to
wait for the unlocking to finish before changing the groups.  (And
using getent instead of grep will hopefully remove a similar race
between Cockpit and the test changing the groups.)

Closes #10327