test: Robustify console error checking
Now that we fixed most React errors, turn the current blacklist into a
whitelist. This will also catch other React errors, and guard against
typos.

We currently only have one remaining React warning left:

    Warning: Can't call setState (or forceUpdate) on an unmounted
    component. This is a no-op, but it indicates a memory leak in your
    application. To fix, cancel all subscriptions and asynchronous tasks
    in the componentWillUnmount method.

The impact of this is very low, and debugging all of these will take
time, so generally whitelist this one for now.

Also, a few tests check error conditions and deliberately produce
console errors. Introduce a per-test whitelist mechanism similar to what
we do for expected systemd journal messages. This uses substring
searching, unlike journal patterns which have to match the *entire*
message. This is often unwieldy, let's move that to matching in a
follow-up.

Move this out of the low-level cdp-driver.js into testlib.py, and merge
it with the already existing "unhandled exception" checking; that is
just a special case of a console error.

Closes #10328