test: Skip mock-sshd unit tests with libssh 0.8.4
libssh 0.8.4 breaks interactive keyboard authentication on the server
side (https://bugs.libssh.org/T117). This breaks our mock-sshd which we
use in unit tests. As there is no way to work around that on the client
side, skip the affected tests with that version. The fix is already
upstream, so 0.8.5 will not be affected.

This should be reverted once 0.8.5 lands in all Fedora ≥ 28.

Cherry-picked from master commit bfae61172329c32c.