Replace our custom Tooltip with PatternFly-React's
Add IDs to them, so that the tooltips become accessible to screen
readers. Note that they don't have to be unique, as there is always at
most one tooltip visible/rendered.

While touching the kdump code anyway, fix the totally broken layout in
the "available, but no kmem reserved" case: The (i) button was broken to
the next line, and there was no spacing between the button and the
"Service is stopped" link.

Drop our own component, it's not being used any more. Also drop the
corresponding Playground demo, the upstream docs at
https://patternfly-react.surge.sh/patternfly-3/ are a much more
comprehensive place for that.

Closes #10452