networkmanager: Fix enabling firewalld on Firewall page
Enabling the firewall on the Network » Firewall page depends on the
firewalld D-Bus client "owner" signal to fire. This works if the
cockpit.dbus() proxy got instantiated at a time when the D-Bus service was
running (or at least activatable), but it's completely dead if not. This
caused the On/Off button to forever remain in disabled state.

https://github.com/cockpit-project/cockpit/pull/9125 attempts to fix
this in a generic way, but it's apparently a bit problematic to land
this. So work around that by re-initializing the FirewallD1 D-Bus client
and its subscriptions when firewalld.service goes to state "running".

Adjust the tests to start with firewalld being disabled, which
exposes the bug in both test cases. Previously the test was cheating by
pre-starting firewalld on the command line. This also requires adding a
third variant of the naughty override for #9722.

https://bugzilla.redhat.com/show_bug.cgi?id=1646936
Fixes #10501
Closes #10511