bots: Install firewalld on Debian/Ubuntu images
Enable the corresponding tests.

This exposes a bug with the Debian 8 docker.io backport when running on
Debian 9: Port forwarding in a container fails with

    iptables failed: iptables --wait -t nat -A DOCKER -p tcp -d 0/0 --dport 3380 ! -i docker0 -j DNAT --to-destination 172.17.0.8:3380:
    iptables: No chain/target/match by that name.

As this is a totally unsupported package and thus nowhere to report a
bug to, just add a hack to disable firewalld for this case.

Also disable firewalld for the libvirt tests, as we don't configure any
rules to allow traffic on the virbr0 bridge. This causes DHCP
requests/responses to get blocked, and thus images take very long to
boot waiting for network configuration.

This exposes that known issue #9722 also affects debian-testing, so copy
the naughty override.

Closes #10512