Automatically eslint-fix indentation/spacing in our *.js files
Eventually we want all our source code to just be *.js files, subject to
one common eslint rule set. As a first step, get the bulk of eslint
errors out of the way by running

    $(npm bin)/eslint --ext .js --fix pkg/

This happend with temporarily setting the "operator-linebreak" rule to
"off" in .eslintrc.json, as jshint and eshint have incompatible opinions
on where to put the line breaks. These can easily be fixed once we drop
jshint in favor of eslint.

Closes #10635