networkmanager: Fix data type handling in Firewall "Add Services" dialog
`selected` starts out to be a Set. Keep it a set instead of turning it
into an arry in `selectionChanged()`. Convert it to a list in the
`addServices()` call, as that cannot deal with sets, only arrays.

This fixes a `TypeError: services.map is not a function` crash when
clicking "Add Services" button in the dialog without any selection.

https://bugzilla.redhat.com/show_bug.cgi?id=1652207

Closes #10642