ws: Add --local-session option
This makes cockpit-ws talk directly to a bridge, with skipping
cockpit-session and all authentication. This allows the web server to
run as any unprivileged user in an already running session, which is
useful for desktop integration or a kiosk mode.

Redirect the original stdout to stderr, so that g_debug() messages and
similar go to stderr and don't interfere with the bridge protocol.

Reorder the `--address` and `--local-ssh` options in the cockpit-ws
manpage to coincide with the `--help` output and how the options belong
together.

Closes #10683