docker: Avoid initialization to "undefined"
Spotted by eslint "no-undef-init" rule.

Only in this case it actually had an effect, as it was declared inside a
`for` loop, but the scope extends to the entire containing function. As
we can't use `let` until we apply babel to our *.js files, move the
declaration out of the loop, to avoid the mental trap about broken `var`
scoping.