machines: Remove target from add disk dialog and disk table
<target> element for disks in libvirt is not the mapping of the disk
device to the guest filesystem in /dev/.
Usually this is the case, but when creating a new disk with specific
target device, libvirt doesn't pass this information to qemu. So we
should not expose this to the cockpit user since it can be misleading.

In addition, the ability to choose the device target for new disks,
is not something admins usually do, so dropping this should not cause
issues.

Closes #10688