tests: skip machines testCreate in debian-stable in case qemu error is missing
debian-stable has old version of QEMU (2.8.1) which contains a bug that is
probably fixed in newer versions. If the host is under a load QEMU binary
crashes in TCG code which will result in not printing any error and only the
libvirt warning.

Skip test if test image is debian-stable and the error message doesn't
contain qemu failure details.

Closes #10649