ws: Drop cockpit.desktop
The current user experience with this .desktop file is rather poor:

 - cockpit.socket is not enabled by default (on purpose).
 - Desktop users rarely see the motd message how to enable it.
 - You will need to log in again in the browser, even though you are in
   an already running session.

We are working on a better integrated solution, with .desktop files for
individual components (and not showing the full navigation and menu).
Until then, remove this to avoid the above confusion on a desktop.

Keep shipping the cockpit.png icon though, it may come in handy in the
future.

Closes #10694