ws: Better error message if cockpit-system/bridge is not installed
If only cockpit-ws is installed, trying to login currently shows an
"Internal error in login process" message, which is not very helpful. In
the case of a "no-cockpit" problem, change this to a more helpful error
"The cockpit package is not installed".

This might not be 100% accurate in some corner cases, but in the vast
majority of these problems this will be the root cause.

https://bugzilla.redhat.com/show_bug.cgi?id=1650467
Closes #10765