build: Move to Babel 7
Enable the "commonjs" modules option as we still have tons of non-ES6
`module.exports`.

Move the babel options into the webpack config and drop the recently
introduced .babelrc again. Babel 7 does not find it any more when
building webpacks in a separate build tree. This also fixes the
regression of not shipping the babel config in release tarballs.

Closes #10825