kdump: Port page and dialog to ct-form-layout
Also add missing `htmlFor` attributes to the labels in the location
configuration  dialog.

This triggers a curious bug/effect in Chromium: The first time the "Test
Configuration" dialog is opened, it immediately closes again, apparently
the button click event somehow "falls through" the dialog that opens.
This happens both in check-kdump as well as interactively. Fix that by
stopping event processing in the handler.

Closes #10930