bots: Enable sudo rule on IPA server
Unfortunately it was decided that the IPA sudo rule would not get
enabled by default. The way to go was to use the new
"enable-admins-sudo" ipa-advise, which got introduced in FreeIPA 4.7.1.
So update the documentation accordingly.

As this is not a temporary hack any more, move it to the ipa image
setup script, to save some time on running the tests. This also exhibits
a structural problem when using Kerberos authentication: As sudo has a
"last matching rule wins" semantics, and nsswitch.conf defaults to
"sudoers: files sss", our "passwordless sudo" hack in testNegotiate()
breaks. Work around that, and extend the comment to explain what's going
on.

Closes #11001