systemd: Quiesce bridge journal message about /sys/class/dmi/id not existing
cockpit.spawn() logs an error when the given directory does not exist.
As this is a legitimate case on non-x86 systems, change the call so that
it's grep that fails, not cockpit.spawn(), and thus the error goes to
the console log instead.

Add a test case that reproduces this.

Part of issue #8807