realmd: Show domain information and less scary leave dialog
Participants in our usability studies expected to get further details
about an IdM domain when clicking on the domain name on the System page,
intead of directly asking to leave it.

So change the dialog to show the domain name, login format (slightly
more human friendly to replace `%u%` with `username`), server and client
software. Hide the "Leave Domain" behind an expander, and show the
warning in there.

Fixes #10063
Closes #11096