bots: Do not show unnecessary information in terminal
When logged in as admin into Ubuntu images a message
`To run a command as administrator (user "root"), use "sudo <command>".
See "man sudo_root" for details.` is displayed on first two lines.

This makes testing terminal harder, as on other distros we expect first
line to contain prompt and other lines to be empty.

After this commit mentioned lines are not displayed anymore.

Closes #11141