bots: Factorize check for Red Hat network
Stop exposing `REDHAT_PING` and just offer a function that says whether
or not we are in the Red Hat network. This avoids duplicating logic. The
function caches its result, so it's fine to call multiple times.

Check the https port of the image server, so that we stop the
requirement of having to expose http just for this purpose. As this
doesn't have a valid certificate, just check if we can resolve the name
and connect to the port.