bridge: Allow data: image and font sources in C-S-P by default
This will allow us to embed SVGs or glyphs for global styling, as
relying on externally shipped graphics in a central place is really
difficult given the API compatibility with older -ws versions.

We already do this in machines, ovirt, and packagekit, and will need it
for moving to `<select>` (PR #11155) and eventually when moving to
PatternFly 4 (which uses background images). This does not weaken our
security as we only load CSS from trusted places, so let's generalize
this.

Note that packages cannot immediately rely on this, as they might be
served from an older cockpit-bridge package. Until then, packages need
to explicitly add this C-S-P to their manifest.

In `test_csp_strip()`, change the manifest override to drop data:, as
that's now the default. Instead, verify that not specifying data: in the
manifest is respected.

Closes #11268