machines: remove wait and noreboot options on virt-install in the VM creation
There is not any reason to block virt-install waiting indefinitely for the VM
to get installed. We should just make sure that guest it's running and
let the creation script to continue.

--wait -1 would probably make sense if we showed in the UI the state of
the VM as creation in progress for the whole time that the VM was getting
installed, but instead when we get the event from libvirt that the VM
started we just show the VM state as running.

So removing --wait -1 will make no difference to the cockpit users.

Lastly if we don't block on the installation to finish with --wait -1
there is not point to call --reboot since it will have no effect anyway
if the installation gets out of control of the virt-install command.

Closes #11272
Fixes #8677