ws: Mark deleted cookie as "Secure" with https
This is completely irrelevant for security, but scanners such as Qualys
complain about it. This makes the deleted cookie have the same
attributes as the real one. See commit 050262f06 for a similar issue.

Adjust test-handlers to verify the non-secure path. The TLS path is
unwieldy to handle in unit tests, so cover that in check-connection.

Fixes half of https://bugzilla.redhat.com/show_bug.cgi?id=1677767

Closes #11279