bots: Drop VirtMachine.add_netiface() vlan option
This isn't supported by newer QEMU versions any more (as in Fedora 29).

`check-networking-bridge TestNetworking.testBridge` was the only user of
that. But as the second `add_iface()` call does not specify an existing
VirtNetwork, this creates a new network anyway, and thus it's already
an independent network.

Cherry-picked from master commit 26156553ac0.