storage: Update file system descriptions
"RHEL 7" is confusing on RHEL 8 or Fedora. According to
https://access.redhat.com/articles/3129891 the main reason to chose ext4
over XFS these days is that it may perform better on low-end hardware
wit limited I/O or CPU performance.

But in general, any 5-word file system description is misleading; e. g.,
users might be tricked into choosing VFAT as it's "compatible with all
systems", but it's a really poor choice for /var or /home. So remove all
file system descriptions.

Stop i18n'ing the file system names; they are proper nouns. As we
changed the strings, this avoids showing up as untranslated again.

Fixes #11481
Closes #11514