ws: Replace pam_shells with running bridge through user's shell
There are some shells like session-recording's tlog or rsh which the
user shouldn't be able to enable or disable, and thus they aren't in
/etc/shells. But they are nevertheless a legit login shell.

So using pam_shells in commit 4c6ffd27015 was too strong. Drop it, and
instead simulate what /sbin/login or ssh do by running our session
leader cockpit-bridge through the user's configured shell. This will
fail (with exit code 1) for e. g. /sbin/nologin or /bin/false; handle
that in on_pipe_close() for a nicer error feedback than
"internal-error".

However, this makes the cases of cockpit-session failing to run
cockpit-bridge, and cockpit-ws failing to  run cockpit-session itself
(broken install) indistinguishable by just looking at the exit code and
problem (1 and "access-denied") in both cases). To decide which case we
have, check if the authorization protocol already started or not.

https://bugzilla.redhat.com/show_bug.cgi?id=1631905

Closes #11575