bots: Map from project through branch to contexts
This commit mainly introduces mapping from project through branches to
contexts. Also if none repo is passed through argument use one that is resolved
by GitHub.__init__(). This means repo is always present, so conditions
as `if not repo...` need to be either rewritten to different logic or
when directly related to fact that bots are part of
cockpit-project/cockpit then just change it to `if repo !=
"cockpit-project/cockpit".

Closes #11581