tls: Add cockpit-tls proxy
This program acts as a reverse TLS proxy for cockpit-ws, so that the TLS
termination and client certificate validation will happen outside of
cockpit-ws (which is too big to trust). Eventually it will run as a
separate system user, but process isolation already improves security
due to standard ptrace protection.

cockpit-tls spawns separate cockpit-ws instances for each client
certificate, TLS with no client certificate, and unencrypted HTTP. This
prevents sniffing and tampering with credentials if attackers can coerce
cockpit-ws to run arbitrary code.

Adjust the systemd service to run cockpit-tls instead of cockpit-ws.

Closes #12068