ssh: Consider the correct remote ssh host
`$COCKPIT_REMOTE_PEER` is the IP address of the browser client from
cockpit-ws' point of view, *not* the remote ssh host to connect to.
Commit 29c0bc33f8 used that to determine whether to ignore unknown host
keys, but this should depend on the ssh remote, not the cockpit-ws
address.

Thus move this decision out of cockpitsshoptions.c (which doesn't have
access to the ssh target host) and into cockpitsshrelay.c.

To be able to test this in test-sshbridge.c, teach the test fixture
about a "hostname" field, and bind the mock-ssh server to that. This
allows setting up a connection to 127.0.0.99, which cockpit-ssh does not
treat as "localhost". Use the existing /ssh-bridge/hostkey-unknown test
for the non-localhost case, and add a new
/ssh-bridge/hostkey-unknown-localhost test that verifies that
cockpit-ssh can connect to an unknown localhost.

This changes the behaviour of the login page from "sometimes accepts
unknown hosts" to "consistently refuses unknown hosts".