machines: Check if properties are available before attemting to read them
Sometimes libvirt-dbus might return part of the properties when calling
the GetAll method. This is expected, as some resource might have been
deleted in the middle of parsing the properties values. Only the
properties parsed before deletion will be returned in this case.

We already peform the checks for the Domain properties. Do the same for
Storage and Network.