ws: Fix /web-server/for-tls-proxy test case
Commit f36fbf571936 introduced this in a weird setup: We do expect
cockpit-ws to redirect to https if it has a certificate. But that test
case set up a certificate *and* tls-proxy mode, which wouldn't be used
together ordinarily. This was hidden by forgetting to pass the
certificate to the CockpitWebServer constructor, so that the fixture's
`cert_file` got ignored.

CockpitWebServer itself doesn't do anything with the for-tls-proxy
property other than distributing it around, so just check that the
proeprty setting and getting works, and that it doesn't break HTTP
requests.