tools: remove tools/tap-gtester
This wrapper used to convert the now-deprecated gtester format to TAP.

Versions of GLib from 2.38 have supported the --tap option, and the
upcoming 2.62 release will default to TAP.  For now, make sure we
continue to pass '--tap' to all testcases in order to ensure we get TAP
output even on older GLib versions.  We can remove that once we depend
on GLib > 2.62.

Cherry-picked from master commit 64d771c6ba39.