common: Drop glib usage from cockpitconf
This makes it possible to share with the upcoming cockpit-tls, which
must not use GLib/GObject.

Rename cockpit_conf_guint() to cockpit_conf_uint() to reflect the data
type changes.

For simplicity this uses a single linked list for storing the data
instead of a hashtable of hashtable. These are being called so rarely,
and cockpit.conf files are so small that it's not worth all this
overhead.

Also, stop supporting parsing the same key strv with different
delimiters. This is unrealistic to happen in practice (and doesn't), and
it would significantly complicate the caching.

Closes #12014