machines: don't filter out non 'dir' type pools in add disk dialog
Previously we were filtering out 'dir' type pools from the dropdown with
the available pools to choose/create volume on.

This commit enables all pool types in add disk dialog.
Some types do not support 'format' option so the code needed to be
adjusted accordingly to take in mind this limitation.

Another issue that we need to take in mind is that not all pool types
support virStorageVolCreateXML API.
So in case a user tries currently to create a new volume on a pool that
does not support this API (aka iscsi) they will get an error suggesting
that the operation is not supported.

https://bugzilla.redhat.com/show_bug.cgi?id=1719556

Cherry-picked from master commit 34b13a.