ws: Make --for-tls-proxy and --no-tls options mutually exclusive
While --for-tls-proxy disables cockpit-ws' own TLS handling, it is
confusing to use this together with `--no-tls` as they are meant for
different use cases: Behind a TLS proxy, and behind a http proxy or
stanalone http server.

So make them mutually exclusive, and also drop the "implies --no-tls"
documentation, as that does not really explain anything.