machines: bring back machine wait=-1 noreboot options on virt-install in the VM creation
Commit d3ec4e8f2b did actually the exact opposite change, but time
showed it was a bad move.

We want virt-install to wait for the installation to finish since it
performs some cleanup jobs, such as removing possible cdrom devices that
were created for the installation (in the case were iso is used for
example).

Finally we want --noreboot option to prevent the domain from automatically
rebooting after the install has completed.

This option should be used only for when we expect installation to take
place. In case of importing existing disks to a new VM there is not
reason for these options, since the are installation specific.
Same for PXE booting.