machines: don't redefine the domain when user deletes it in the middle of the installation
Only try to redefine the domain when the installation failed and the
domain is appearing in the 'virsh list' command.
In case the domain is not present, the user manually deleted the domain
and in this case we should not undo the user's action.

Relevant to https://bugzilla.redhat.com/show_bug.cgi?id=1715399