containers: Bump unit-tests to Ubuntu 19.04
This tests on a less ancient toolchain, more recent libssh, and gives us
a more recent pyflakes which is able to spot a lot more errors.

We now use a i386 image for the i386 testing instead of crossbuilding
from amd64.  We use `setarch` to make sure we get the correct
personality on i386, even with a 64bit host kernel.

The main repo -dbg packages are mostly gone now, so enable the dbgsym
repository (as per https://wiki.ubuntu.com/DebuggingProgramCrash) and
install the corresponding -dbgsym packages instead.

Improve our use of 'n' and uninstall the system version of nodejs once
we have the lts version installed.

Use eatmydata to install the bulk of packages for faster container
creation.

Use a new setup.sh script to perform the bulk of the setup tasks to
avoid the gigantic (and growing) mess of shell in the Dockerfile.

Display some hints about the host and container distribution and kernel
versions at entry to the container, which might be useful when examining
logs of failed tests.

Finally, this commit also changes the process of building and running
the test containers, adding two new scripts: `build` and `start`.  With
the `start` script, particularly, it is now possible to run the unit
tests on a fresh checkout of cockpit without magical docker
incantations.  The old Makefile targets have been removed.

This commit is a combined effort, based on work from Martin Pitt.

Closes #11855