machines: on Stopped signal call Undefine action only if getVm failed
The only scenario during which we need to call Undefine action on Stopped
signal is for transient resources. For this scenario getVm will fail
since shutting off the domain will undefine it automatically.

This commit limits the cases where Undefine action is called. The
current behavior is not affected but less state updates leads to code
with less races.