networkmanager: Ensure that endianess is always set
Initializing the endianess in parallel meant that a lot of IP addresses
in the UI got rendered with `byteorder === undefined`, and thus being in
reverse order on big-endian machines.

Make these errors apparent also on LE machines by asserting that the
value is set correctly. In the UI, initialize byte order first and defer
everything else until that finished, via a new "preinit" promise.

Fixes #6041
https://bugzilla.redhat.com/show_bug.cgi?id=1728213
Closes #12382