sytemd: Dial down checking for available software updates
The PackageKit `GetUpdates()` call is apparently fairly expensive on
embedded systems like the RasPi (taking several seconds, as opposed to ~
0.3 s on laptop/server class hardware).

Now check at most once per hour, and after refreshing package indexes or
applying updates on the Software Updates page. Remember the state
through a sessionStorage variable.

This is a trade-off: This will now present out of date information if
software updates are being applied outside of Cockpit, or in the
Terminal.

Fixes #12413
Closes #12422