bots: Drop obsolete subscription copying in image builder scripts
Commit 5f6a1168bf dropped the consumption of the subscription
credentials from rhel.setup. Also drop the corresponding copying of them
into the VM, so that they don't stay around in the final image.

This also eliminates libguestfs for most images.  Even if we want to put
back subscriptions into our images some day, this doesn't require
libguestfs -- these can be copied in the kickstart file instead.

The only other part of virt-customize was the upload of the ifcfg
scripts. This is unnecessary as the kickstart already copies them.

Closes #12543