tls: Add test for blocked TLS handshake
Add a test to ensure that we can't wedge the cockpit-tls server by
blocking during the handshake phase (which could even theoretically
happen by accident if a connection is dropped, for example).

This was a bug before moving to the threaded connection model.  The
lock-up eventually cleared when hitting the gnutls handshake timeout,
but 40 seconds is a long time for the server to be blocked from handling
any other incoming requests.

Closes #12674