tls: Robustify /server/tls/client-cert-expired test
This test was brittle across different distributions, GnuTLS releases,
and architectures. We are not testing the specifics of GnuTLS here, we
just care about TLS requests failing with expired certificates. So don't
expect precise error codes at precise times, but just that the TLS
handshake or first read/write fails when expected.

Closes #12743