tools: Fix systemd dependency
cockpit-tls' .service file (commit 3f395e1054) uses RuntimeDirectory=
with creating a subdirectory. This has been supported since systemd 235,
so declare this dependency. This prevents people from trying to run this
on e. g. RHEL/CentOS 7, where the unit fails.

This is not strictly a build-time dependency, but bump it as well so
that the compatibility check is done for people who build from upstream
releases.

Fixes #12750