bridge: Fix valgrind memory leak in bridge's signal handler
Setting up the signal handlers before forking to start the ssh and
polkit agents causes the corresponding GSources to never be free'd in
the child processes. This is not a problem in practice as eventually the
agents will execve() -- but our unit tests are so quick that this
doesn't happen fast enough, and thus valgrind sees these GSources a
memory leak.

To mitigate that, set up the signal handlers after forking. Delete the
corresponding suppression.