Use Promise.all() instead of cockpit.all()
This was fixed properly in Cockpit 186. The earliest version that we
still support anywhere (RHEL 7.7) is 195, so it is safe to use
Promise.all() now. To ensure this, bump the manifest's cockpit
dependencies accordingly.

This still leaves some cases where `cockpit.all()` is being used to
deliver promises with a progress() method. These need to wait until the
pages get ported to React.

Closes #12795