test: Skip Machines selenium tests on Edge for now
Since PR #12573 (PatternFly 4 Tables), the Machines page crashes on the
old Edge 42 on our windows-10 image. Edge refuses to update on that test
image, so we currently have no way of testing with 44 (where this crash
does not happen). See issue #12864 for details.

Until we have a better fix for that, skip the Machines page tests on
Edge. We've ignored the Edge test for some weeks now, it's too dangerous
to introduce regressions to other pages now.

Closes #12880