machines: When the recommended resources values are below 1G convert the unit to MiB
To be honest I never noticed an actual Operating system which has
recommended memory or disk size below 1G so this was not affecting real
users. However the tests modify the osinfo-db data in order to allow VM
creation in low resource availability test environment so that was
spotted during testing.