tls: Run separate cockpit-ws instance for each client certificate
This got dropped with the move from direct spawning to systemd socket
activation in commit 4c40dad1a8.

Move the single cockpit-wsinstance-https.{socket,service} to a template.
Introduce cockpit-wsinstance-https-factory.socket that spawns new
instances of the former with each connection and tells the caller the
new Unix socket name to connect to.

Implementing the factory service is a bit involved: This could be a
simple `/bin/sh -c 'echo -n https@%i.sock >&3'`, but unfortunately that
doesn't have (or can be made to get) the `cockpit_ws_exec_t` target
context that the current SELinux policy expects. So re-use the
cockpit-tls binary for that with a new hidden --instance-factory-respond
option.

In cockpit-tls, bring back the TLS certificate → https ws instance
socket name map from before commit 4c40dad1a8, and use the instance
factory for each newly encountered client certificate. This is now
much simpler as before, as due to systemd's magic, cockpit-tls never
needs to consider or track instances going away -- even if the
cockpit-ws behind them times out or crashes, socket activation will just
bring it back on the next connection; and if there's something broken
with ws, the connection will just fail normally.

Adjust socket-activation-helper.c to provide two instances (this can
easily be bumped if test-server.c ever needs more), and a stub API for
the factory. This doesn't faithfully reproduce systemd's real behaviour,
but doing that is too involved -- let's rely on the integration tests
for that.

Add integration test for correct socket permissions.

Closes #12972