base: Remove target in Makefile for creating patternfly-additions.css
Previously we were duplicating a lot of CSS by including css code that
was already present from other files in the patternfly-additions.css file.

Note:
This commit moves around the less imports in patternfly.css.
There are two things we need to be carefull about:
* Make sure to always apply the variable overrides in the end.
* We need to remove the first occurrance of variables.less which is
imported twice (looks like that the fact that this file is imported
                multiple times leads to the first occurance to be
                applied in only)

Closes #13050