bridge: Avoid excessive duplicate queries for the D-Bus cache
Specifically, don't call Introspect for things that we already have a
reply for, and don't retrieve properties more than once per batch.

These changes were prompted by the flood of UnitNew and UnitRemoved
signals that systemd sometimes sends, such as during a reload.  When
such a flood comes in, we call cockpit_dbus_cache_poke on
o.fd.systemd1.Manager for each.  If at that time the cache doesn't
already have introspection data and property values for the manager,
it will call the Introspect and GetAll methods once for each signal.

The Introspect calls are done sequentially, but the GetAll calls are
done in parallel.  This might overload the "pending reply limit" of
the D-Bus daemon, at which point things might start to break.