bridge: Explicitly cancel fswatch GFileMonitor before unreffing
When GFileMonitor uses inotify, it asynchronously cancels the monitor on
disposing. This races with creating a new monitor somewhere else, and
frequently runs into deadlocks when the cancellation of the old and the
creation of the new monitor both try to acquire their internal mutex.
See https://gitlab.gnome.org/GNOME/glib/issues/1941 for details.

Interestingly there was a related bug in 2017
(https://bugzilla.gnome.org/show_bug.cgi?id=740491) for which we still
had a workaround in fslist. That exact same workaround applied to this
fswatch deadlock as well. However, the original bug got fixed long ago,
in all glib versions we care about. We don't need the main iterations
any more, just the explicit cancelling. So harmonize the workaround and
comment in both files.

Fixes #13146
Closes #13160