test: Check reading SMT status on Fedora CoreOS
On Fedora CoreOS there is no way to persistently change kernel options:
https://github.com/coreos/fedora-coreos-config/issues/234

But at least run the portion of the test that parses the current status,
and that it gives the expected error message.

Add checks for the `mitigations=nosmt` variant that is being used on
Fedora CoreOS. Our parser code already handles that correctly.