tls: Export current client certificates
Bring back the export of client certificates for current connections
into the runtime directory. This got dropped in commit 4c40dad1a.

Export them to `$RUNTIME_DIRECTORY` under their peer certificate
fingerprint name. That can be derived from the systemd unit (cgroup) in
which cockpit-ws and cockpit-session run. Use flock() to implement a
reference count on each file.

Each exported certificate file will only last as long as there is an
active connection with that client certificate. The authentication stack
can thus rely on certificates belonging to a current Cockpit session --
at least the websocket connection will be open all the time.

If cockpit-tls itself quits, systemd will clean out the
RuntimeDirectory= for us, to avoid getting stale data.

Drop the restricted RuntimeDirectory permissions, so that they use the
default 0755. This allows the unprivileged cockpit-ws to check if there
is a client certificate, and thus if it should request authentication
with that.

Add an unit test that opens 20 parallel connections, to ensure that the
certificate refcounting works and nothing deadlocks:

 - The first variant of the test only waits until after the TLS
   handshake. It can happen that one of the connections did not yet
   complete the certificate exporting part before the test starts
   closing fds.  In this case, when we go to check to verify that the
   certificate file "still" exists, it might simply not yet have been
   created. Thus wait for it to appear.

 - The "alternate" variant uses a new client certificate to connect to
   an "alternate mode" of the socket activation server which, instead of
   running cockpit-ws, writes a "hello" message and pauses.  We use
   reception of this "hello" message as an indicator that the service
   has been fully initialised.