tls: Fix disabled idle timeout
Initialize the idle timer fd to the expected "unset" value of -1,
instead of 0, as the rest of the code expects.

Let all unit tests except the explicit idle test run with a zero idle
timeout, so that it doesn't inadvertently stop the server when the tests
take longer than a second (as they do under valgrind).