cockpit/ws: Provide auto-start on boot with podman
When running with podman (not `atomic install`), make the "INSTALL"
label create a systemd service on the host that starts/stops the
container.

Don't use "podman generate systemd" for that, as that just stops/starts
the container. This both wastes resources when it's not running, and
seems to trigger some podman bugs with reattaching network interfaces
onto existing containers.

Closes #13210