tls: fix dup() bug in socket-activation-helper
The code here fails in case the dup()'d file descriptor ends up being 3,
because of the close().  Only do the dup2()/close() if the fd is not
equal to 3.

To be clear, dup2(3, 3) works properly: it's the close(3) after that's
the problem.  Since we're conditionalising the close() anyway, we may as
well also not run the no-op dup2() either (even though it's harmless).