cockpit-session: correct utmp entry writing
`LOGIN_PROCESS` seems to be intended to be used for login(1)-style
processes that are showing a login prompt on a given tty.  One can see
this from the current code in the output of `who -a`.

Use `LOGIN_USER` instead.

Additionally, set the "line" field to "web console" to provide a hint
about the nature of the login.

Add a testcase.

Closes #13242