test: Test OTP 2FA login with FreeIPA
This demonstrates how to set up FreeIPA to enable two-factor
authentication. On RHEL, this is the only supported method, as there is
no google-authenticator package.

Use HOTP instead of the default TOTP as that's more predictable for an
automated test.

This has worked for a long time already, but let's make sure it stays
that way.

Closes #13321