lib: Add React component for privileged UI elements
This provides a React counterpart to the jQuery `update_privileged()`
function. That isn't appropriate for React, as React renders stomp over
the jQuery updates. The generic `<Privileged>` component can wrap
anything, but isn't easy to use as it can't encapsulate the
`cockpit.permissions` logic -- there is no way to push the disabled
attribute to arbitrary child elements.

As most of our use cases involve link-type buttons, provide a
`<PrivilegedButton>` as well.

Use the "CPU Mitigations" link button approach, which works well. Refine
that a bit to not switch between button and plain span, but always use a
(PF4) Button and disabled it when necessary.