systemd: Fix Restart disabling when not privileged
The Button needs the `disabled` attribute instead of the class.
(Annoyingly, the Dropdown only works with the class). To fix the jQuery
vs. React race condition, move this to use the Privileged component.

Keep the data-stable attribute as the tests look for it.

This eliminates jQuery from the core overview page.

Closes #13328