tls: Support separate certificate key file
If a found ws-certs.d/ *.crt/*.cert file has a corresponding *.key file,
use that instead of expecting a merged private key in the certificate
file. This makes it easier to use Cockpit with LetsEncrypt or other
automatically maintained TLS certificates.

These days we mostly only need that for cockpit-tls. But fix it for
cockpit-ws as well, so that we keep remotectl happy (it complains
otherwise that it cannot find the key).

Drop the explicit `keyfile` parameter from `connection_crypto_init()`,
and let it try the two "separate key file" and "combined cert file"
cases by itself. Do the same EAFP approach in cockpit-ws.

Fixes #8922
Closes #13388