systemd: Replace OS updates health card entry with a page notification
In PageStatusNotifications, check notifications from the "updates" page
as well. On the Updates page, set the status to "Checking.." with a
spinner, instead of an empty notification, while updates are being
loaded.

Both of this replaces the rather elaborate hardcoding and code
duplication of OS updates in the health card with our standard
mechanism. It also avoids a forever-dangling "Checking..." spinner if
cockpit-packagekit is not installed.

Fixes #13400
Closes #13401