systemd: Respect page notification custom icon/link in Health notifications
If a page notification has a details.icon, show that instead of the
"type" icon. The Update's page's "up to date" status uses this (it
does not have a type as we don't want to show it in the nav menu).

Likewise, if a page notification has a details.link, use that for the
jump target instead of the page name. The Update's page's "not
subscribed" status uses this.

Also provide a way to not make a notification a link altogether, by
setting `link: false`. This is useful for transient "in progress"
notifications that don't require attention.