ws: write a login messages json blob to a memfd
Print out information about the number of failed logins, the last failed
login (if there was one), and the last successful login into a json blob
inside of a memfd which will be passed to cockpit-bridge via an
environment variable.

Currently, we only consider failed logins which have happened since the
last successful login, which means that if there was never a successful
login, we don't check for failed logins at all.  This might not be the
correct behaviour.