users: When entering user card clear up roles_changed status
This is set up to true when roles were changed for a user. However, when
you then go back to Accounts page and then to different user, it would
still be set up to true and showing alert that this user needs to
relogin as well.

Cherry-picked from master commit 675f2cdbfd7