test: Ignore iptables crash on debian-testing
This affects a lot of tests, and is too frequent and also too unspecific
for a naughty. See https://bugs.debian.org/951477 for details.

As our unexpected message filtering is line-by-line, this also needs to
ignore the whole stack trace. Make the patterns specific enough to not
match on just any stack trace.

Closes #13560