Replace PF3 buttons with PF4 version
Most of the changes here were automatically done with sed to replace:
* the PF3 button classes to PF4
        - .btn => .pf-c-button
        - .btn-default => .pf-m-secondary
        - .btn-primary => .pf-m-primary etc.
* The React component module import from patternfly-react to
@patternfly/react-core

For consistency all Cancel buttons in the dialog were adjusted to be
pf-m-secondary, some of them used to be pf-m-primary.

Where buttons had directly applied pficon or Font Awesome classes they
had to be moved in a span or icon element as child of the button
element.

Closes #13612

Remove button-sm from several lists