test: Fix race condition in TestFirewall.testNetworkingPage
firewalld and libvirt start up in parallel, and libvirtd asynchronously
adds its zone into firewalld after starting up. So wait until it did
that, so that we get a stable zone count.

Otherwise it often happened that the initial `--get-active-zones` did
not yet include the libvirt zone, and the later assertion about what's
shown on the Networking page failed due to a number mismatch.

Conditionalize the waiting on libvirt being a new enough version to add
its own zone.

Closes #13619