test: Fix cleanup of scsi_debug
add_ram_disk() interfered badly with the clenaup centralization in
TestMachines, as with `addCleanup()` it tried to remove scsi_debug
before running the TestMachines cleanup handlers. This caused the device
to be busy as it was still part of storage pools.

Move the scsi_debug cleanup to MachineCase.nonDestructiveSetup(), so
that it happens as (one of the) last things.