systemd: Port to PF4-React Button
The "Instantiate" button for template systemd units in the service
details was previously completly unstyled. Move it to a list group
(i. e. line) by itself, as otherwise it's directly attached to the input
without any margin. There's nothing else on the page anyway, so move it
below. This is more mobile friendly, too.

Keep the following manual `pf-c-button`s for now:

 - in terminal.jsx, as `<Button>` does not understand `ref=`.
 - in pkg/systemd/overview-cards/motdCard.jsx, that's the close button
   for the alert. This will better be rewritten using the actual <Alert>
   component.

Closes #13670