packagekit: Use our EmptyStatePanel component
This will move the "Ignore" secondary action in the "Ask for restart"
panel below the primary button. This matches the PF4 design. Drop the
`&nbsp;` hack to put it next to the primary button.

Also, stop putting the "Package information" list there inside of the
empty state component. It's not necessary (it looks almost the same) and
simplifies the component structure.