build: Consistently use absolute source and build dir
`$(abspath $(srcdir))` and `$(abs_srcdir)` don't do the same thing if
there are symbolic links in the path. In particular, if /home is a
symlink to /var/home (like on OSTree), the former is
`/var/home/<user>/...`, while the latter is `/home/<user>/...`.
This caused broken tarball builds.

Consistently use `$(abspath)` so that the values agree on the same
prefix. Unfortunately webpack insists on getting absolute paths.

Closes #13735