accounts: Properly dismiss tooltips
The tooltips were placed at the end of the body element on hover. When
someone clicked on the checkbox, that triggered actions and the groups were
re-rendered. By re-rendering, the page lost the context of which tooltip
belongs to which element (as those elements were created from scratch).
That kept a dangling tooltip at the end of the body after each click on
these checkboxes.

With this patch the tooltips are rendered in the same DOM element, so when
groups are destroyed, those tooltips are destroyed as well.

Fixes #12262
Closes #13764