machines: Remove --noreboot parameter to virt-install for VM installation
We were passing this parameter so that when the VM installation finishes
it won't automatically reboot. It will stay alive and if the user tries
to reboot it, it will just shut off, because the installation VM config is
always set up with <on_reboot>destroy</on_reboot>.

Some more info for understanding the virt-install logic here:

When creating a new VM through virt-install we have two steps:
1) the install phase, where the guest is set to boot off the fetched/supplied kernel/initrd
2) the post-install phase, where the guest is set to boot off <disk>

In order for the post-install config to actually take effect, the VM needs to be
fully powered off, after which libvirt will replace the 'install phase'
config with the 'post-install config'.

So, virt-install will force on_reboot=destroy, so that if anaconda
(or other installation program) requests a reboot, libvirt will instead
fully power off the VM, so the post-install config can take effect.

This patch allows virt-install to have this default behavior.

Note: the tests had to be adjusted because of:
https://bugzilla.redhat.com/show_bug.cgi?id=1818089
Which causes virt-install to restart the VM when it's stopped during
installation.

Closes #13795
Relevant to: https://bugzilla.redhat.com/show_bug.cgi?id=1750637