systemd: Fix showing the "More" diagnostics in realm join dialog
Commit d4b4c7d340e was missing the realms-op-diagnostics field, which
also uses "hidden", and thus never appeared when calling .show(). The
effect was that clicking on the "More" link to get the verbose
diagnostics just made the whole error message go away.

Change the "hidden" instead of the "display" attribute to fix that. Do
the same for realms-op-error for consistency.

Add a test for the error handling: Initially we just want to show the
error message and the "More" link, and when clicking we want both go to
away and show the verbose "diagnostics" log.

Part of https://bugzilla.redhat.com/show_bug.cgi?id=1813136

Closes #13802