selinux: Offer automatic applying of sebool fixes
setroubleshooter's catchall_boolean plugin often offers multiple
solutions that the user has to choose from, so due to the API being
restricted to only auto-apply a fix for a complete violation per plugin,
the catchall_boolean plugin does not offer automatic fixing.

However, it seems both safe and very convenient to provide an UI to run
the provided `setsebool` command. Special-case this (it can be extended
for other safe commands later on) and run it through the command line.

The existing `SELINUX_ALERT_SCRIPT` in the test fits right into that use
case. Refine the test to tell apart the auto-fixable sebool solution
from the non-auto-fixable generic "file a bug" solution, and make sure
that the sebool solution has the desired effect. Rename the two alert
script variables, as they are both fixable now.