test: Sort non-destructive tests
Running tests in a more or less random order has already bit us more
than once (at least PR #13843 and #13598): A new test that succeeded in
a PR just because it happened to run late, and then it fails later on
when it runs earlier.

At the same time we don't just want to test a single ordering, as that
would hide accidental test dependencies.

Sort the tests ascending or descending based on an one-bit hash of
`$TEST_OS`. This will evenly distribute both directions and still be
robust and reproducible.

This requires some re-engineering of check-testlib, as this now causes
the expected strings to change depending on the image. For the tests that
expect a particular oder, override TEST_OS, and test both directions.

Closes #13844