logs: Enable to stop following (and start it back up)
When users examine logs they may need/want to stop new messages from
appearing. This is especially useful, when they inspect some logs from
lower priorities and when the even already happened.

Also enable the users to start following again. This will also read
messages that happened from the last visible entry.

Also it is possible to pass `follow=false` in the URL to prevent the
following in the first place.

Fixes #13830