tools: When filtering manifest files look for '.in'
In language files (like 'de.po') we refer to manifest files like
`pkg/sosreport/manifest.json.in`. This is later on used with `msggrep`
to only include relative files. Looking for `manifest.json` did not
return any results and therefore we would not translate things in shell
like labels or help menu items. (We would translate it only if the same
string comes also from different location, which was the case for
majority of label as <title> usually has the same string.)

Cherry-picked from master commit 1cb74637a8e7