test: Fix Machines testNICAdd permanent attachment
Specify a type and source in the "permanent attachment to running VM"
check. Otherwise it's going to default to "Direct attachment" and try to
use whichever real interface happens to be first. This is almost
certainly going to fail, as e. g. "lo" can't be attached (resulting in
"invalid argument") and vnet0 or most other interfaces are already busy.

Closes #13907