pkg: Fix cockpit-components-log-panel some
When limiting the log display, we want to cut off at the end, not the
start.  Ouch.

We want to keep "max + 1" entries since the day header doesn't really
count.

Using __MONOTONIC_TIMESTAMP for the key doesn't work since it is not
guaranteed to be unique.  Let's use __CURSOR instead.

Putting React nodes into the component state is perfectly fine, just
like they are fine to be put into properties.

Closes #13914