test: Make network team tests nondestructive
Very similar to the bond tests in commit 37127ec6196f7. Except that for
creating a team it's not possible to specify/copy the underlying
device's MAC address, so we can't expect to get the same IP.

Adjust the test to just check for a valid DHCP IP. This was previously
hidden behind known issue
https://github.com/cockpit-project/bots/issues/66 , but this was never
really meant to apply to this particular scenario.