systemd: Add instructions how to test Active Directory
The "services" image recently grew a Samba AD server [1], document how
to use it.

Manually testing Cockpit, realmd, or any other part of the OS against
Microsoft Active Directory can easily be done with an AWS managed domain
controller, document that as well.

Also add the missing /run-freeipa step that was introduced in
https://github.com/cockpituous/bots/commit/9ebea751e

Closes #13920

[1] https://github.com/cockpit-project/bots/commit/b13ed9db