test: Add initial Active Directory tests to check-realms
Re-use the common TestRealms tests.

This currently uses a custom sudo rule for giving sudo access to domain
admins. This is not very elegant yet, but initial search all pointed to
this method. There ought to be a centralized method on the domain server
side, similar to `ipa-advise enable-admins-sudo`.

Adjust the expected error messages in testUnqualifiedUsers() when giving
a wrong password, as they look different when joining AD.

This also detects the AD joining bug fixed in PR #13803, so skip
testUnqualifiedUsers() on rhel-8-2-distropkg.

Closes #13921