test: Add workaround for broken sudo with password with AD user
Starting the root cockpit-bridge fails when sudo asks for the password
of an Active Directory administrator. As a workaround, to run at least
the other parts of the tests, configure sudo to not require a password
for domain admins.

See <https://bugzilla.redhat.com/show_bug.cgi?id=1825749>. Revert this
commit once the bug gets fixed.