services: Reduce the API calls that the initial load of Services page does
For displaying the services list we can just rely on the results of
the `ListUnits` and `ListUnitFiles` API calls, without calling GetAll
properties for all units.
Only when a Unit details page is loaded, we should query the rest of the
properties.
That should reduce the amount of API calls that are happening during
`Services` page initial load.

For the sake of this simplification, we needed to drop the tooltip on
static services state, which specified which units require that static
unit, since for that property we need to call the expensive 'GetAll'
that we dropped in this commit.