containers: Move unit-tests to Debian stable
Ubuntu 19.04 has been EOL for over a year. The current versions (≥
20.04) don't support i386 any more (but we still want to run tests on 32
bit), and also don't include chromium any more (it's a snap now), so
move to Debian stable instead.

This also allows us to go back to the distro version of nodejs.

Move turd-polish to python3.

Drop a few build dependencies which have not been used in ages.

Closes #14852