tools: Add watch mode to webpack-make
Watch mode makes fast iterations much faster. Compared to plain `make`,
watch mode and disabling eslint reduces e. g. tuned from 19s to 3s, or
systemd from 48s to 17s on the first rebuild, and 5s (!) on the second.

Having to specify `-d` that way is still a bit ugly, but this will be
addressed in a subsequent commit.

Mostly drop the obsolete and duplicate instructions from
tools/README.dist. Calling plain `webpack` is utterly broken and not
supported. tools/webpack-make is easy enough to call now to make that
the official interface.

Closes #13975