tests: (TestService) wait for the reload to finish before executing other unit actions
When we enable/disable units and the daemon-reload takes place we fetch
information about all units from scratch (listUnits). We call setState to update
the unit's object in the state only if the properties have changed. However if
a JobNew/JobRemoved signal is received after we got the results from
ListUnits but before we updated the state, we might run in an incorrect
update, where the event handler from JobNew/JobRemoved will update the
state correctly, but then the listUnits will update the properties back
resulting in a wrong state.

This is an actual issue but I don't have a good solution for this, so
let's make sure the test won't run in this race.