Fix keyboard navigation
This rule is actually very useful, as it points to things, that are not
keyboard navigable. We were introducing new breakages since we disabled
this rule (like services and overview pages).

Here is list of things that did not work before, but do now:

- service-tabs.jsx: The whole navigation (targets, sockets...) was not
reachable by keyboard.

- page-status.jsx: Items in 'Health' card were reachable by keyboard,
but could not be activated by keyboard.

- usageCard.jsx: Text 'View graphs' on 'Usage' card was reachable by
keyboard, but could not be activated by keyboard.

- systemInformationCard.jsx: Text 'View hardware details' on 'System information'
card was reachable by keyboard, but could not be activated by keyboard.

- insights.jsx: Text 'Not connected to Insights' on 'Health' card was reachable
by keyboard, but could not be activated by keyboard.

- storage-controls.jsx: The 'things' menu on storage page could be
opened by keyboard, but the elements inside were not reachable by keyboard.

- overview.jsx: On storage page non of 'Devices', Drives'... and any
other devices were not reachable by keyboard. Now they are reachable and
also it is possible to activate them by keyboard. (This includes also
'Show all' buttons)

- diskEdit.jsx: These prevents were pointless so I dropped them.
Otherwise we would need to do the same for keyboard events.

- containers-view.jsx: Items could not be activated with keyboard.

- cockpit-components-listing.jsx: Give the event listener to appropriate element.

- fsys-tab.jsx: This is actually unused but making it link-styled button
seems like a correct approach to make it accessible.

Part of #12954
Closes #13996