test: Avoid libvirtd crash in TestMachines.testLibvirt
libvirtd often crashes when stopping it right after starting [1]. As a
workaround, be a little less demanding, and wait for domain interfaces
to have initialized before stopping the service again.

Also make sure that the page recognized that the service started up
again and the empty state goes away.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1828207

Fixes #12978
Closes #14043