packagekit: Refine reaction to superuser change
With the PackageKit client reconnecting automatically,  we don't need a
full page reload any more, as there are no other permanent connections
to system APIs.

However, the page reload had the side effect of getting out of the
load/update error screen. This is common when trying to check for or
install updates when unprivileged, and polkit rules don't allow this
action to the user. It makes sense to reset the view when switching to
privileged mode, so add that special case.

We don't want to hardcode `superuser: true` to these operations, as this
is a common polkit rule customization. This should be replaced with a
proper polkit privilege check in the future.