tests: robustify TestMachines.testDetachDisk
When live-detaching disks the guest OS needs to cooperate so that we can
see the disk getting detached in the UI.

For this reason adjust the check in the test before detaching the disk,
so that we wait until we see the user login-prompt. At that point the OS will be
fully responsive.

In addition, increase the timeout, so that the test will not be flaky
when under IO stress in the CI.