common: Fix deadlock in channel pressure handling
When a channel applies back-pressure to its input because its output
queue is getting too large, it relies on the consumer to acknowledge the
reads by ponging to the pings that the channel sent out.

However, it can happen that the last ping was sent for a sequence number
*before* the one that the channel expects for lifting the pressure. Then
the receiver could drain the output queue and dutifully respond to all
pings, but the last sequence number would still be smaller than the
treshold for lifting the pressure. But the channel isn't sending
anything else either and thus never generates new pings.

Fix this by sending an additional ping when going into pressure mode, to
ensure that the last pong sequence number is within the out_window
again. Ensure that this only happens on edge trigger, so that we never
send out the extra ping more than once. This also avoids raising the
pressure-on signal more than once as a side effect.

This can also be reproduced/investigated with this test-spawn-proc unit
test:
```js
QUnit.only("stream large output - find", function (assert) {
    const done = assert.async();
    assert.expect(2);

    cockpit.spawn(["find", "/"], { err: "ignore" })
            .stream(function(resp) {
                console.log("got block", resp.length);
            })
            .then(function(resp) {
                assert.equal(resp, "", "no done data");
            })
            .always(function() {
                assert.equal(this.state(), "resolved", "didn't fail");
                done();
            });
});
```
But this is not appropriate for actually committing -- the find often
fails at the end, and takes too many IO resources for a unit test. The
integration test confirms the fix, though.

Fix the integration test to happen before resetting the terminal, as the
subsequent checks expect it to be empty.

Fixes #12580
https://bugzilla.redhat.com/show_bug.cgi?id=1751783

Closes #14079