test: Adjust Machines tests for test VMs with 1 GiB RAM
Some Machines tests check adjusting the guest RAM size to the maximum of
what the host has available. The formatting of the usage donut then
depends on whether that results in a size < 1 GiB (in which case the
usage will be formatted in MiB) or > 1 GiB (then it will be GiB). Thus
replicate the cockpit_format_bytes() logic in the test to work with
testbeds that have more RAM.

This doesn't affect the creation dialog itself, just the donut check; so
move the text computation logic out of the VmDialog class.

Cherry-picked from master commit 87c18da7968ee.