test: Check overlaying cockpit-ws RPM on Fedora CoreOS
This didn't use to work in Atomic times, but does now -- so let's keep
it that way. So un-skip some check-connection tests on fedora-coreos and
instead have them disable the container, rpm-ostree install cockpit-ws,
and use that instead.

Splice testBasic() into a "normal" mode (using the container on ostree)
and an "ws overlay" mode with the same test.

The only problem is an SELinux denial. Let's figure this out later and
disable SELinux for the time being.

Generalize the dbus-daemon message hack in testlib.py to apply to all
Fedoras. One of these days we need to fix this properly..

Closes #14211