common: Set strict Cross-Origin-Resource-Policy
There is no good use case for including parts of Cockpit's resources
into websites of a different origin. Cockpit does not expose any images
or scripts which would work or make sense there. Thus set the
`Cross-Origin-Resource-Policy` header [1] to `same-origin` for both
static and authenticated paths.

Do this even for "ping", as this doesn't matter for that API -- CORP
leaves the complete response intact, it just instructs the browser to
discard the response contents.

The only cross-origin use case that we have is to embed Cockpit frames
into other websites. The CORP header does not affect `<iframe>`, so this
continues to work fine.

[1] https://developer.mozilla.org/en-US/docs/Web/HTTP/Cross-Origin_Resource_Policy_(CORP)

Closes #14215