tls: Add test/demo CA
Add an OpenSSL configuration and generate.sh script to build a CA and
two user certificates that are signed by that CA. This is useful for
both automatic and interactive certificate authentication tests.

Use these pre-generated certificates in
`TestIPA.testClientCertAuthentication` instead of dynamically generated
ones. That makes manual testing easier, as the certs don't change all
the time; and they are also more representative as they are CA-signed.
Don't generate/copy a .der format explicitly, as a base64-encoded DER is
just the PEM with the `----BEGIN/END` lines stripped off.

Also use these in test-tls-server, instead of mock-client-expired.crt
and alternate.{crt,key}.