test: Fix race condition in TestSOS.testCancel
Clicking the Cancel button merely .close()s the sosreport process, it
does not synchronously wait for its termination. So  it is often still
around right after that. Give it 10s to terminate.

sosreport with the full configuration takes about two minutes, so the
chances of this catching the regular (non-cancel) end is negligible.

Closes #14364