test: Drop selenium kdump test
Despite doing a lot of stuff, it does not actually validate much in
Cockpit. The NFS server setup  is in vain (as it just checks the
generated kdump config file), and the sysrq-trigger is a practical
no-op. The verify/check-dump nondestructive test has similar coverage
and supersedes this.

This gets rid of depending on kdump.service succeeding, which is not
currently true in Fedora due to
<https://bugzilla.redhat.com/show_bug.cgi?id=1859575>. But this breaks
*all* selenium tests.

Closes #14401