tests: Robustify TestMachines.testCreateThenInstall
This test suffers from the following race condition:

* A VM installation is running
  (virt-install is running through pkg/machines/scripts/install_machine.sh script)
* The test clicks 'Force Off' button from the UI which makes virt-install
  command exit and install_machines.sh script should continue to try to
  re-define the domain in order to inject the metadata.
* Then the test tries to delete the VM and expects to see an empty list after
  the delete operation

-- Race condition --
Sometimes the deletion was processed before the install_machine.sh
script redefined the domain, resulting in non empty list after the
delete operation.

There is nothing really to be fixed in the source code, this test just
asks for problems.
Robustify it by waiting for the metadata of the VM to get set, which
ensures that install_machines.sh script has finished.

Cherry-picked from master commit fba9a531.