test: Define 'default' network before trying to start it (#14484)
`startLibvirt()` tries to start the `default` network, but if it was
removed by some previous test it fails. The line I moved was introduced
in #14404 with exactly this same intention, but it needs to happen
sooner.

Hopefully fixes https://src.fedoraproject.org/rpms/cockpit/pull-request/29

Cherry-picked from master commit 861472dba