machines: IP should not be same as network or broadcast.
IP address value should be between network address and broadcast address.
The IP address used in the configuration is automatically set to the bridge
interface and is used by dnsmasq to listen on for DNS and DHCP queries.
In case DHCP is used, then invalid gateway is provided to VMs because they
already have a route containing the Network IP so a default gateway cannot be
added with the same value.