machines: DNS PTR record for this virtual IP network
Treat all DNS PTR record requests for this virtual IP network.
Without this attribute only known IPs from the virtual network will be resolved.
Unknown IPs will be forwarded to external DNS, but since we are in a private
network all IPs belonging to the virtual network should be treated only by the
local virtual DNS.

As example consider "192.168.100.0/24" network IP with DHCP enabled would
register automatically the localvm with "192.168.100.123" IP as
"123.100.168.192.in-addr.arpa." in the local DNS, any other query like
"192.168.100.234" would be forwarded to the hosts DNS server
(ie. router, ISP, Google, TLDR) as unnecessary traffic.

On the flip side, we may use a network name IP as "8.8.8.0" which would resolve
localvm with "8.8.8.1" IP as "1.8.8.8.in-addr.arpa." but would forward
externally "8.8.8.8".

Closes: #14564