all: drop our journald logging handler
GLib grew support for logging to the journal in 2.50.  cockpit's own
journal-enabled logging handler has some features not enabled by default
in the GLib one (such as the ability to log to the journal, even if
stderr is on a terminal), but these minor functionality changes are
probably still worth dropping this huge chunk of code.

One feature that we don't want to drop: in case we ssh from one cockpit
host to another (including the bastion host case), we want to ensure
that any log messages or stderr output stays on the remote host, and
doesn't get sent via ssh back to the client side.  We can accomplish
that with a very specific and small piece of code only in
cockpit-bridge.

Closes #14390