build: change our GLib dependency logic
Adjust the way we handle our glib version macros in configure.ac in
order to be able to depend on a newer GLib version (2.50) without
causing problems with deprecation warnings (due to our continuing use of
GSimpleAsyncResult).  Additionally, we drop the separate GIO version
number, because it's already much lower than our required GLib version.

Make sure we bump the version numbers in our distro packaging to match.

We require GLib 2.50 for the journald logging support depended on in the
next commit.