test: Wait for firewalld service state
There seems to be some issue with this part of test where it times out
waiting for the change in the UI. There are 3 possible explanations:
1. The service fails to stop/start
2. There is no event delivered (or is delivered but we process it wrongly)
3. It just takes a bit longer to propagate in the UI

This PR does not fix any of these issues (possibly the last one) but it
should point if the issue is no. 1 or no. 2. Then we can do more
debugging.

Closes #14611