jsonprint: fix escaping again
More unit testing discovered that we were overlooking the ASCII DEL
character (decimal 127, hex 0x7f) and writing this non-printable
character directly into the output.  There's nothing theoretically wrong
with that, but let's not print a non-printable character.

There's no longer a benefit to using signed characters, since 0x7f is a
positive value in any case, so switch the logic to unsigned to make it
more understandable.