test: Fix check-embed to work with chromium>=85
The default "Referrer-Policy" for chromium is now
strict-origin-when-cross-origin, which caused the previous version of
the test to fail as js from the local embed-cockpit.html file to the
iframe was considered cross-origin.

By making the embed file the default shell, it is no longer
cross-origin.

Cherry-picked from master commit ca1f9787

Closes #14743