machines: fix testSnapshotRevert flake (#14801)
* machines: fix testSnapshotRevert flake

The flake was caused by the fact that snapshot create commands would
get spawned at the same time, so the creation time would be same.
When the snapshots have different time creation, they are sorted
in descending order. When they have same time creation, they are sorted by
name, which is in ascending order. This causes a race condition. Fix
it by forcing a different time of creation.

Also change snapshot names for code clarity:
snapshot0 to snapshotA
snapshot1 to snapshotB