storage: fix position of spinner it dialogs
Now the footer element from PF4 is styled as `flex`, therefore `pull-right`
class was not taking effect.

Make sure that `dialog-wait-ct` which is the wrapper element for
spinners and idle messages, is displayed in the same line with the
buttons.

Also make sure the the spinners always appear in the right of the
messages in footers, previously some dialogs had them on the left.
Closes #14815