machines: Use MiB for memory/storage if recommended value is not a whole number
By default GiBs are used for memory in 'Create VM' dialog. However if the recommended memory value is
floating point, ex. 1.5GiB, it's floored to 1GiB which then results to error
and user having to fix it manually. Prevent this from happening by using MiB in this case.
Fix it also for storage.

Note: Use fake libOsInfo values which convert to a whole MiB number

Closes #14770