shell: Implement navigation and host switching in more sensible way
There was this weird event listener in `pkg/shell/base_index.js` that
was listening on events from react components created in a completely
different place. It was very hard for other maintainers to find where
this is being handled.

Also this function had a bunch of hacks like ignoring `button`s clicks
or we had to have `event-eater` hack.

Also it was not working ideally as the hack with `button` did not work
out as expected when we clicked on the icon in that button.

This patch just does it "the right way" where we handle events on
appropriate elements when we create them. This does not need any of
those hacks and works more reliably not mentioning it is much better
readable.

Closes #14848