users: Enable to delete the whole user name
It was not possible to delete the last character of the name. If you
wanted to change "foo" to "bar" you could not delete "foo" and then type
"bar", but had to delete just "oo", type "bar" to get "fbar" and then
move cursor to the start of the string and remove the "f".