bridge: locals cleanup in cockpitinteracttransport.c
Do some more g_autoptr()-based reworking to fix a Coverity defect
report.  In this case, it's a false positive (caused by Coverity not
understanding that the g_bytes_new() call in parse_frame() can never
return NULL), but this is a nice cleanup anyway and will avoid the
warning without having to add extra hacks or annotations.