Revert "base1: Fix race condition in "watch barrier" D-Bus unit test"
This was a misunderstanding of what the test was aiming to do -- the
whole *point* was to create a "race" between setting up the watch and
doing the call, and ensuring that D-Bus does a proper ordering between
the watch and the call.

That change also did not help at all to fix the test.

This reverts commit 3317a06cd851ff0979a0deb37c353da585b07e16.