base1: Fix D-Bus "watch barrier" test
In that test it could happen that after the first `notify` signal with
the expected properties there is a second `onnotify()` invocation with
an empty property update dict (which is unnecessary, but formally
conforming with the spec, and mostly harmless).

I broke that in commit 01317a769d, which replaced a deep update of the
cache object with a shallow override: I wrongly thought that there
should/will only ever be one invocation.

Go back to a "deep update" to accomodate multiple notify signals,
similar to how the "watch interfaces" test works.

Fixes #15095