check-dist: don't try to open /dev/stderr
Running this script could fail because /dev/stderr could be a symlink to
a socket or a file that we don't have permission to open().  In the case
of running it under toolbox, for example, it's a link to a pty which is
inaccessible to the current user.

In the case that the check fails, just run the command again to get the
debugging output.