terminal: utilize Toolbar PF4 component for the terminal header
This allows us to completely drop PF3 imported CSS.
This commit also makes the terminal work as expected in small screen
sizes, whilst previously the toolbar section was overflowing.

Note: <Select> component from lib, does not have a className property.
'theme-select' class was never applied so this commit just dropped the rules
for it as well.