machines: properly show a dummy VM in the VM details during the VM installation
Internally, when the user clicks the 'Install' button for a VM we
`undefine` it and run the `virt-install` script which will shortly
`define` it again and then run the installer.

Therefore there is this short period, that the VM being installed is not
listed by the libvirt API.

In order to give the user a good UI experience we had invented in the past
the concept of *dummy* UI Vms, which we inject in the list of the Vms the
API gives us in order to make sure that the user still sees an entry in
the VMs list while virt-install is performing the initialization
process for the installation.

These dummy UI VMs contained so far just a VM name and a connection type,
but this commit extends the dummy VM object to contain all details in the
case of the installation being seperate from the VM creation. That
allows the user to still be in the VM detals page, even when the 'real'
VM is not listed by the API for a few seconds.

Related to rhbz#1916120