lib: drop our custom select implementation in favor of PF4 FormSelect
The FormSelect does not support inline cursor dividers, just groups,
so where the divider was used before it was either just replaced with
FormSelectOptionGroup or dropped completely.

Secondly, FormSelectOptions do not support option descriptions as our select
did.
Our custom 'Select' was using the 'title' attribute which is actually
not valid child of the 'select' element and therefore should not be used.

There were luckily just two usages of that problematic 'title' property.
These were:

* VMs: PXE boot entry used 'title' to show why it's conditionally
disabled. I just removed it when it was not available instead of disabling it

* VMs: Some pool types don't support storage volume creation and were
appearing in the dropdown but disabled with description.
These will now keep appearing in the dropdown, I can't think of a better solution for this.

The tests were also affected because the previous select was exposing
the data and data-label attributes which are not used any more.