lib: Utilize TableComposable instead of Table from PF4 for more control over properties
The motivation behind this:

Our tests have spotted that the dialogs whose code is part of table
cells close sometimes unexpectedly. That is the reason behind some flaky
tests, like TestMachines.testDiskEdit.

This issue is following us for a while and was in the past addressed by
commits like 81267a676a, which tried to keep the status of the modal
outside of the react component rendering. Unfortunaly this workaround
did not count with the fact that even if we maintain the open status of the
modal in a static way when a component gets unmounted and re-mounted the
state get's reset anyway, resulting in all the fields that the user
already filled by that point to be lost.

So we really need to ensure that there is not unexpected unmounting of
react components.

The reason behind this behavior was that the table row cells did not
have the expected unique keys that React uses to identify components.

This commit adds the necessary 'key' attributes to solve the above
mentioned issue, but in order to do that the whole implementation of our
Table component needed to be re=done, since the Table PF4 component is
too implicit and does not allow us to control cell keys.

Note: this commit moves only the basic lists to the TableComposable
component, the selectable, expandable, sortable and any other magic
tables stay with the old implementation, just because rewriting will
make this commit too complicated.