lib: workaround missing dataLabel property in <Th> PF4 component
When we use <Th> as cell elements, the dataLabel property is not applied
correctly and we got a warning in the console. In addition in the mobile
mode the missing `data-label` was causing the cell labels to not exist.

Use the original `data-label` property to fix the visual regression until it's fixed upstream by:
https://github.com/patternfly/patternfly-react/pull/5406