Build `make dist` artifact for every master commit and PR
The npm install/webpack steps are very expensive, and annoying to do
locally for reviewing PRs or updating a local checkout. Introduce a new
"build-dist" workflow which creates a downloadable dist tarball for each
interesting commit (PRs and pushes to master for now).

Split out the building part from scenario-distcheck.sh into a new "dist"
scenario, so that we can finish this workflow in the least possible
amount of time, as we soon want to trigger the tests after this
workflow, to re-use the built dist tarball.

Skip building the cockpit-cache*.tar.xz for this workflow -- we don't
need it for CI and it's expensive to build.