build: Drop gzipped dist tarballs and make-source usage
Since the introduction of `$XZ_COMPRESS_FLAGS` in commit 7a9870cbef and
supressing the build of cockpit-cache-* tarballs in commit 0b9ecf7ea
there is no reason to build a .tar.gz dist any more, as `xz -0` is even
faster than gzip.

Drop the `dist-gzip` make target, adjust the "dump file name" target for
xz, and switch the .spec and image-prepare to using the tar.xz
targets/tarballs.

bots/make-source defaults to `dist-gzip`, is rather hard to work with
for dist-xz even with a custom target, and most of it is a big cockpit
special case (and not used in any other project). Let's phase it out,
and just do the few build steps ourselves -- these are project specific
anyway, so let's not pretend we can do an "one size fits it all" script.
While at it, replace make-source's static `-j8` with a dynamic CPU
count.

This unblocks building images straight from the pre-generated dist
tarballs in GitHub action artifacts.