build-copying: adjust to not disting node licenses
Adjust the build-copying script to continue to directly scan the
node_modules that we do ship, but also scan (from the source directory)
the node modules that are included in files we ship in dist/.

Fix a couple of bugs along the way (repeated entries were being included
due to some incorrect quoting), and use variables and a loop to avoid
running half of the script in a subshell.

Also, move the rules around a bit to make sure that the README and the
dist/ directory are definitely in place before running the script.
Deconditionalise the inclusion of the README, so that we flag an error
if it's missing (which it was before, due to it being copied after the
script ran).