machines: correct few details in storage page when pool's deactivated
It's impossible to create a volume when a storage pool is deactived.
But right now, user has to fill out the form only to find this out.
Also when storage pool is deactivated, the "No storage volumes defined
for this storage pool" string is shown in volumes list. That's incorrect.
There are no volumes because libvirt API cannot fetch any information
about volumes for deactivated storage pool.
Therefore to fix both problems, just hide the storage volumes list if pool
is deactivated and instead show a warning telling to user to activate the
pool