storaged: Replace --with-vdo-package configure option with manifest map
Ideally we'd ask the system if there is a "vdo" package available, but
PackageKit takes too long to figure this out if it has to first download
the entire package index. So for now, continue to decide based on the
host OS (RHEL/CentOS), but move the decision from ./configure to a
per-OS map in the manifest. Do this generically by teaching get_config()
to handle OS → value maps.

This gets rid of one part of dynamic manifests, and thus makes them more
portable (~/.local/share/cockpit) and easier to generate by webpack. It
also avoids an "unbreak my build" configure argument for CentOS/RHEL.
The manifest is a much more "quirkable" text file which can be handled
through overrides by admins who know what they are doing.