apps: Replace --with-appstream* configure options with manifest map
This gets rid of the remaining piece of manifests which depend on
./configure options. This gets us closer to making them fully portable
and buildable by webpack.

This uses the same approach as the storage page in commits
0f0744be3fceb and 8844634cf7e: Put a per-distro map into the manifest,
which is easy enough to tweak/override by admins, and patch by
distributors. Borrow and adjust the get_config() function from
pkg/storaged/utils.js for that.