lib: Check installed Red Hat products instead of yum config file
The `enabled=` option in subscription-manager.conf is not a reliable
indicator whether the installed system needs any subscriptions. On at
least the qcow image, the dnf plugins (both subscription-manager and
product-id) are disabled by default, and the first invocation of
subscription-manager auto-enables them.

Instead, ask RHSM if there are any installed Red Hat products on the
system. That list is empty on Fedora/CentOS. As before, if RHSM itself
is absent, quietly consider this as "system does not need subscription".

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1931429