bridge: Use json template expansion machinery
Remove the explicit management of the envp and argv arrays and replace
it with a simple JSON template expansion.  This allows us to use the
existing JSON hash and equality functions (instead of rolling our own)
and seems to be a good fit, since we were always converting the data
back into a JSON blob again anyway to pass it on to CockpitPeer.

It will also allow us to add an additional level of transformation in
the near future in an obvious and clean way.