test: Fix retry reason printing in run-tests
Commit 8aeac4d7ad changed the retry_reason printing to happen after
flushing stdout. This kept it stuck in Python's IOStream buffer until
more output (from the next few tests) was printed, and thus eventually
put the marker onto the wrong test.

Fix this by flushing after the retry reason in the `!print_tap` case.
With `print_tap`, the flushing happens at the end of the function, and
it is ok and desirable to buffer the entire output and flush it out just
once.

Fixes #15414