bridge: force package configs to be immutable
Make sure that the package config JsonObject (ie: the manifest) is
immutable at the point where it's passed into the router.  It's not
likely that anyone will try to change the values after the fact, but if
they do, it would have some deeply weird effects, so let's forbid it
explicitly.

On the caller side, we need to seal the manifest after reading it and
possibly applying overrides, and also add seals to a few places in the
tests.