bridge: clear superuser_rule before destroying it
If cockpit_packages_reload() is called then the bridges loaded into the
router will be refreshed.  During that process, the existing rules are
destroyed and new rules are created.  The self->superuser_rule pointer
is permitted to continue to point to a destroyed rule, which is a memory
error.

Miraculously, we never noticed this before, because the rule would be
immediately recreated, with a reference to the very same config object
as it had before, allowing lookups into that object to occur
successfully.

A future commit will introduce a variable substituion step which will
cause the config to potentially be regenerated on each reload, resulting
in a different object at a different address, and a crash.