test: Fix Run button click in TestMachinesSettings.testCPUModel
The "Run" action is a primary button on the page, not part of the kebab
menu. Thus performAction() is the wrong tool for this. This was slightly
wrong before commit 7b82e6d849d0, as the code first opened the kebab
unnecessarily and then clicked the button next to it.