build: fix `make uninstall` if no debug is present
We have some custom rules in `make uninstall` to deal with removing the
installed files from dist/, including a separate rule for map files.
We don't always install map files anymore, and due to the changes in
5580cb0a, don't even create the directory in that case.

This doesn't play nicely with our uninstall rule which assumes that the
directory will definitely exist.  Add a check.

Add a check also to the equivalent rule for the normal installation
files: this makes `make uninstall` nicely idempotent.