build: use manifest.json as the webpack stamp
Now that manifest.json is written to the dist/ directory by webpack, we
can use it as a stamp instead of having a separate stamp file written by
make.

Conditionalise the building of the manifests as part of `make` by using
the usual `SKIP_NODE` tricks attached to `noinst_SCRIPTS`.  Remove
`WEBPACK_INSTALL` there: it's redundant.

Finally, make sure the manifests are in `EXTRA_DIST` so that `make dist`
depends on them (and triggers the webpack build) even if `make` isn't
run separately first.  That's needed for working `image-prepare` on the
bots.

This change gives us a nice new property: everything that gets written
to dist/ during `make` comes from webpack or webpack-make now.