webpack-make: keep manifest.json out of _OUTPUTS
825a38e2 contained a modification to webpack-make which prevented
manifest.json from being deleted as it was created.  This change had a
side-effect of causing manifest.json to appear in the _OUTPUTS varibles
in Makefile.deps.

We don't need that to happen since manifests are already handled
separately.  Furthermore, it will complicate our upcoming work of
turning the manifest itself into the stamp, due to a circular
dependency.