storage: Improve size slider text input
We no longer convert the text input into a number immediately on every
change.  That did interfere severely with the user input.  For
example, it was impossible to type a decimal point "." at the end of
the text because it would immediately be removed again when converting
the text to a number and back again.

Instead, textual input is converted to a final number in the validate
function.

Fixes #15409