machines: reapproach forcing the resource limits when creating new VMs
* Recommended resource limits as fetched from osinfo-db are not used any more.
They usually are 20GB storage and 4GB memory, which is an exageration
for for cloud installations.

* Minimum resource limits are not used neither. They were usually 10GB
storage and 2GB memory, which is is an exageration as well. Just show
these values as a warning helper text if the user inserts something
smaller, but do not force these.

* Force node's max memory as an upper limit for the VM's allowed memory.

* Show as a warning the host available storage, if the user tries to
insert a bigger value. Do not stop user's creating disks with size
bigger than the available space on the target storage.

* Lastly drop sliders for memory and storage size since we don't enforce
low limits at this point.