test: Drop ovirt tests
cockpit-machines-ovirt has been an evolutionary dead end. The code in
this branch is frozen, and in the unlikely chance that we'll ever have
to touch it again, it will have to be tested manually anyway -- our
ovirt image is many years old and has been unbuildable for almost as
long (and thus has an ovirt version which is long out of support).
Relying on that will not increase any confidence in code adjustments.