ws: Document utmp string termination in btmp_log()
It is explicitly documented in utmp(5) that these fields are not null
terminated if they use the full length.

On RHEL 8, Coverity complains about that with a BUFFER_SIZE error, but
that's just how it works; so add corresponding annotations and a
reference to the glibc bug.

Part of https://bugzilla.redhat.com/show_bug.cgi?id=1938693