test-static-code: test more python files
The recent changes to FMF (which included adjusting the exclude pattern
for the python static code checks) triggered an interesting (yet
very understandable) limitation in git: if you create the pre-push hook
in the recommended way:

  $ ln -s ../../tools/test-static-code .git/hooks/pre-push

and then use worktrees, then git will always run the check out of the
primary worktree, instead of the one you're working in.

That led to a bunch of newly-introduced files not being properly ignored
during the pycodestyle checks and a bunch of errors being thrown.

Investigations of this problem revealed that in addition to sometimes
scaning files that we shouldn't be, we were *not* scanning a majority of
our python files.

Change our detection logic to try to catch all of our python files, and
avoid needing to maintain an increasingly unwieldy exclude list.