test-static-code: some W504 adjustments and notes
Fix up a single case of using + for string concatenation when
juxtaposition will do just fine.

This was just about the only useful case kicked up by trying to remove
our W504 exception from pycodestyle.  The others are not improvements.

There are two interesting things to note here:

 - W504 is ignored (along with many others) in the default --ignore
   list, so we could actually drop our --ignore parameter and get more
   permissive checking than we have now

 - W504 complains about "line break after binary operator" whereas W503
   (also disabled by default) complains about "line break before binary
   operator".  It's pretty clear that we can't have both of these
   enabled, so let's just leave this one disabled.

As such, drop the blanket TODO item in test-static-code and replace it
with an explanation about why we intend to keep --ignore=W504 and a more
specific TODO about E204.