build: Clean up --enable-polkit configure option
Don't call `pkg-config polkit` when building with --disable-polkit, as
it's confusing to have two messages

    checking for POLKIT... yes
    checking whether to build with polkit... no

Drop pkg-config auto-detection, and default to enabling polkit. Distros
or users who want to build without it now have to specify
--disable-polkit explicitly.

This allows distributions to reliably build with/without polkit support
without having to rely on auto-detection.

Fixes #14052