lib: replace onHover custom function in plots with an inline tooltip
Previously when hovering on the resources, we highlighted them in the
relevant part of the screen. This has two main issues:

* The lists can get long so nothing guarantees that the resources of interest
are in the visible part of the screen

* The highlighted section was just changing background color which is
not helping people with limited vision.

This commit replaces this highlighted sections on hover with an inline
tootip that appears when the plot is focused.

Note: title is also read by screen reader software therefore is an accessibility
enhancement from our previous state.

Fixes #4006