tls: split out certificate handling code
Create a new file — certificate.c — and split the code for loading the
certificate file into it.

The main benefit of this change is that we will be able to use the same
code to help check for the existance of the certificate from our new
remotectl replacement.

Use the chance to add a wrapper structure around the certifcate data
with support for refcounting.  This will come in handy later.