tls: set permissions of 10-ipa.cert on install
This was never being explicitly done, under the assumption that the
certificate file, as created by IPA, would be world-readable by default.
It's not, so let's do it ourselves.

The reason this is working is because remotectl is coming along after
the fact and setting the permissions, incorrectly (under the assumption
that there was also a private key in there), but in a way that would
allow cockpit-tls to load it, at least.

Remove that code from remotectl now as well: we will soon stop running
remotectl in the normal case.