tls: tidy up cockpit-certificate-helper
When we were originally creating this file, it was a very direct copy of
the shellscript that was being assembled in-place by browser-side
javascript code.  It was never meant to be a proper script.

Perform some general cleaning up, as well as splitting some of the code
into functions.  Introduce the use of a permissions-restricted working
directory (which will be cleaned up on exit).  Step up our automake game
a little, and support installations not in /etc.

Finally: remove the +x bit.  It's meaningless on .in files; the script
will be set +x on installation.