Revert "testlib: Always output unexpected journal and browser messages"
This leads to many test failures not being retried, as with that commit
they started to fail twice: First for the real reason, then because of
whichever journal messages caused the failure.

In many cases that happens because tests tend to call
allow_journal_message() at the end, which they never get to if they fail
in the middle. But even if we'd change that, it's still conceptually
wrong -- tests should have one back trace and failure reason only, not
two.

This was introduced for
https://github.com/cockpit-project/bots/issues/1453 , but we don't
actually require that -- the naughty can match on the error that the
browser console writes.

This reverts commit b75cfe652873cdf578e19094ff7289aaca8bea14.