users: Fix lastlog parsing - April the 1st edition
This parsing has been changed in 03314e6b1d on 23.3.2021.
Therefore we only tested it on dates with double digit dates which
failed to notice that when date is single digit it does not work correctly.

We split the string on space, which produces:
`Tue Mar 23 14:49:04 +0000 2021` => ["Thu", "Mar", "23", "14:49:04", "+0000", "2021"]
`Thu Apr  1 08:58:51 +0000 2021` => ["Thu", "Apr", "", "1", "08:58:51", "+0000", "2021"]

When we then take last 5 fields in the latter case we don't get month
which breaks subsequent parsing.