various: pycodestyle changes
LGTM results from our entire code base found a lot of issues that would
have been uncovered by tools/test-static-code if we'd been running it on
all Python files, and not just those in tools/ and test/.

In preparation for expanding the coverage, start fixing some issues.

In addition to purely whitespace fixes, we have a couple of cases of:

 - drop trailing ;
 - one import per line
 - if not x in y → if x not in y

This patch is a mix of automated fixes from autopep8 and manual changes.