test-static-code: more pycodestyle adjustments
Our recent changes to test-static-code were causing node_modules/ and
bots/ to be scanned.  Steal some code from bots/ (thanks Matej) to make
sure we only scan files that are in git.  That should also speed things
up a little bit.

This causes us to find a couple of new files, so make the usual tweaks
to fix those ones up.

If we're not in a git checkout, #skip the test.