base1: Add option to disable file reading for cockpit.file().watch()
This is useful when watching large files or non-readable devices, i.e.
when the caller is only interested in the actual change notification.

Add a `read` option to watch(), which can be set to `false` to disable
reading. If not present, keep the original default to not break
backwards compatibility.

Fixes #14188