.gitmodules: add node_modules as a real submodule
Up to this point, node_modules has been sort of like "half a submodule"
in the sense that it has a gitlink, but no corresponding .gitmodules
entry.

We get away with that because we have our own logic for finding and
checking out the content.

Add a .gitmodules file to enable people to use standard git commands to
do this for themselves if they'd like to.