tools: Fix build-debian-copyright to work against build tree
We had to run it against the `make dist` tree in the past because that
had a reduced set of node_modules/ (without devDependencies), and
running against the full module set failed.

That changed with commit 8605d56e131, where each webpack explicitly
enumerates included-modules. With that we can run the script against the
build tree with identical results. This makes it easier to change/test
the script, and also opens the possibility of having a pre-generated
copyright file.

This allows us to build tools/debian/copyright in a more standard
fashion.

Replace the iteration over all dist/ with iterating over all
`included-modules` files. This more precisely gives us the webpacks, and
weeds out unrelated stuff like "README.md" or "guide". (Also, "ssh" and
"pcp" do not appear in dist/ any more.)