test: Unpack pre-built debian/copyright from our dist cache
This can't be built without a full node_modules/, and was the only thing
that breaks `make dist` with a pre-built dist.

Unfortunately tar's `--touch` order seems to be unpredictable, and not
the order that is given on the command line. Ensure that the unpacked
debian/copyright is newer than the dist manifests.json.