test: Use grubby to set crashkernel= option
This now works on all our Fedora/RHEL/CentOS images, and grubby is the
official tool to do that. Directly editing /boot/grub2/grubenv is not
supported, as it says prominently on the top:
"WARNING: Do not edit this file by tools other than grub-editenv!!!"

Both editing grubenv and specifying boot options in /etc/default/grub
stopped working on the current RHEL 9.0 image.