test: Quick-fix dist tarball unpacking
The tar timestamp workaround from commit 0e1a264f42c2 with only touching
webpacks, but not node_modules/ and package-lock.json was completely
flawed: This would only work on an already existing cockpit.git
checkout, but not on any clone that happened after building the dist
cache.

So we have to touch all the files, which also simplifies the code quite
a bit.

To work around the `tar --touch` bug, add a sleep to ensure that the
unpacked files are at least one second apart. This is a giant hack of
course, but let's unbreak our CI and investigate this in an orderly
fashion.