services: handle better transient units
The garbage collection from systemd will automatically unload the unit
configuration and state when a unit is not needed anymore.

However, by default failed units are not unloaded until the user invoked
systemctl reset-failed on them to reset the failed state.

When setting CollectMode=failed-or-inactive this behaviour is altered:
in this case the unit is unloaded even if the unit is in a failed state,
and thus an explicitly resetting of the failed state is not necessary.

This commit makes sure that transient services when the get
inactive/failed get handled properly.

Closes #15801