tools: Fix expected mode of SELinux policy directory
Create the compiled directory in the build root, so that `%ghost` can
see the expected mode. This fixes the `rpm --verify` error, as so far
the "expected" mode was 0 due to the directory not being present in the
build root.

    .M.......  g /var/lib/selinux/targeted/active/modules/200/cockpit

Drop the `--nomode` hack in `TestConnection.testWsPackage`, so that from
now on the package must be fully valid.