metrics: Add switch for exporting metrics to network
This enables Prometheus/Grafana and similar metrics
collection/visualization tools to read metrics from the machine.

That network API is provided by pmproxy.service. It additionally needs
redis.service to run for the metrics parts of the pmproxy API, so always
start/enable this as well together with pmproxy.

Add a new testPmProxySettings() case to test all the fine-grained
state combinations of the involved services. Change `TestGrafanaClient`
to use the UI instead of CLI to set up pmproxy.