workflows: Run containers with docker
The most recent hack to run podman as root (commit 5470ce34c040) is
still not enough, podman still falls over on the i386/debian image now.

Let's give in, that third-party OSBS podman backport is just not
reliable/supported enough; run everything through docker, which is the
supported platform on Ubuntu 20.04. We can revisit this with Ubuntu
22.04 LTS, as podman is now in Ubuntu proper.

Set the `$docker` variable, which the unit-tests scripts use to select
between podman and docker.