storage: use SHA256 for Tang fingerprints
For some time now, `tang-show-keys` has been showing the user key
fingerprints based on SHA256, while we've been computing them in Cockpit
using SHA1.  Our advice for the user to compare these two values isn't
very useful in that case.

This was exposed by a change to jose to default to using SHA256 instead
of SHA1 if no hash algorithm is specified.  We invoke jose directly from
the integration tests.

Switch cockpit over to computing hashes using SHA256, and adjust our
invocation of jose to explicitly specify the algorithm as SHA256 as
well, so that we remain compatible also with older versions.