reposchutz: use pull_request.base.sha
github.sha seems to refer to the current state of the main branch of the
repository, not the base branch that the PR was proposed against.  That
means that if someone else lands a github workflow change in the
meantime, unrelated PRs will start going red.

Change that to be based on github.event.pull_request.base.sha.