build: drop building -cache tarball
Now that we host the node_modules in a separate git repository and
include it as a submodule, we can always get exactly the set of files
that a previous release was built against.

We'll also start tagging the node-cache repository at the same time as
the Cockpit release, with the same version, to make that even easier for
people working with tarballs.

Now that we're down to building a single tarball again, we can fall back
on the normal `make dist` machinery and remove our overrides.  This
means that our home-baked `XZ_COMPRESS_FLAGS` variable is now replaced
with the standard `XZ_OPT`.  Adjust test/make_dist.py accordingly, also
taking the chance to drop the now-superfluous `NO_DIST_CACHE=1`.

The produced tarball contains one less file:

  test/common/axe.js

which is a result of our custom `tar` invocation including symlinks and
the standard automake one not doing it.  We don't actually dist that
node_module anyway, though, so it's harmless to lose this.