test: Fix PackageCase NetworkManager setup
The libvirtd code path was incomplete: It's not enough to start
libvirtd, it would also need to ensure that the `default` network is
actually up.

However, we don't actually need this any more -- just adding a dummy
connection works on all Debian/Ubuntus, so let's simplify this. On the
Fedora family this is not necessary at all, as NM already manages the
main interface there.