test: More robust pixel-tests pushing
Use --force to really blow away the index.  Git otherwise refuses to
remove things from it that are neither in the HEAD commit nor in the
working tree.

Fetch the "empty" branch before using it.

Always push to the remote, even when there are no changes. The
previous push might have failed and this is an easy way to try again.