cockpit-po-plugin: replace function(n) with a lambda
In the output .po, we provide a function for deciding plural forms.
This is basically the perfect case for a lambda expression, so use that
instead.

This approach completely ignores the nplurals= part of the plurals form,
since it's not interesting to us.  Indeed, including this under the old
approach was basically dead code, since the plurals expression isn't
allowed to reference it, anyway.