cockpit-po-plugin: stop using po2json, rework
Stop using po2json, in favour of using gettext-parser diretly.  po2json
is basically just an output transformation on gettext-parser anyway,
which is a little bit silly for us, since we then apply yet more output
transformations.

Importantly, using gettext-parser directly allows us to access the
references on the translation object returned by gettext-parser (which
will be useful in a following commit).

Also stop pretending that our output is JSON: we were using some ugly
hacks to get our plural forms function declaration into the output
"document".  Move instead to a chunk-based approach.

This refactor makes some additional reasonable assumptions about the .po
files that we will be processing, for example, that they will always
have a header, and will always declare a plurals form.  It also makes
assumptions that the consumer of the output is cockpit.js.

For our current set of .po files, this commit preserves the output
exactly as it was.