test: Fix race condition in TestGrafanaClient initialization
On Fedora/RHEL systems, starting pmproxy and redis in parallel does not
work: pmproxy needs to start *after* redis is already running, so that
its metrics API actually works. So use the same code structure on all
operating systems.

PR #15812 will fix that properly and get rid of that manual
initialization, but that still takes a while to land.