systemd: fix _global-variables.scss import path
46b403b03c756a79215ddde7a12fc7457be6beef introduced an @import of
"./_global-variables.scss" in reporting.scss.

I'm not sure why this got imported with "./" but since that's not in the
same directory, it seems wrong.  Furthermore, although it works under
sassc, it breaks with Dart Sass.

Other users seem to import this as "_global-variables.scss" and rely on
path resolution to find it in pkg/lib, so let's do the same here.