test: check out the bots on an as-needed basis
Most python things that need to use the bots first import parent.py to
set up the path to enable importing things from the bots directory.

Use this as a nice place for a hook to check out the bots, if they're
not already there.

With this in place, we no longer need the blanket call in autogen.