testlib: force use of prepared images
For images not explicitly specified by a provision, and in case
--machine wasn't given, assume that we want to use an image created by
test/image-prepare.

Throw an early exception with a useful message, for ourselves, if it
doesn't exist.

Note that passing a pathname as the image name is explicitly supported,
and the basename (without extension) will be used for setting the
Machine.image variable.