tls: add a "client-certificate" field to metadata
In case the client presented a certificate, store the basename of the
certificate file as the "client-certificate" field in the metadata blob.

We can use this from cockpit-ws to determine whether or not we should
perform client certificate authentication, removing the need to check
for the certificate ourselves.  We pass the certificate filename along
as part of the "tls-cert" authorization type, but cockpit-session will
ignore it for the time being: that will be addressed in the following
commits.