metrics: Replace moment with Intl.DateTimeFormat
moment is deprecated, and the metrics page only uses absolute
dates/times.

Clean up the generation of the "last 14 days" selector to avoid the two
useless back-and-forth conversions, and handle the "Today" case in a
more symmetric fashion.

Avoid any locale based formatting altogether with the journal timestamp
computations, as they expect a fixed format anyway. (We could also do
this with `@unixtimestamp`, but this does not generate human friendly
values in the Logs page search box.)

See #15979