docs: rework (and unbreak) dist hook
We have a hook to display an error message if you try to do `make dist`
with `--disable-docs`.  The way the logic is written causes errors about
"warning: dist-hook was already defined in condition TRUE" when running
autoreconf.  That's because automake doesn't know about GNU-style
double-colon rules and that multiple definitions of them is OK (and even
expected).

Even worse, though: the check is broken, because in the case that docs
are disabled, nothing actually causes the warning dist-doc-hook to be
called, and additionally, it wouldn't have a failing status.

Adjust this to eliminate the warnings and fix the broken check.  The
resulting code is not exactly nicer to read, but it works a whole lot
better and avoids the warnings.