guide: don't dist fonts if docs disabled
We include our fonts in EXTRA_DIST unconditionally, but only define the
rule for copying them if ENABLE_DOC.

This results in an error about now knowing how to make those files in
--disable-doc builds, if you run `make dist`.  This happens before the
dist-hook which would output the proper message about why, and leaves
the user guessing about what's broken.

Conditionalise the inclusion in EXTRA_DIST: the build will still break
on `make dist` with `--disable-doc` but it will do it in a much more
deliberate and helpful way now.