make-bots: support symlinking to another bots dir
We probably all have a checkout of `bots/` somewhere already.  Martin
mentioned that he likes to setup a symlink to it when creating a new
worktree.  Let's automate that process in an elegant way.

Modify `tools/make-bots` to check for the presence of a symlink at
`~/.config/cockpit-dev/bots`.  If one is found, use the directory that
it points to as a target for a symlink from `bots`, as an alternative
to checking out our own copy.

A small matter of taste: we print the (relative) directory name of the
specified bots directory in the log message, but we actually link to the
`~/.config/cockpit-dev/bots` symlink itself.  That way you can update the
symlink as-needed (if working on another branch, for example) and all
existing cockpit repository worktrees will automatically switch over.