build: fix `make install` with `--disable-doc`
Commit 5b0a778b0199762a712424c518ec95678882082a rearranged the
conditionalisation of our docs targets and accidentally introduced an
unconditional dependency on dist/guide/html/index.html to the
install-data-local: rule, even if docs were disabled.

We can't conditionalise the rule itself, for fear of bringing back
automake's complaints about duplicated definitions, so introduce a
variable.  This isn't pretty, but it works.

Fixes #16064