lib: shutdown-restart dialog: use date-fns instead of the deprecated moment for parsing localized strings
So far we were using Intl.DateTimeFormat JS builtin date localization
functions, but there is no way to parse an already localized string to a
date Object with that.

We need the above functionality for providing PF4's DatePicker with a
parseDate function and localized placeholders. Start using date-fns for
this reason.