session: Move PAM and session handling from session-utils into main program
The original code split-out to session-utils in commit 60bdb8e15aaf47
had the intention of sharing as much code as possible with separate
cockpit-<type>-session programs (in particular, cockpit-cert-session).
But that approach was a dead end, and never landed.

Let's start to organize this a bit more logically: All PAM and session
process handling and related helpers are now back in the main session.c,
and session-utils.h keeps the more generic helpers and JSON protocol
bits.