ws: Add alternative error message filtering in cockpit-session
In recent krb5 versions, with an usage mode other than `GSS_C_ACCEPT`,
the error message looks like this now:

    No credentials were supplied, or the credentials were unavailable or inaccessible
    (No Kerberos credentials available (default cache: KCM:))

This is `GSS_S_NO_CRED`, so accept that as alternative expected
error message. The "Keytab is nonexistent or empty" string that we get
on older krb5 releases is the generic `GSS_S_FAILURE` with an
indecipherable numerical minor code, so keep string comparison for
these.