containers: Properly run test-static-code during unit tests
Since commit da2ae76990c6cb our unit test runs did not run pycodestyle
any more, as that started to skip the pycodestyle check in a non-git
tree (such as the unpacked tree that `distcheck` runs in). As
test-static-code does not run in `check-memory` either, none of the four
scenarios ended up running it.

So explicitly call it in the distcheck scenario, and relatively early so
that failures (which happen quite often) are easier to spot in the log.