session: Drop cockpit-session- prefix from client-certificate.[hc]
This is not an API for anything other than session.c, and now is already
namespaced by src/session/. So rename the file to be shorter. This is
now nicely parallel to src/tls/client-certificate.[hc].

Also fix the obsolete (and now non-matching) docstring of
read_cert_file(), and add a docstring to the _map_user() function.