bridge: Fix PCP archives on 32 bit architectures
`label.ll_start.tv_sec` is a 32 bit type on armhf and other platforms.
Multiplying with 1000 causes an integer overflow, which completely
messes up the time stamp. Fix this by casting to a 64 bit type first.

This worked in the /metrics/archive-timestamp unit test because that
starts with time stamps at the epoch. Add a new
/metrics/archive-timestamp-now unit test which works with real-time
"now" timestamps, to ensure that works as well.

Fixes #16083