test: allow for existing /var/cache/krb5rcache
It's true that we clean up /var/cache on creating images, but sometimes
we install packages onto the system after image creation (for the copr
variant, for example) and that can recreate the directory.

Use `mkdir -p` to avoid the error.

Also: use `mkdir -Z` instead of running `restorecon` separately.