build: re-prevent node from messing with stdio
node does some ioctls on startup which seem to be resulting in our
terminal landing in a non-blocking state in some situations.  This
problem has plagued us for a long time, but was brought back again
(somewhat) recently with 94a2269188, which removed use of the timeout
command (which, as a side effect, gave us a protective pty wrapping).

Add a small python script which should suffice to get us this protection
back again.