webpack-make: convert -d to a position argument
webpack-make has an awkward mandatory `-d` argument which you need to
specify like so:

  tools/webpack-make -d dist/static/Makefile.deps

Internally, it disects this pathname to get to the part that it really
cares about: the "prefix" (`static`) which it uses for many things.

Change things around to just specify the prefix directly as a mandatory
positional commandline argument:

  tools/webpack-make static

which serves to simplify and robustify the logic, and increase
user-friendliness.