test: use .restore_dir() instead of env redirects
TestConnection.testCommandline() is from a time before we had
.restore_dir() and used a temporary directory and environment variable
overrides to acheive the same effect.

Use .restore_dir(), which get us a lot closer to actually testing the actual
production config.

We need to reorder the test so that the checks that run on the
"unmodified" system run before we make the modifications.