tools: Provide eslint mode for webpack-watch
Since commit 66c76af8f3af8 it is not possible any more to run
webpack-watch with eslint, as there is no option to enable it, and the
program overwrites `ESLINT=1` from the environment.

Drop the weird inconsistency between webpack-{make,watch}), and default
to running eslint in both modes. Then `--no-eslint` can be used to
disable it.