lib/table: fix miscalculated colSpan from PF4
PF4 does not seem to always calculate the colSpan properly for the
expanded content. An example is the selinux alerts expanded rows, where
the colSpan should be 5 but it's 4.

In pkg/lib table implementation we currently use for the expandable table variant https://www.patternfly.org/v4/components/table/react-legacy

Since this is a legacy component, we can't open issue about the
miscalculated colSpan, but we need to port to the new Table
implementation from PF4 at some point.

Till then let's have a local fix for the colspan of the expanded rows.