test: Verify that cross-domain frame embedding does not work
This reproduces the scenario in #16122. Try to load a cockpit frame from
a different domain (localhost:12346, from a local python web server),
and ensure that this does not work beyond the login page. We really want
it to refuse to load the frame at all, but as we currently don't set
`X-Frame-Options:`, this currently does work, in both Chromium and
Firefox.

But make sure that the default browser/cookie policy prevents loading
the frame and leaking credentials. This test only works in Firefox, as
Chromium's CDP driver does not even expose cross-origin frames (even
though the browser itself does show them).