test: Add pixel tests for Overview page
Compare the complete layout of the Overview page in a steady and
relatively well-defined state. This ensures that card layout, fonts etc.
are as expected.

Add a separate check for the motd box, as we have some non-trivial code
and CSS for sizing/aligning with the cards, and ensuring a fixed-size
font formatting.