po-refresh.yml: split out pot upload to weblate
The current po-refresh job does two things:

  - generates an updated cockpit.pot and commits it to the weblate repo

  - copies .po files from the weblate repo and opens a PR with them

These two things are actually completely independent from each other, so
let's separate them out to allow for more fine-grained use of
credentials.

Introduce a update-pot job to the existing po-refresh workflow which
takes care of updating cockpit.pot and committing the results to the
weblate repository.  This job requires very few permissions: it only
needs to write to the weblate repository and it does that by way of a
deploy key.

Meanwhile, remove the commands from the `update-pot` rule in
po/Makefile.am to prevent this job from being done twice.  We need to
keep the target itself because bots/po-refresh expects it to exist.