systemd: Drop computer-ou input field from realm join dialog
This has never worked -- literally the first-ever commit (cbeb24173b206)
got this wrong (the code queried `.realms-join-computer-ou`, while the
HTML defines `#realms-op-computer-ou`). This never had an integration
test, testing it does not work with our Samba server (as it does not
define any "Computers" OU and `adcli show-computer` does not show the OU
either). Also, that feature is incomplete (there are many more fields
that one can set [1]), and nobody complained in all the years.

So let's just remove this, and only bring it back in the future when
there is some actual demand. This also gets rid of the
`.realm-active-directory-only` class special case.

[1] https://freedesktop.org/software/realmd/docs/gdbus-org.freedesktop.realmd.KerberosMembership.html