test: Fix race condition in check-realms testClientCertAuthentication
Wait for the auto-detection to be finished before typing a manual
address into the input line. This was forgotten in commit 017dcfd1e2a;
that changed `Browser.set_val()` (which does not work with PF4 React
components) to `.set_input_text()`, which does not clear the input
first. So it could happen that autodetection happened, and the
set_input_text() just appended after that.