webpack: Use relative resolve path for npm 7 compatibility
npm 7 changed how it resolves dependencies, and cockpit fails to
build with lots of unresolved peer dependencies of PatternFly.

With an absolute path, `resolve.modules` will only look in that
directory; the default is a relative path "node_modules" that just
works [1]. We need to keep the `$SRCDIR` support, but convert the path
to a relative one to keep the old recursive search behaviour.

[1] https://webpack.js.org/configuration/resolve/#resolvemodules