metrics: Initialize top most timestamp to something usable
When the initial load of the last 12 hours does not return any data
then after one minute when we try to update the last minute we load all
the data there is.
The updating logic calls `load_data` with `this.most_recent`. This variable
has initial value `0` and is updated when some data are parsed.
Calling `load_data(0)` translates to loading all the data it can find.

Fixes #16462