services: show user owned systemd units in the services page
Closes #13217

Some notes:

* With the current implementation toggling the owner select will remount
the main service's page component. This was a deliberate implementation decision although it
might seems suboptimal user experience. The reason is that this
page is already quite slow: by keeping data for both connections we
would introduce an extra killer for the page's load times. So, let's keep it simple
for now, by allowing the component to keep only one of the two
connection states at once.
This makes the code changes also rather trivial. We can consider a more
elegant solution in the future.

* The root user does not have user services. Hide the new toggle buttons
when logged in as root.

* Hide for now the 'Create timer' button although it's totally possible to
create a user systemd timer unit. Let's enable the feature in a
followup.

* Hide service logs for user services. These are available but the
LogsPanel component will need some extension for supporting it.

* tests: some parameterization was required and some parts of the tests
were skipped as they were relying on some specific system services
(NetworkManager, tty etc). However, since operations like filtering are not
affected anyhow by the 'owner' selection, the fact the we skip some
parts on the user test is not problematic.