test: Generalize set_input_text() blurring
Generalize the blurring of the input element after typing keys in
set_input_text(). We keep running into the same problems with pixel
tests, and if tests need a particular item to be focussed, they should
be explicit in that.

The time/date validation checks in check-system-shutdown-restart *are*
dependent on keeping the focus, so provide a flag to disable the blurring.

Update all affected pixel tests which previously showed focused input
lines.