bridge: Block signals initially for the default CockpitDBusCache
The CockpitDBusCache for the default name of a dbus channel is already
careful not to process signals from other sources than the default
name, but it starts doing that only as soon as the unique name is
known (or known to not exist).  Before that, it will accept all
signals, which might still lead to journal messages like

    org.storage.stratis3: couldn't get managed objects at /org/freedesktop/UDisks2: GDBus.Error:org.freedesktop.DBus.Error.ServiceUnknown: The name org.storage.stratis3 was not provided by any .service files

Thus, we block signals initially as well, until the unique name is
known.