bridge: modernise ssh-agent spawning code
Port the code for spawning ssh-agent to GSubprocess, merging it with the
(now-very-similar) code for spawning dbus-daemon.

We spawn ssh-agent with -D so that we have better control over its
lifetime.  In particular, the `prctl()` that we were previously doing
had no effect, because without `-D` ssh-agent `fork()`s immediately.
Closing all file descriptors also had no effect because this was already
the default behaviour of `g_spawn_sync()` and is also the default
behaviour of `GSubprocess`.