packagekit: Show "Web Console restart" for Arch and derivates
On Arch Linux and derivates cockpit-ws is included in the cockpit
package which makes users miss out on the "Web Console restart" warning
message. As these distriubtions all use the "alpm" backend, the
component now on load fetches the configured backend (can only be one)
to determine which package to check. Using packagekit was chosen over
parsing /etc/os-release or checking if pacman is installed on the
system as we already interact with packagekit over Dbus.