storage: Use PF4 pf-c-table__action for content row actions
This finally produces a useful mobile layout.

The "compact" table variant is very cramped when used with actions, so
let's use the default non-compact one.

Also, the empty <Td/> that takes the place of the expander in
non-expandable rows needs to be styled like the real expander toggle.
This ensures that the mobile layout works as expected with
actions.  (Otherwise, this empty <Td/> would show up as a empty row in
the mobile grid.)