tls: add internal https redirect handling
Add an internal helper to send the 301 Moved Permanently replies to
clients who try to speak http when we're expecting them to speak https.

We still allow localhost to connect, but now we base that strictly on IP
address, and no longer look at the Host: header (which -tls doesn't have
access to).  The AllowUnencrypted cockpit.conf option still overrides
the check.