metrics: Show top CPU core usage
Often some application maxes out and gets bottle-necked on a single CPU
core. On systems with a lot of CPUs this is not visible in the average
CPU progress though. It is visible in the tooltip, but first this is not
very discoverable, and second there is no indication for the user to
actually check this, as there might be dozens of idle cores next to the
single 100% one.

On multi-core systems, show a second CPU progress bar for the currently
busiest CPU core, and show the average and maximum usage in the label.
This provides a good balance between pointing out the problem and
avoiding cluttered UI with showing all of them all the time (which is
not even all that interesting).

Avoid showing the usual Progress "danger" icon next to the progress bar
for CPU usages > 90%. It destroys the layout and is not actually that
helpful.