test: Robustify TestCurrentMetrics.testMemory
In the "have swap" code path of this test it often happens that the
memory usage after killing both memhogs does not go back quite to the
same number as at the start of the test. In our noisy test VM this could
have any number of reasons. So make the check less strict by measuring
against the usage before starting mem-hog2 (which is higher, as memhog
is already running). The point of the check is that current memory usage
goes down after stopping the services, but we can't pinpoint an exact
number.