test: Work around Debian testing pam-sss log permission bug
Debian testing's sssd has a subtle bug that messes up the permissions of
/var/log/sssd/p11_child.log (https://bugs.debian.org/1001377) under some
conditions which are triggered by TestLogin.testClientCertAuthentication.

Work around this by removing the sssd logs for every individual check.
This also makes debugging failures easier, as the logs then only
contain the messages relevant to the current scenario. But still
restrict it to Debian testing, so that we know if/when that bug happens
in other distros in the future.