test: Add pixel tests for top navigation bar
Add reference images for privileged and unprivileged mode. These are
currently identical in mobile mode, as we don't show a privilege
indicator there. But still keep them as separate reference images for
simplicity, and just in case we will ever add a mobile indicator.

Note: The superuser dialog is already covered.