unit-tests: split entrypoint into a separate script
Instead of generating /entrypoint from setup.sh (where we do a single
variable substitution for the personality), copy the file in as a
separate step.

In addition to being cleaner, this helps with caching: hacking on the
script is a lot easier now because the entire image doesn't need to be
rebuilt as a result.