webserver: rename WebRequest to CockpitWebRequest
Rename our internal helper struct WebRequest to CockpitWebRequest.  All
of the methods that operate on this struct are a little bit confused
about if they're methods on the server or methods on the request, so
rename them and make use of 'self' as the first parameter name to show
that they're meant to be taken as methods on CockpitWebRequest.