test-webservice: substantial restructure
It turns out that our websocket code contains support for parsing a
request and headers, all on its own.  This code is only used by tests,
many of which make connections between the webservice and a websocket
client, directly.

Restructure all the tests in test-webservice to make use of the
CockpitWebServer that is already being created (but used for nothing).
We now make connections via the webserver and wire them through to
cockpit_web_service_socket() from a handle-stream handler.

This means we're testing much closer to what would be happening in the
"real world": in particular, this gives us a much nicer way of running
the origin tests against various combinations of the config file,
X-Forwarded-Proto headers, and "for TLS proxy" mode.

We perform a similar adjustment for a single testcase in test-handlers.
For that, we need to make a small tweak in cockpithandlers.c: a guard
there was assuming that if `headers` are non-%NULL then `ws` must be
too.  That's no longer true now that we feed headers in via this
testcase.

Unfortunately, we can't drop the parse-my-own-headers code from the
websocket server, because we use it from quite a lot of worthwhile cases
in test-websocket.