common: start using closefrom(), keeping our hack
47bc8b93b7604b0831090e17e043286c4a81f1cb introduced our own version of
close_range(), but this turned out to be the wrong API.  closefrom() is
meanwhile available on modern versions of glibc.  It features a
simplified API, with no chance of failure.

Rename cockpitcloserange.c to cockpitclosefrom.c and move the definition
into cockpithacks.h.  Introduce a configure check for closefrom() and
port all users to the closefrom() API.  In case that's not available, we
redirect those calls to our own cockpit_closefrom() (which is now only
conditionally-built).

Since neither the glibc API nor our own code have any chance of failure,
we can drop all the error checking we were doing before.

Incidentally, this should fix warnings we were occasionally seeing from
Valgrind that we were finding and attempting to close its log fd.
Modern Valgrind wraps closefrom() and prevents this from happening.