webpack-jumpstart: don't touch unrelated files
When doing a partial jumpstart, we touch any files that have changed vs.
the state of the tree when the jumpstart was built.  This is a little
bit silly for .c files and the like, so make sure the name of the file
in question is actually a dependency of the webpack build before
touching it.

This check could theoretically have false positives (like if a filename
appears in the Makefile in some other context), but it should never have
false negatives.