shell: Fix changing of an uncontrolled input to controlled
Warning: A component is changing an uncontrolled input of type <input> to be controlled.
Input elements should not switch from uncontrolled to controlled (or vice versa).
Decide between using a controlled or uncontrolled input element for the lifetime of the component.

If `state.user` or `state.address` is `undefined` it is considered as uncontrolled.