test: Fix TestConnection.testCockpitDesktop on arch
Arch does not have the polkit shim for reading the *.pkla format, it can
only read JS rules. OTOH Debian/Ubuntu don't have support for the JS
rules, so we need to write both kinds of files.

This only worked sometimes on arch because of the race condition fixed
in the previous commit.