tools/test-static-code: major overhaul
The current situation in this file is a bit unfortunate.  The number of
test cases (for the plan) has to be tracked manually, and each test is
responsible for outputting its own ok/not-ok lines manually.

The output is also not really TAP compliant: we often have diagnostic
lines which are printed without the required leading '#'.  We also
freely mix stderr and stdout.

The tests themselves are also often more complicated than necessary.

Do a major overhaul where each test is now defined as a function with a
name starting with `test_`.  We provide a framework which automatically
finds and runs all of the functions, deriving the test path from the
function name.  Each test can signal failure by a non-zero status, or by
producing any output at all.  A unified skip mechanism is also provided.
The output is now TAP compliant, and is written exclusively to stdout.

Finally, we give up on the idea that it's meaningful to run these tests
outside of a git checkout.

Given this new framework, most tests are now reduced to one or two
lines.