test-static-code: clean up python checks
Move the Python checks to be located beside each other, and use a common
function for determining the list of filenames.

The new function now avoids duplicating files which end in '*.py' and
also have a #! line, and also avoids problems with whitespace in
filenames.

Running `python -m pyflakes /dev/null` as a check turns out to be really
slow, so switch to trying that via `python -c 'import pyflakes'`, which
is about 3x faster.  That's not a noticeable change now when we're
running each of these on every python file in the tree, but when we
write our new pre-hook which only runs on the changed files, and runs on
each commit, it becomes more noticeable, particularly for commits which
don't change python code.