packagekit: Show a spinner while checking for kpatch support on RHEL
Testing for kpatch support can take quite long (UX review reported 8
seconds), and it is apparently confusing/not discoverable for users to
wait that long until they see the status of kpatch.

During detection of kpatch support on RHEL (the only OS where we
actually expect this to succeed), show a spinner as status. Also show a
disabled "Edit" button to avoid the layout jumping around once it gets
enabled. In the (unexpected/unlikely) case when kpatch packages are
unavailable on RHEL, show a popover with that information.

Don't show the spinner on other OSes, as we expect it to fail there, and
it is both confusing and also changing the page layout when the "Kernel
live patching" entry entirely disappears from the Settings card.