static-code: add warning about disabled `set -e`
Because of the way we invoke test functions:

  output="$(${test_function} 2>&1)" && test_status=0 || test_status=$?

we end up with `set -e` disabled for the body of the test function.

The entire point of the above construct is to avoid the entire script
aborting in case a single test fails, but it has the unintended side
effect of not just protecting that single line, but disabling `set -e`
for the entire body of the test function.  The fact that it's invoked in
a subshell doesn't help us, either.

This turns out to be very difficult to solve (and I've tried several
exotic ideas).  Since most of our testcases are essentially one-liners
anyway, let's just leave it alone.