testlib.py: make a tweak in login_and_go()
login_and_go() accepts a host kwarg and passes it as the first argument
to machine.start_cockpit().

start_cockpit() is defined in the bots repo as:

  def start_cockpit(self, atomic_wait_for_host=None, tls=False):

which means that our `host` gets slotted in to the
`atomic_wait_for_host` parameter there.  The presence of other kwargs on
that function signature make this look like a bug.

Pass host in explicitly as a kwarg with the correct name.

Meanwhile, the start_cockpit() implementation on
BastionContainerMachine() accepts and ignores position arguments.
Nobody calls .login_and_go() for that machine anyway, so that variable
is already unused, but now that we always pass kwargs, it's even less
used.  Drop it.

Note: After a quick search through our projects, I wasn't able to find
any user of login_and_go() which is providing a host kwarg.  Maybe we
should actually just drop this completely.