cockpit.spec: undo stderr redirection during build
e39cb1fb85a26b53d333e10b8d183937b6c79fc9 introduced

   exec 2>&1

to the %build and %check stages to silence warnings so that we'd get a
clean output with rpmbuild --quiet.  It's true that our checks produce
an awful lot of output on stderr, but everything that is produced on
stderr during the build phase is worth seeing.

In fact, we've been masking some issues that we ought to fix up.

Let stderr be shown during the build phase, but leave the redirection in
place for the checks.