debian/control: pin pages to the same bridge version
At Debian package build time, we run a Python script which scans all of
our manifest.json files and returns a single number representing the
highest version of the bridge that any of the packages depends on, and
then we add this as a dependency for all of the pages packages.

It's a bit arbitrary that we take the highest version and apply it as a
dependency for every single pages package, when we could scan the
individual manifest for the page in question and generate the dependency
more precisely.

On the other hand, however, we gain very little value from any of this:
the new versions of the binary packages are all released at the same
time.

Let's drop this and just pin the pages to the same cockpit-bridge
version.  This also lets us drop our `override_dh_gencontrol:` rule from
debian/rules.