git-hook-post-commit: deal with deleted files
Deleting files will currently fail the post-commit hook (and therefore
the pre-push hook) with a message like

  fatal: pathspec 'tools/gen-spec-dependencies' did not match any files
  tar: This does not look like a tar archive
  tar: Exiting with failure status due to previous errors
   - 3f6fecd00 build: produce static cockpit.spec for release

This is caused by the fact that the filename shows up in the list of
changed files, but can't be extracted from the commit via git-archive,
because it doesn't exist anymore.

Replace our use of git-show with git-diff-tree, which is a better fit
anyway).  It also has a very nice feature to filter certain types of
events (eg. --diff-filter=d for deletes).