packit.yaml: stop clobbering git config
It turns out that it's possible to install packit locally and run

  packit srpm

right in your git working tree.  Our current packit.yaml is a bit
unfriendly for that, though, since it clobbers your git config.

Now that webpack-jumpstart has a default EMAIL environment variable
fallback, that shouldn't be necessary anymore, so stop doing it.