test: Give starting kdump.service more time in TestKDumpNFS
Enabling kdump for the first time rebuilds the initrd, which is
inherently slow. On busy CI machines it sometimes takes more than a
minute.

The `TestKDump` class already does this by bumping the global timeout.
Be more specific here, to avoid hiding UI issues for operations that
should be fast.