test: Fold remaining relevant piece of check-system-s4u into TestIPA.testClientCertAuthentication()
check-system-s4u started out as a playground/proof of concept for using
S4U kerberos certificates to transcend smart card/certificate
authentication to ssh and sudo. Since then, this functionality has been
implemented in cockpit, and covered by TestIPA.testClientCertAuthentication().

The only remaining piece was to check that when ssh'ing to a remote
machine, the S4U ticket works for authenticating to sudo as well. The
forwarding is currently misssing from both `ssh -K` and cockpit-ssh;
this is tracked in https://issues.redhat.com/browse/COCKPIT-643. But
it's useful to keep the test as a reminder how it *should* work
eventually. So add that last scenario to testClientCertAuthentication().

This allows us to drop the check-system-s4u check entirely. It is one of
our most expensive tests.