pkg: stop including cross-package translations
A great number of our pages contain this:

  <script type="text/javascript" src="../*/po.js"></script>

in their .html files.  The effect is that the page pulls in the
translations for every package that the bridge knows about, including
those that may have come from a different source package.

This was introduced in #5898, but unfortunately the PR is a bit thin on
rationale for why this was done.

Each page really only needs it own translations, plus the translations
for base1.  Adjust accordingly.

See also discussion in #13906.

This changes the caching behaviour and aggravates the Chromium caching
bug for PO files (#8160), and thus now depends on fixing this behaviour
(commit b915d694de63ea). Thus skip TestApps.testL10N on
rhel-8-6-distropkg, as a simple browser reload is not sufficient any
more to reload the po file.