test: Eliminate run-tests --batches option
Always allocating about half of `$TEST_JOBS` to nondestructive tests works
poorly in a lot of cases: It wastes a lot of capacity for projects which
have mostly nondestructive (like cockpit-machines) or mostly/only
destructive tests (like subscription-manager).

Now that the global machines clean themselves up as soon as they are not
needed any more, we can optimize this. Start with a simple approach of
getting the  tests out of the way as fast as possible, and let the
destructive ones fill in the remaining available slots. This is easy
enough to tweak now, by applying a different heuristic to `num_global`.
Let's check the statistics in a few weeks and let this be data driven.

We can get away with dropping the `--batches` option because (1)
test/common API is versioned, and (2) none of our projects actually uses
that, they all rely on the builtin default.