all: Fix breadcrumb navigation
The shell will turn the empty hash "#" into "#/", and when it sets the
"src" attribute of the iframe to "#/" that has just set its location
to just "#", the browser will treat that as a navigation event and add
a history entry.  So let's use "#/" to avoid that if possible.

If we really need a onClick handler, don't use "to" but style the item
as a link.