test: Fix virbr0/firewalld inconsistencies in network pixel tests
The previous networking pixel tests were wildly inconsistent in their
assumed network interfaces: The references for basic and team had no
"virbr0", while the references for bond and bridge did. This was not a
conscious decision, just an arbitrary result of in whichever order
the tests happened to run when generating the references.

Stop the libvirt default network (i.e. virbr0) at the beginning of all
of these tests. In addition, the basic test assumed that firewalld was
running, so start it to actually make sure of that.

This makes these tests independent from whether
check-networkmanager-firewall runs before or after them.