test: Fix race condition in TestFirewall.testNetworkingPage
This test often failed because starting firewalld asynchronously started
the libvirt default network, and then added a new "libvirt" zone for it.
This then broken the zone count between the test's initial
`active_zones` count and what the UI actually displayed later on.

There was a special case for ubuntu-2004 which fixed that, but it did
not run on any other OS (the same thing happens on ubuntu-stable and
both debian-*).

Commit dad691c85a5 was supposed to fix this for all networking tests,
but it did not apply to TestFirewall as that just skipped NetworkCase's
constructor. Call the proper parent constructor to include that, restart
firewalld to make sure it picks up the dropped virbr0 interface (it
otherwise takes a while for it to realize), and drop the now obsolete
ubuntu-2004 special case.