eslintrc: reenable array callback return rule and fix all related errors
Reasoning:
A map() creates an array, so a return is expected for all code paths (if/elses).
If one does not want an array or to return data, they should use forEach instead.
Similar for filter().

An expection in our codebase was cockpit.transport.filter which confused
the linter, thinking that its an array filter. I disabled linting on these
lines for this rule.