users: Call `date` instead of using Date() to parse `lastlog` output
Date() has trouble parsing some dates/times; e.g. Firefox 91 and 97:

    > new Date('Jan 4 11:52:37 -0500 2022')
    Invalid Date

    > new Date('Jan 4 11:52:37 +0100 2022')
    Date Tue Jan 04 2022 11:52:37 GMT+0100 (Central European Standard Time)

The Date documentation [1] also explicitly discourages using it for
parsing dates if they are not in ISO format (which the above is very
much not).

As `lastlog` has no way of outputting timestamps (or machine readable
output, for that matter), call `date` to convert the date/time string
into a time stamp. However, still keep the JavaScript dissection of the
lastlog string, as cutting out the string does not work at all with
`cut` and `sed`, and is way too complicated with `awk` even.

[1] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/Date