python: major refactors
Split things into separate files.

Improve class hierarchy for Channels (new Endpoint class) and also for
Protocol subclasses.

Make match rules more explicit and more flexible, closer to how they are
in the C bridge.

Add the start of a Peer functionality.

Very many changes to function names, virtual methods, style, and other
things.

This commit also introduces an as-yet-unused file `bootloader.py`.  This
is intended to be run on remote systems via ssh.  If it detects that the
Python bridge is not yet cached, it will request that it be sent from
the connecting session (ie: `cockpit-ws`).  Then it will start the
bridge.  This is the main bit of how we actually get our .pyz onto the
remote machine.