networking: Clarify return values of call_object_method
This is a continuation of adb9726823751011327ae41414b114fc8869685b.

All callers of call_object_method now explicitly convert the array of
return values into a single return value, as appropriate for the
specific method.

This still uses Cockpit's promises, but we don't go back to the old
multi-argument "then" functions because standard promises don't
support them. Thus, the code is now correct again and also ready to be
moved to standard promises.