man: change the language= tag of config examples
Several config examples in the cockpit.conf manpage were tagged with
language="js".  It turns out that this is double wrong: not only is this
not javascript code, but even if it were, the Docbook XSL stylesheets
would expect "javascript", not "js".

Change it to "ini" (which is supported by docbook-xsl).