bridge: compile common metrics code once
We currently create two libraries (libcockpit-bridge.a and
libcockpit-pcp.a) which share a great number of source files.  Because
of differing CFLAGS, those files are compiled twice: once per each lib.

Create a libcockpit-metrics.a to contain those files, which are now only
built once, and depend on the new library from cockpit-bridge and
cockpit-pcp.