ws: Don't cache po.js files forever
This causes at least Chromium to not ever re-request the file from the
webserver, so that switching languages does not have any effect.

This is a heavy hammer, as it generally disables caching po.js even
when not switching languages (the common case). So this should be
replaced with a better design at some point.

Fixes #8160