build: lift docs rules out of toplevel Makefile
This `if ENABLE_DOC` soup has no business being in the toplevel
Makefile.  Its reasons for being there are somewhat complicated, and
have to do with automake's distaste for ::-style rules mixed with
conditionals.  We've often written rules like

    some-hook::
    if ENABLE_DOC
            do_stuff
    endif

for this reason.

Invent a new approach that involves unconditionally adding a single
dependency-only line for each hook, having it depend on a per-hook
_TARGETS variable.  The individual targets can now easily conditionally
add themselves to the variable without triggering automake's wrath,
which allows for things to be written much more cleanly.

Neaten up the docs stuff accordingly, and move it to the makefile in
doc/guide.  Also use the dist_ trick (instead of EXTRA_DIST) to get
files in the tarball regardless of which conditions are enabled.