build: use EXTRA_DIST for testing node_modules
Exempt these from our tar machinery and use EXTRA_DIST to make sure they
end up in the tarball.  The tar approach doesn't work nicely with
out-of-tree builds, because we end up with the name of the srcdir
prefixing the file in the archive.

Add a rule to pkg/build so that make knows to run the package-lock.json
rule in order to populate files in node_modules/.