tools/make-dist: new script to make a tarball
This creates a tarball in the least-invasive way possible.  If the
working directory is already configured, then it just runs `make dist`.
If it's not, then we autogen the current directly, but we don't
configure it.  Instead: we create a temporary directory, do the minimal
"mock" configuration, and run `make dist` there.

This means that we always are able to get a tarball, and often with a
very small number of dependencies, but never break the working tree
which a human might be hacking in.

There is currently some amount of duplication with test/make_dist.py and
autogen.sh, but it will be removed in following commits.