build: remove custom `install` rules for webpack
Now that we don't have to deal with on-the-fly compression of .gz files,
we can use a much easier approach to installing our webpack output.

automake's nobase_ prefix is pretty much exactly designed for this sort
of "preserve tree structure" installation.  There's only one catch:
since we have no way to drop the `dist/` prefix, we do a workaround:
create a symlink from `cockpit/` to `dist/` and install the files, via
that symlink, directly into `$(datadir)`.

That gets us names like `/usr/share/cockpit/base1/cockpit.js.gz`, as
before.

It also means that it's now possible to point XDG_DATA_DIRS directly at
the build directory, for testing purposes.

One issue this introduces is how to deal with the manifests for pcp and
ssh: the tar rule managed to strip the pkg/ prefix from those, but now
we have to install them a bit more explicitly.