test: Reduce multi-VM test memory
Our test VMs get 1.1 GiB of RAM by default, which is enough to fit more
complex tests. However, this demands a lot of unnecessary RAM in tests
which provision multiple VMs.

Explicitly specify the RAM for these tests; strive for making the sum of
all VMs not exceed 1.1 GiB. For example, the networking tests only start
a second machine to provide a DHCP server -- that its completely happy
with just 256 MiB. This isn't always possible -- e.g. candlepin or
FreeIPA on the services image need a lot of RAM, and the host
switching/multi-machine tests run three cockpit machines, which can only
be made that small. But at least severely reduce their size, so that
they don't exceed 2 GiB in total.

This makes it easier to plan capacity for our infrastructure (for
calculating maximum allowable parallelism), and also makes it easier for
developers to run these tests on laptops which don't have tons of RAM.