metrics: Use cockpit.format_bytes() for RAM/Swap sizes
Use our standard formatting API to get consistent results, and more
legible numbers if there is very little (< 1 GiB, for VMs) or a lot
(TiB, for large servers).

Also drop the explicit base-10 formatting for top memory consumers. It
is confusing to show the total memory in base-2, but the service memory
in base-10. We will change this soon, but in a global manner. In the
meantime, adjust the integration test's expected numbers from MB to MiB.

Accept any false-y value for `swapTotal` as "no swap present". `null` or
even `0` is equivalent, and this fixes the inconsistent comparison
against `undefined` or just being false-y.