metrics: History metrics react to visibility changes
Immediately refresh the history metrics when the Metrics page becomes
visible. This gives a better UX than having to wait for the next
one-minute auto-refresh timer.

Conversely, don't reload/refresh the history metrics at all if the page
is not visible, as that's just a waste of CPU and bandwidth.