test: Lower default CDP timeout to 15s
One full minute is way too long for our UI to react for our UI
standards. In general, our UI should react very fast even on loaded CI
machines. The main exception are waiting for OS APIs that genuinely take
a lot of time (such as formatting a drive or joining a FreeIPA domain),
but these should (and often already do) have an explicit wait_timeout().

This also makes development of tests less annoying, and speeds up
known-failing tests (with naughty matching).

Lower the default timeout to 15s, drop the redundant low timeout in
TestConnection, and add explicit timeouts for legitimately long-running
steps. This makes it easy for us to grep the tests for things that are
slow, and worth looking into.

The biggest offender is the Services page, which is known to be dog
slow. Until we fix this properly, liberally and widely sprinkle
increased timeouts over the tests.