test: Fix pam_faillock configuration in Debian/Ubuntu
Don't overwrite the standard pam_deny/pam_permit lines. They are there
for a very good reason, and we should keep the config changes to a
minimum.

In particular, `auth sufficient pam_faillock authsucc` does *not* fail
the login, so it needs the `pam_deny` fallback. Alternatively, the
module could be `required`, but let's stick to what the manpage says.

This resolves https://launchpad.net/bugs/1966416 and the corresponding
https://github.com/cockpit-project/bots/issues/3146