storaged: handle nfs flake due to switching elements
On packit and sometimes in our own CI the nfs related tests flake due to
to the a race condition between the Combobox changing from an input to a
select. As switching elements is also not user friendly for users as the
inserts text can be discarded drop this behaviour. The select element
already has the option to create a so not showing the input still allows
users to input share if showmount does not work.

Closes #17142