tools/git-utils.sh: new tar_to_cache() function
This reads a tar file from stdin directly into the cache, turning it
into a tree object, and returning the object ID on its stdout.

Nothing uses it yet.  We introduce it here in a separate commit to
explain some finer points about the implementation.  The points below
are discussed in context of `node_modules/` which will be the first user
of this new API.

First: on my system, `mktemp --tmpdir` creates files in `/tmp` which is
a tmpfs with 16GB free.  A typical `node_modules` checkout is on the
order of a few hundred MB, so it will fit there quite nicely, while at
the same time providing high performance.  We're very careful to make
sure that we always delete the directory, regardless of how the script
exits (error, signal, anything else).  During development I haven't seen
the temporary directory get left behind even a single time.

Next: the way that git creates commits is essentially grouped into three
stages:

  - read the objects into the object store, computing blob checksums
    (which are stored into the index)

  - turn the index into a tree object

  - create the commit referring to this tree

In order to complete the first step, you don't really exactly need a
working directly: an index will do.  We use the GIT_INDEX_FILE
environment variable to create an index in our temporary directory which
we'll write to while writing file contents directly into the cache. This
index will disappear automatically when the temporary directory gets
cleaned up.

A possible alternative approach to this (and the one that I tried first)
would be to just create a `node_modules/` subdirectory, initialize it
as a git repository, extract the .tar file there, and commit the result.
This worked, but it wasn't great:

 - the node_modules/ directory ended up in a different state than if we
   had just used `tools/node-modules checkout`, which impacted its
   treatment when running `git pull`, for example

 - the produced commit would need to be synced over to the cache anyway
   before pushing or verifying it (concepts introduced in future
   commits)

 - we needed to add extra consistency checks that were already performed
   by `tools/node-modules checkout`

Reading the result of `npm install` directly into the cache means that
we can simply commit it, and then use the `checkout` commit to get it
into the tree in a way that's entirely consistent with normal operation.