reposchutz, tools/node-modules: a new security approach
Use the newly introduced `tools/node-modules verify` subcommand to
perform a much more thorough check for `package.json` vs `node_modules/`
consistency from reposchutz.  This check needs to pass before any commit
can land.

Because the check is now much more expensive, we run it only on commits
that actually changed `package.json` or `node_modules/`.

Since we now invoke the `tools/node-modules` script to do our work for
us, we require a checkout of cockpit in order to do the work.  We also
stop abusing the shared git directory that we were using to download the
objects from the node-cache: we let the script use its cache in the
usual location.

This new approach to verifying the pushed `node_modules` commits means
that we can now safely allow anyone to push these commits to the
`node-cache` repository.  Add a `tools/node-modules push` command to
enable that.