workflows/npm-update(-pf).yml: change the environment
Instead of running the npm-update workflow in the `self` environment and
using SSH to push, push via HTTPS (with the `contents: write` permission
now granted).

This frees us to use the `node-cache` environment, which lets use use
the `DEPLOY_KEY` secret from there with `tools/node-modules push`.

Eventually we may want to create an environment with both deploy keys
inside of it so that we can force-push the commit after we create the
PR, in order to trigger the testing workflows to run.  Since we're not
actually doing this force-push at the moment anyway, let's take this
approach as the easier one.