workflows/npm-update{,-pf}.yml: force push via SSH
Add a new environment in .github/deploy-keys.sh which has access to push
to both the main repo and to the node-cache repo.  Use this new
environment from the npm-update workflow.

The main benefit of pushing via SSH is that after we use the GitHub
token to open the pull request, we can force-push the branch with the
deploy key.  That should trigger a synchronize event on the PR, causing
all of the tests to start running.  This doesn't happen when pushing
with a GitHub token (as a guard against infinite recursion).

Note that we have a bit of a workaround at present: making both SSH keys
available while running the `bots/npm-update` script produces a problem:
GitHub doesn't know which user to authenticate us as when we connect.
We avoid this problem by doing the first push via HTTPS.