reposchutz: use fetch-depth: 0
We use `--not` with `git log` to find the merge base of the base branch
with the proposed branch, which doesn't work if we fetched them shallow.

I had thought that `git fetch` would fix this, but it turns out not to
sufficiently de-shallow the already-checked-out target branch.

Just do `fetch-depth: 0` from the start to avoid this non-sense.