build: pass --nonet to appstream-util validate
We validate our metainfo files as part of `make dist`, which tries to
access the network to check if the screenshots for Cockpit Client look
OK.  That breaks when trying to `make dist` inside a container without
network access (and could be a cause of instability anyway).

Use --nonet to make sure it stays off the network.