tls: avoid static port in test-tls-server
We've run into issues where port 9123 isn't available on the builder
we're using, so let's use a dynamically allocated port instead.

We accomplish this by passing 0 as the port number (which already
worked) but we need to have a way to figure out what port we actually
used.  We add a minimal accessor to the server code to let us get the
listener socket fd, which we can call getsockname() on to determine to
where we ought to connect.