containers: Simplify package installation for bastion container
The install-rpms.sh script is way more generic than it needs to be,
still from a time when we had a kubernetes container. It also does not
work on a standard fedora:36 container any more due to the `util-linux`
dependency.

Replace this with a direct rpm/dnf invocation in install.sh. Drop the
`--nodeps` mode, we don't need this and it's conceptually very
questionable.