git-hook-pre-push: bring back realpath -m
This reverts a tiny piece of 18e27b0d5f2d4b5f4d0776afd4fb0185fdbf1e3b.

This script is typically run by git from a symlink located in the
~/.git/hooks/ directory, which results in `$0` pointing to that link, and
not to the original file.  In order to make sure we resolve the path
relative to the location of the original file, go back to using
realpath, as before.