packit: Bump FMF test plan ref on release
Fedora dist-git FMF test plans have to specify a git repo
and fixed ref for where to find the tests to run. This is bad, and at
some point it should instead just discover the tests from the upstream
source tarball instead. Until https://github.com/teemtee/tmt/issues/585
gets fixed, we need to bump `ref:` in dist-git's plans/upstream.fmf.

Cockpituous had a hack for this [1], which we need to transplant to
packit now. Unfortunately this reintroduces `files_to_sync:`, but all of
this will go away when the follow-up issue
https://github.com/teemtee/tmt/issues/978 gets addressed.

Create the file in tmp/ instead of plans/ (from where it would be easier
to sync) as packit refuses operation on an unclean tree. tmp/ is in
.gitignore.

[1] https://github.com/cockpit-project/cockpituous/commit/2ef3f6c9991