test: Don't write successful test output in run-tests
A considerable part of our test logs consists of the same browser
messages and boilerplate over and over again. These are uninteresting,
so don't write the test output for successful test in TAP mode. Only
write the last line, as it contains the test's runtime (which we parse
for metrics).