client: fix dist in presence of unannotated tags
We use `%(if)%(taggerdate)` to make sure we only output data for
annotated tags, but unfortunately, with `--shell`, we get an empty blob
in case unannotated tags are found in the tree.

For some complicated reasons, this leads to a very confusing failure
that looks like:

    releases.extend(format_release(r) for r in get_releases())
        TypeError: expected sequence, not "generator"

Let's filter the releases so that we don't try to process empty ones.