tools: Eliminate required_base computation from spec file
The use case for this is obsolete: In RHEL 8 packaging we haven't had an
actual version de-sync between BaseOS/Appstream, and in CentOS/RHEL 9
this cannot happen at all any more as there is just a single source
package.

We only still need this for testing rhel-8-N-distropkg, but this doesn't
justify the overhead any more. Hardcode the current version requirement
(it does not change that often).

Once we drop supporting RHEL 8 from main, we can just make all RPMs
depend on a current (`%{version}`) bridge and shell like we do in the
Debian packaging.