cockpit.js: Support different URL root for the websocket
We eventually want to support running cockpit-ws behind https://www.3scale.net/.
This requires two different API paths: `/api/<service>-<port>/` for HTTP
traffic, and `/wss/<service>-<port>/` for websocket connections.

cockpit.js needs to know which URL/path to contact for the websocket, so this
cannot be done with path rewriting on the server side only, or even with
multiple UrlRoot= values.

Thus introduce support for reading a `websocket-root` meta tag from the
webserver, similar to the existing `url-root` one. Nothing sets that tag
right now, as we are still not entirely sure how we want to configure
this. See https://github.com/cockpit-project/cockpit/pull/17473 how all
of this fits together.

This is the only part required on the client side (which is much harder
to change and get into distros than the ws container), so land this now.
It is harmless, and easy enough to revert if we won't use it after all.