base1: Drop legacy PF3 and cockpit.css
We will replace this with committing patternfly.css (aka cockpit.css) to
RHEL 8's dist-git instead. There is little sense in rebuilding the same
file over and over again, PF3 does not change any more, and we don't
even test this. So it's actually safer to ship a known-good version in
dist-git. If needed, we can rebuild it from an old rhel-8.y branch.