.github: run static code checks
Shockingly ./test/static-code does not run at the moment. In
83ae6ac5c44f5187a1 the way we run tests changed from a bash script to an
Python script which does try to run ./test/test-static-code but as we
are not in a .git directory it does not run. For now let make check run
again, in the future we probably want to optimise this to only run the
required linters.