test: Fix run-tests scheduling with little or no parallelism
Commit 7f7e21fc0 introduced a regression when the number of parallel
jobs was smaller than the cost of a test: These tests would never be
scheduled, and at the end the scheduler failed the "parallel_tests
should be empty" assertion.

Fix that by also scheduling the next parallel test if the number of
running tests is empty. That will cause the costly ones to run serially,
as intended.