test: More thoroughly wait for defined VM
Merely waiting for a VM to appear in `virsh list` is a race condition:
virt-install has two phases, and it un- and redefines the VM in
between. This races with the explicit deletion in the UI: If that
happens while virt-install is still running, the second phase would
re-define VM, and the test times out waiting for the VM to be gone.

Current cockpit-machine's assertDomainDefined() is much more thorough:
it waits for the VM to have the expected properties. Backport that.