python: remove "dbus_internal" hack
Remove the special-casing of the "internal" dbus channel and implement
it as a peer-to-peer connection inside of the bridge.

The other side of this connection lives inside of a new file,
`internal_endpoints.py`, which contains the objects that we want to
appear on the internal D-Bus.

Most of these are still stubbed out, but the parts that were implemented
before have been copied over.

Bump our dependency on systemd_ctypes for the required APIs for
peer-to-peer connections and object exporting.