test-channelresponse: use pidfd to watch for child exit
Sending signals to old pids is unsafe (due to recycling) and ineffective
(due to zombies).  The waitpid() approach messes too much with GLib's
internals.

Let's avoid both of these and use a pidfd to do our watching in a way
that's guaranteed to be race-free and effective, without mangling the
process state in a way that confuses GLib.