shell: Set iframe body background to PF's dark theme
When switching between pages, we create an iframe and display it when
it's still loading as the default html background is white this causes a
flash while switching. As we can't easily force a background color in
for example pkg/systemd/index.html as we disallow inline styles and
setting it via JavaScript might be too slow.