python: port stream channel to ProtocolChannel
Replace the StreamChannel with two new subclasses:
SubprocessStreamChannel and UnixStreamChannel, using the existing
'open' message key-based restrictions mechanism of the Channel class.

Add SubprocessTransport and SocketTransport subclasses of our fd-based
Transport class to support the two new channel implementatins.

We need a special helper for handling stderr, so add that as well.  This
is a significant improvement on the existing code which just did a
one-time read of stderr when sending the close message: that code would
have caused the process to block if the stderr buffered filled up.