test: Avoid multi-arch conflict on mock-pam-conv-mod.so
In Fedora you can install multiple architectures of a package
simultaneously. This causes a file conflict when trying to install both
the i686 and x86_64 architecture of "cockpit-tests", as they both ship
/usr/lib/cockpit-test-assets/mock-pam-conv-mod.so (and this is not a
magic multi-arch directory).

This is of no practical importance other than that rpminspect complains,
which causes a nasty red stain in gating.

Install the file into the proper PAM dir right away. This is properly
multi-arched in both Debian and Fedora, and tests then don't need to
know the exact directory any more. For people installing from upstream
this is still sufficiently guarded, as `make install` does not install
that module (only `make test-install` does).