build: patch <releases> in metainfo on install
When doing the flatpak install, check if the `DOWNSTREAM_RELEASES_XML`
environment variable points to a non-empty file.  If so, then patch the
installed metainfo for Cockpit Client to use the release history from
that external file.

Add a sources entry to the flatpak manifest file to inject the
`org.cockpit_project.CockpitClient.release.xml` file from the current
working directory when building the flatpak.

Make the prepare script touch this file if it doesn't exist.  An empty
file will be ignored, but if the file already exists, its contents will
be used for the version history of the build flatpak.