workflows: run pytest on pull requests
We run this in the unit-tests container, just like the other unit tests.

Add a couple of `make` targets for `pytest` and `pytest-cov`.  This
makes it easier to add ourselves to the existing test matrix (since all
other tests are based on `make` targets).

(Add also a missing .PHONY marker for the previously existing pycheck
target).

Mention the new targets in HACKING.md.