python: tighten up typing in samples.py
Very strictly speaking, the type of 'samples' being Dict[str, Any] is
correct.  It's not very helpful, however.

Define a type alias

  Samples = collections.defaultdict[str, Union[float, Dict[str, Union[float, None]]]]

which reflects the fact that samples is expected to be a defaultdict.
The keys are strings.

In case of non-instanced samples, the values are numbers ('float'
includes integer values, per PEP 484).

In case of instanced samples, the values are, themselves, dictionaries,
mapping instance names (strings) to numbers, or None.  Writing None here
is a supported operation — the cgroup sampler does it in some cases.

This change unleashed very many instances of this sort of error:

  src/cockpit/samples.py:295: error: Unsupported target for indexed assignment ("Union[float, Dict[str, Optional[float]]]")

in response to this sort of code:

  samples['network.interface.rx'][iface] = int(fields[1])

Those are legitimate errors — we have no way to be sure that the name
we're using there to fetch the dict (`network.interface.rx`) wasn't
written to by another sampler as a numeric (ie: non-instanced) value.
We should figure out a way to clean that up a bit.