python: return empty stderr on stream channel
When we request the stderr to be returned via the message field on a
stream channel, make sure we always return it — even in the case that
the stderr is empty.  Previously this possibility was being lost,
because we were using the empty string to mean "we didn't collect
stderr".

Use `None` as the "we didn't collect stderr" case, instead.