tests: fix periodic test-chan.html failure
The "filter message in" qunit test in test-chan.html fails periodically
due to a race.  The source of this race is a workaround that was added
in 6aa3fb6cb6c759fc43148666d8e504fd85d72e61:

  // HACK: send the reply asynchronously, so that mock_peer addEventListener() gets to see it before our own onrecv()
  window.setTimeout(() => self.send(null, JSON.stringify(reply), 0));

Unfortunately, this hack was only applied to the control messages, and
not data.  This meant that for our "filter message in" test case, for
example, we'd usually see the "one" "two" "three" data messages echoed
back to us immediately — and terminate the test — before receiving the
"ready" control message.

Every now and then, due to timing, we'd get the extra "ready" control
message, though, and get an error looking like:

  not ok 154 - Expected 11 assertions, but 14 were run, expected: undefined, got: null, test: filter message in, module: , source: at http://localhost:32847/dist/base1/test-chan.js:2:148292

The race seemed more likely to happen when the python bridge was in use,
and particularly more likely to happen while running tests under pytest
with coverage enabled (which the next commit will introduce).

Since window.setTimeout() makes no guarantees about delivery order,
let's move to a full-powered ordered queue for sending replies.  Now
that "ready" is always guaranteed to be received before the data
messages, bump the expected assert count to 14.