test-server: Change python bridge detection logic
Move the logic for detecting if we're running on the Python bridge out
of the testcases where it's used (4 separate locations) and replace it
with a single boolean value calculated inside test-server.

Instead of looking for "cockpit-bridge.pyz" in argv[0], assume that any
value containing "py" is python.  This will apply to the .pyz but also
to a direct invocation of python3.