flatpak: preserve xml declaration in installed metainfo
The flathub appdata validation requires an <?xml?> line.

If we want one of those, we need to pass the `xml_declaration=True`
kwarg to `ElementTree.write()`.

To make sure we catch issues like this for ourselves, validate the
in-tree metainfo before patching it, and validate the installed copy
after patching.