webpack-jumpstart: tweak partial mode
When we do webpack-jumpstart --partial, our previous approach was to
touch any file that had been modified relative to the version of the
tree that the jumpstart tarball had been built against.

This is theoretically sound, but removing the target file completely is
probably a more robust choice.  Let's just delete any webpacks that
mention changed files.

See also #17462