python: minor tweak for mypy happiness
On account of the fact that we initialize the `Channel.channel` field
with `None`, mypy thinks that this field must be typed `NoneType`.

That's false: it's a string.  Moreover, the string will always be filled
in, since the first thing that happens on the channel is that the 'open'
handler gets called.

Replace `None` with the empty string.