python: remove 'watcher' global state
We were setting a _watcher class variable on the SubprocessTransport as
a cache of our chosen child watcher.  That mostly worked, since we kept
one event loop for the duration of the bridge, but it was causing
problems in the unit tests, since each of them got a fresh loop.

We were working around that in TestSubprocessTransport by always
resetting the cached object back to None, but as we add more tests that
use subprocesses (like the upcoming support for superuser bridges), this
approach would be more and more unwieldy.

Drop the global state entirely, and switch to attaching our watcher to
the loop, using a qdata-inspired approach.  That way, our object will
live as long as the loop, and go away when the loop goes away.