Allow root login on RHEL 8
With the disallowed-users feature we now disable root login by default
in Cockpit. In RHEL 8 however, a default install can login as root over
ssh with password authentication so allow this as well in Cockpit.

The /etc/cockpit/disallowed-users file is kept as then administrators
can optionally easily disable root login when wanted.