shell: fix white flash regression by dark theme switcher
On `readystatechange` is not soon enough for users to notice a white
flash while switching. The setting of the iframe's background color was
removed in the dark theme switcher commit f8761acb1e029b08 while it's
still needed. As we show the iframe while the content is still being
rendered it will show a white flash.