systemd: disable create timer button on unit load
In our tests we have a flake where `boot_timer.timer` does not show up
after it has been created. This is caused due to `loadingUnits` being
`true` so when we press create we do not refresh the list again. This
race condition is now resolved as we don't allow a user to create a new
timer if we are still loading units.