test: Run lintian on Debian/Ubuntu packages
Let's make sure we catch regressions when they happen. Ignore
"initial-upload-closes-no-bugs" which is due to our generic packaging
without changelog; and "newer-standards-version" which is due to
backporting.

We also have to supress "mismatched-override" for older releases. This
is usually a good hint to update obsolete overrides, but Debian 11's and
Ubuntu 22.04's lintian does not yet complain about two of our source
files and thus triggers these warnings. But we run it at least on newer
versions.

Also enable informational messages. There is a few things to clean up
there, but they are not important, so don't make them fatal.