tests: pixel-testing: introduce a scenario for RTL
This commit introduces a new pixel test scenario for RTL (Right to Left). This
scenario keeps using English for language but changes the direction of the language
to be RTL.

The reasoning behind this choice is that pixel tests take screenshots in the
middle of tests execution, by dynamically adopting the theme, browser
size etc for the different scenarios.
However in order to change the UI language one needs to reload the
page, which effectively means losing the application state. For this
reason, we would need to set the language at the very beginning of
each test if we wanted to properly test RTL with a language like Hebrew.

Since this means making a whole test run in another language, we decided
that we will for the time being test RTL with English.
This is obviously an unrealistic scenario but allows us to view spacing
issues at least.