test: Restore /var/log/wtmp for nondestructive tests
Previously, `TestAccounts.testUnprivileged` left behind recent logins of
the "anton" user in wtmp. This broke `TestAccounts.testBasic` when it
ran afterwards, as the pixel test did not expect existing "Last logins:"
entries.

Users are already cleaned up/restored between tests, so restore wtmp as
well to fix this in a generic way.