python: Debug-log watch errors
While we send errors from "cockpit.dbus.watch()" to the bridge client,
our current cockpit.js seems to ignore these. This makes it really hard
to debug broken ObjectManagers.  Add a debug message for that, so that
they appear in the journal in debug mode.

We shouldn't log them with a higher priority, as this can be legitimate.
E.g. the storage page routinely tries to talk to stratis, and just
ignores if it's not present.