python: Add temporary ${libexecdir} expansion in manifests
This provides the moral equivalent of what commit f87f57960bb4 did for
the C bridge. However, we want to avoid parameterizing the py bridge
with ./configure arguments, to keep it buildable with pip. Detect the
path at runtime instead, there are only a few plausible options anyway.
This is a temporary hack until we eliminate all `${libexecdir}` from
manifests.

This fixes joining an IPA domain, as it now finds
cockpit-certificate-helper. TestIPA.testQualifiedUsers now gets much
further, unfortunately it still fails on the remote host part of the
test, as cockpit-ssh does not work yet.

It also fixes the cockpit-{ssh,pcp,askpass} paths in the manifests, but as we
don't do that routing yet, it does not fix any of the ssh/pcp tests yet.
But it is a necessary prerequisite for actually implementing these.