webpack.config.js: don't gzip .html pages
The python bridge doesn't know how to decompress gzip, and -ws requires
HTML to be decompressed so that it can patch it while serving it to the
client.

All of our .html pages are small anyway, and since we're serving them
uncompressed, we may as well also have them on disk uncompressed, saving
ourselves the trouble.

This fixes TestTerminal.testOnlyTerminal with the python bridge.  That
test writes a cockpit.conf containing:

    [WebService]
    Shell = /system/terminal.html

which breaks unless cockpit-ws is able to patch html files (which it can
only do when the file is served in uncompressed form by the bridge).