users: Fix calling `last` in endless loop
Using `[]` as "uninitialized" state for `logins` caused the useEffect()
to run in an endless loop, calling `last` hundreds of times per second.
It's legal for a user to not have any recent logins, in which case the
effect would call `setLogins([])`, and re-triggering itself.

Robustify this by moving to useInit().