systemd: Always use reportd on the system bus
The old code would connect to either the session or system bus,
depending on whether the session has admin privileges or not, but in a
racy way.

There is a time during initialization when superuser.allowed is "null"
and if a ReportingTable would be constructed in that time, it would
erronously connect to the session bus and not get any workflows.

Sine a ReportingTable is only shown in the first place when the
session has admin privs, let's just always connect to the system bus,
just like we always connect to the system bus for problemd.