selinux: Fix React crash on API/CLI errors
When cockpit.script(runCommmand) or deleting an error fails, we catch an
"error" object. Don't try to stuff that into the `<Alert>` as-is, as it
will make React crash with

    RuntimeError: Error: Objects are not valid as a React child (found: [object Error]).

Convert the errors to a string first.

That does not fix the underlying reason/race condition why applying the
solution sometimes fails in TestSelinux.testTroubleshootAlerts, but it
will make the failure a lot more useful both for the user and the test.