python: reduce use of Any in typing hints
We use things like Dict[str, Any] in a few places around the code,
typically for things which we parsed from JSON.  This basically tells
mypy to stop paying attention.

Change that to the more correct type of Dict[str, object].  This results
in mypy complaining more, but that's good -- there are legitimate type
safety issues here.