packagekit: Add manifest condition
We only want to show this page if Packagekit is installed, and the
system is not an OSTree.

Update `TestUpdates.testNoPackageKit` to test the new behaviour:
Previously, the page would just show "PackageKit is not installed", now
it does not get shown at all.

Test the old and new behaviour based on `$TEST_SCENARIO`, so that the
test will fail as soon as we make the Python bridge the default and drop
the C bridge. This will remind us of updating the test and docs,
instead of letting it silently break (with a simple skip).

First half of https://github.com/cockpit-project/cockpit-ostree/issues/295