python: Fix typing in superuser.py
Fix `self.superuser_rules` type. The dict key is not optional, but the
label is (first tuple value).

Make the SuperuserRoutingRule params type stricter. Despite of that,
mypy is still not clever enough to see that `name` must be a string
after all the fallback computation. Help it by adding an assert.