python: kill subprocess on closing its transport
The standard library version of SubprocessTransport sends SIGKILL to the
spawned process when the transport is closed.  Let's do the same, but
with SIGTERM (to better match historical bridge behaviour).

We need to be a bit more careful about handling the _process variable:
if we throw an exception while spawning the process, we'll end up with a
half-initailized SubprocessTransport instance which (via `__del__`)
needs to work properly with `_close()`.  Set a default value of `None`
and check for it.