python: catch BusError on Properties.Get
For DBus interfaces that notify property invalidations (without
specifying the new value) we immediately try to "Get" the new value so
that we can send it to JS.  The call to do this was missing a try: to
catch the BusError which might result if that Get call fails.

Catch the failure and log a debug message.

Fixes #18358