test: Ensure that ABRT runs in TestJournal.testAbrt*
This isn't the case when the package just got installed, like in TMT
runs. On our test VMs this is already the case, thus a no-op.

Reset the sysctl settings for the tests too, as something resets
/proc/sys/kernel/core_pattern in between installing the packages and
starting the ABRT tests on TF machines. The kernel just comments that
with

```
Unsafe core_pattern used with fs.suid_dumpable=2.
Pipe handler or fully qualified core dump path required.
Set kernel.core_pattern before fs.suid_dumpable.
```

It's not clear what exactly causes this, but it's more robust and does
not hurt to do this right before the test.