systemd/services: Eliminate duplicate usePageLocation
The top-level ServicesPage already maintains a usePageLocation object,
we don't need to have another one in ServicesPageFilters. We already pass
the "path" half of it through ServicesPageBody to ServicesPageFilters,
so just pass on the "options" half as well. This avoids duplicate event
handlers and makes debugging event handling a bit easier.