systemd/services: Read CLOCK_BOOTTIME with Python, fall back to /proc/timer_list
To correctly interpret monotonic timers, we need to read
CLOCK_MONOTONIC. This cannot be read with shell tools, so try to read it
with Python (which we will soon depend on anyway).

If that fails (e.g. on CoreOS), fall back to reading /proc/timer_list.
This is readable by root only, so it's not a good primary source of
clock reading.

Fixes #14818