test: Disable preloading for release builds as well
In TestHostSwitching.testBasic, the auxiliary machines often fail to
load software updates. This breaks the pixel test, as there is an
additional icon in the main menu which also changes the column sizes.

Change setup_provisioned_hosts() to always respect the `disable_preload`
argument. If necessary, we can do the `self.is_devel_build` in the
caller, but it's not really relevant: We already have several explicit
integration tests to check the preload functionality (e.g.
TestUpdates.testBasic).

The recent commit 39b8be0ba939 also overzealously moved a comment. Move
it back, and adjust it for the updated situation above.