HACKING: Emphasize webpack-watch based development process
Most contributors want to work on the web (HTML/JavaScript) parts of
Cockpit. For that it's not necessary to deal with autogen.sh, make,
etc., as that's really slow and irrelevant.

Thus move the webpack-watch part to the top, and stop treating it as
alternative workflow: It's the primary and recommended way to develop the
session pages. Don't make new contributors suffer through building things with
`make`.