base1: Drop cockpit.dbus.publish() and .meta()
Nothing in cockpit or any related project publishes a D-Bus interface.
This is only exercised in the unit tests. The Python bridge does not
implement this either.

It's hard to imagine a good use case for calling .meta() without
publishing an API. In theory, it could be used for calling D-Bus methods
on services which don't have introspection data by themselves, but (1)
this is exceedingly rare (not observed anywhere), and (2) better handled
by explicitly giving type signatures with .call().