lib: Make cockpit-po-plugin work with multiple entry points
Change the `subdir` option to be a `subdirs` list. Change the `wrapper`
option to be a subdir => string function instead of a static string.

This allows us to use the plugin with multiple entry points in the
future, which can all be built simultaneously.

This only affects cockpit itself. starter-kit like projects only have
one page/entry point and don't use these two options at all. The default
behaviour did not change.