build: Drop Makefile.deps mechanism and build dist/ in a single webpack run
This gives a massive speedup compared to building them individually, as
parsing/compiling `node_modules/` takes by far the most time:
`node_modules/.bin/webpack` is able to build the whole dist/ in 2
minutes (production with terser), instead of 27 linear minutes, or ~ 4
minutes with 8 parallel processes.

It's even faster in development mode without linters, then it just takes
about 40s.

`make` will now rebuild the whole dist/ if anything in pkg/ changes --
the same approach that starter-kit and friends have used for years. That
is already faster for a full "make dist" or jumpstart CI build. It is
slower for incremental builds, but (1) the documented and recommended
way to do that is with tools/webpack-watch and just building a single
page; and (2) it will not matter any more once we switch to esbuild.

We keep tools/webpack-{make.js,watch} for the time being. They are now a
rather shallow wrapper around invoking webpack, which we can simplify
further -- but that will require changing our habits and the
documentation, let's keep this for a follow-up.

This also paves the way for esbuild, as we don't have a mechanism to
generate Makefile.deps there.