packit: Re-add post-upstream-clone to copr_build
In commit cfb7ca737 I dropped the copr_build's `post-upstream-clone:`
action as it became identical to the global one, and I thought it would
be inherited. But it turns out that specifying `actions:` replaces the
global object entirely. That made release COPR builds fail on missing
spec file.

Copy the post-upstream-clone command and add a comment why it's
necessary. We'll have to keep this until
https://github.com/packit/packit-service/issues/1505 gets fixed.