python: make sure the bridge has an SSH_AUTH_SOCK
If the main bridge (ie: not the root one) gets started without
SSH_AUTH_SOCK set, then make sure we start an ssh-agent of our own for
the purpose.

I've tried a number of different approaches to who creates the socket
path and how the agent should be signalled to exit.

Sometimes (in several test cases, for example) the bridge gets killed
with SIGTERM, and in that case, Python more or less terminates
immediately (ie: default signal handling).  That means things like
context managers and TemporaryDirectory won't work.  We could try to
intercept SIGTERM and trigger cleanups, but we could quickly find
ourselves in a situation where we may have to clean up *too* much (ie:
cancelling and collecting all running tasks, etc.) and I'd rather avoid
it.  For that reason, let's leave SIGTERM as it is and handle the agent
exit via prctl().