python: remove mandatory message from CockpitProblem
Protocol errors always have a human-readable error message, but the sort
of errors we'll throw from failed SSH connections usually don't come
with such messages.

Change this into just another kwarg in the general case (but use it as
the main message for the exception if it's specified).

Now "problem" is the only field which is absolutely always present.