test: Mark TestStorageHidden.testHiddenSnap as nondestructive
Keep testHiddenRaid() as destructive for now, as it is too hard/brittle
to tear down the MD device if the test fails in the middle.

Also, testHiddenLuks() often runs into memory errors in a long-running
VM, due to the expensive KDF. So keep that one as destructive for now as
well.