build: fix `make install` for relative `DESTDIR=`
The rule for installing our bundles looks like:

   cd $(srcdir)/dist; find -type f -exec install -D -m 644 '{}' '$(DESTDIR)$(datadir)/cockpit/{}' \;

which is fine if $(DESTDIR) is given as an absolute path (or omitted).

If $(DESTDIR) is relative, however, then its meaning will change as a
result of the `cd $(srcdir)/dist`.  For example:

  `make install DESTDIR=install`

will end up copying the bundles into `$(srcdir)/dist/install/`.

That's particularly problematic because if you repeat it, the future
invocations will consider the files it finds in `dist/install/` to also
be web bundles, and copy them yet again, creating exponential growth
over multiple invocations.

We can avoid this issue by wrapping `$(DESTDIR)$(datadir)` in
`$(abspath)`. That gets evaluated by make before the working directory
is changed. Including `$(datadir)` makes sure that things will work also
when `DESTDIR` is empty.