test-server: stop using /tmp/dbus-tests as tmpdir
This used to work with older versions of GLib, even if the directory
didn't exist, but it seems to have stopped working with more recent
versions (ie: Fedora 38).

It's also a security (or at least denial of service) bug: if someone
else creates that directory before us, we can't use it.

Just use the top-level /tmp directory instead, which solves both
problems.