python: migrate setuptools config to setup.cfg
The version of setuptools required to support pyproject.toml isn't
available in any current version of RHEL, so let's take a step back for
now and port our config to setup.cfg.

This lets us drop our setuptools dependency to 39.2.0, which is,
conveniently, the version available in RHEL8 as well as the minimum
version that supports dynamic version numbers (which we will soon use).

While we're at it, switch our package-finding logic to the
fully-automatic form, adding an __init__.py to cockpit.data to make sure
it gets included.