networkmanager: Recreate the D-Bus watch when the service restarts
The watch doesn't survive a restart of the D-Bus service, let's
recreate it when a new owner appears.

This didn't seem to matter much in the past since NetworkManager used
to have a very incremental startup and would send out
PropertiesChanged notifications for a lot of things.  It is sending
less now, apparently.