python: de-indent a function in samples.py
This function has a long and inaccurate name (it has nothing to do with
cgroups) and doesn't need to be at class scope.  Move it to the
top-level and give it a more accurate name.

This is mostly about avoiding excessively-long lines, but it's logically
sensible in its own right.