test/static-code: fix our vulture setup
The way we've been using vulture is pretty ineffectual.  Cranking the
min-confidence all the way to 100 is skipping a lot of potentially
useful messages.  We've had to do that because vulture does
whole-program analysis and we run test/static-code from our post-common
hook only on the files that changed.

Let's move our vulture configuration to pyproject.toml, removing the
minimum confidence level while we're at it, so we get a lot more
suggestions.  We have to add some suppressions to deal with a lot of
those, but that's how the tool is intended to be used.

In order to avoid running vulture in the post-commit hook (where it only
has access to a subset of files) use an environment variable to signal
test/static-code that we're in this situation an skip vulture in that
case.

We effectively have two sets of configuration for vulture now:

 - what happens when you run `vulture`: this basically runs on `src/`
   and `test/pytest`, excluding the integration tests and testlib.  We
   do this because by default vulture only looks at *.py files, and our
   integration tests are executable scripts without extensions.  That
   would lead vulture to falsely conclude that most of testlib is
   unused.  This is also how vulture will run from ALE, if enabled.

 - when happens when you run `test-static/code`: in this case, we do
   more aggressively scanning for python files based on #!-lines.  In
   this case we scan all files (except for fmf_metadata) which means
   that the testlib appears (correctly) well-used.