testlib: drop several old and unused APIs
Some APIs were dropped from use, over the years, without being removed:

  d838d280bf "Table rows react now on onClick"
    - Browser.mousedown

  ed4db97600 "Drop cockpit-kubernetes"
    - Browser.dialog_complete
    - Browser.upload_file

  b7a908475b "lib: Redesign On/Off switch"
    - Browser.wait_attr_not_contains

  bdedda51db "services: Redesign details page"
    - Browser.click_action_btn
    - Browser.wait_action_btn

We drop most of those, but keep .get_checked(), .upload_file() and
.wait_attr_not_contains().  .get_checked() is used in cockpit-machines,
cockpit-podman and cockpit-certificates.  The other two aren't used in
any of our projects, but fit nicely with the API and we want to keep
them.

For those three cases we add a suppressions file (vulture calls this a
"whitelist") in order to indicate that we consider these functions
"used".