python: update ferny, implement cockpit.send-stderr
ferny stopped implementing our send-stderr command, which was only used
for us, so we have to implement it ourselves now, from our
InteractionResponder in superuser.py.

On the plus side, we can now use interaction_client.py to send custom
commands.  According to the ferny documentation, we should prefer this
over sending messages for ourselves.

Note: in comparison to the old send-stderr handling inside of ferny, our
new `cockpit.send-stderr` command no longer implies that the ferny
interaction is complete — we need to send `ferny.end` separately.  This
isn't actually required for the interaction to exit (since we'll send
our "init" message very soon anyway) but is required in some situations
to prevent ferny from interpreting our dropping the stderr socket as an
unexpected exit.