test: drop code from mock-bugzilla-server.py
This was flagged by vulture as being unused, and removing it doesn't
seem to break the test that it was designed to be used as part of.

It remains as an open question if this file is used for anything at all
anymore, but let's remove the part that vulture was upset about, as a
first step.