python: abort outstanding sudo auth in various cases
Let's make sure that we don't get blocked on an outstanding `askpass`
interaction when doing various things:

- hitting the Stop method on the login attempt

- reloading the packages and noticing that the auth method disappeared
from the manifests

- logout of the bridge

This is not the most elegant way to do it, but it's something like the
easiest.  It also doesn't solve a very similar problem we're bound to
have with SSH connections.

This fixes a couple of tests.

Some extra info:

The crux of the problem is this: when we shut down the bridge via EOF,
we try to close all of the connections that the bridge has open.
Connections that have been queued up on a frozen endpoint count as
"open" as far as the bridge is concerned, so when we send the "close"
messages for those, it'll be queued up right after the frozen "open"
message and will never go through.

Meanwhile, the endpoint will never be unfrozen because it's waiting for
an askpass response.  In turn, askpass will never exit, because it's
waiting for a D-Bus method call (the answer to the prompt) that cannot
possibly occur anymore.

I tried various solutions involving cancelling the entire ferny agent
interaction (part of which we will want to keep, for other reasons) and
tracking users of the "client" side of the internal bus and notifying
the exported objects on the "server" side when they're all gone, but
everything else got too complicated, which is why I settled on the
imperfect but very simple solution in this commit.

Fixes #18676