metrics: Fix y-flipped history polygons for saturation, and RTL orientation
SVG polygon's coordinates have Y go downward, but in our graph we expect
the time coordinate to go upward. So we always have to scaleY(-1).
Changing the numbers to be aligned to 100% instead of 0% is more
computational effort -- all transform()s are a single matrix
multiplication, which happens anyway.

Flip them on the X direction depending on RTL and whether it's use or
saturation.

This fixes the saturation graphs being the wrong way around in LTR, and
flips the horizontal orientation in RTL: i.e. in RTL, usage goes to the
right and saturation to the left.