python: various fixups to exception chaining
Writing code like this is almost never right:

    try:
        do_something()
    except PermissionError:
        raise ChannelError('access-denied')

The reason for that is because the explicit `raise` is indistinguishable
from other errors which might occur (unintentionally) inside of the
`except` block.  This leads to ugly-looking tracebacks with the familiar
"During handling of the above exception, another exception occurred"
language.

The better way to write this is using the familiar pattern:

    try:
        do_something()
    except PermissionError as exc:
        raise ChannelError('access-denied') from exc

which produces a more accurate "The above exception was the direct cause
of the following exception" message.

There's also another form with explicit `from None`, as follows:

    try:
        do_something()
    except KeyError:
        raise ChannelError('not-found') from None

which is what we should write when we explicit prefer to only see the
second exception, without reference to the original cause.

Make sure all of our re-raising is using one of the above two forms.

This is one of the changes required to enable bugbear linting (ruff
`select=["B"]`).