python: include clean-ups for isort
Theoretically, we want to separate things into five buckets (stdlib,
third-party, vendored, cockpit, local-folder) but we also want to take a
more or less stock isort behaviour, so we do this:

In our testcases, we want the "local folder" bucket to refer to the mock
utility classes in the pytest/ directory, while first-party refers to
cockpit.  Add __init__.py to test/ and test/pytest/ to ensure that we can
do relative imports (which we should have been doing already).

In src/ we make sure to always refer to our vendored packages by their
full name (`cockpit._vendor.*`) and to do relative imports from cockpit
itself.  That ends up putting our vendored packages in the first-party
bucket and cockpit into local-folder, which is approximately correct.

The only issue here is that if we directly include vendored packages
from our testcases, they end up in the same bucket with cockpit itself.
That's not a big problem.