test: Don't use CSS to replace DOM elements with arbitrary text
The tricky CSS tricks turned out to be a bit too tricky in the case
where the production CSS already uses ::after or ::before pseudo
elements.  Our ListingTable component uses those alot for handling the
mobile layout, for example.

Let's use to most straightforward way: setting the text of an element
via injected JavaScript.

This might fail when React decided to render inbetween the call to
set_mock and capturing of the actual screenshot.  But the tests keep
the pages quiet during pixel testing anyway, so hopefully this will
not happen in practice.

Another way this might fail is that React gets confused the next time
it renders after the call to set_mock because the DOM has changed too
much.  This can hopefully be prevented by only replacing the text of
elements, and not their children.