doc: Describe $XDG_CONFIG_DIRS
When set, Cockpit uses the first XDG_CONFIG_DIRS entry which contains
the requested file (such as /cockpit/cockpit.conf or
/cockpit/machines.d). Spell this out in the main cockpit.conf.xml
documentation, and shorthand reminders and a link to that to all the
other sections.

At the same time, drop the placement of an override file next to a
manifest as the primary method. It's a giant no-no for /usr/share,
pointless for ~/.local, and all but useless for /usr/local/share.

Fixes #18633