ws: Don't capture stderr for Cockpit Client when debugging
We capture the output of our ssh command in order to provide useful
error messages to the user of Cockpit client in case of failures.
Setting COCKPIT_DEBUG will result in a flood of messages to stderr,
however, and it's more useful to see those on the terminal than in the
login window with their formatting destroyed.

This makes it possible to do something like:

    flatpak run --env=COCKPIT_DEBUG=all org.cockpit_project.CockpitClient