pybridge: Respect $XDG_CONFIG_DIRS
Use XDG_CONFIG_DIRS for looking up package overrides and the machines.d/
directory as well, like the C bridge did. Note that this is not matching
the spec [1], but the current behaviour of ws and the C bridge has been
established for a long time already.

Eliminate the ETC_COCKPIT constant, and use a generalized version of the
`$XDG_CONFIG_DIRS` parser instead.

While at it, change that to return a `Path` object, and adjust the
callers to use it.

[1] https://specifications.freedesktop.org/basedir-spec/latest/ar01s03.html