pybridge: Detect `no-cockpit` peer error
When the remote peer program fails and EOFs, the process has not been
wait()ed for yet at the time when `eof_received()` gets called. This
makes it hard to tell apart different error cases. In particular, we
want to know whether e.g. a remote (SSH) cockpit-bridge is not installed
(shell code 127), or whether it failed for other reasons. The former is
supposed to fail with problem code `no-cockpit`.

To do this, defer the channel closing until `process_exited()`, where we
can check the exit code.

This introduces a new assumption about cockpit peers: That they exit
soon after closing their stdin. This may not be true for arbitrary
programs, but it's safe for cockpit-bridge/pcp/ferny. After all, the C
bridge waits for the exit code as well.

This fixes the "cockpit is not installed" failure in
testTroubleshooting.