test/verify: teach pytest how to load check-*
Add a conftest.py file to test/verify with a hook that allows us to load
our `check-*` files.  This is a little bit complicated because the `-`
in the name and the lack of `.py` means that they're not recognized as
normal python modules, so we need to do some hacks (similar to the ones
we do in test/common/run-tests).

With this in place, it's now possible to run the integration test suite
via pytest with something like the following:

  $ test/common/pixel-tests pull
  $ test/image-prepare
  $ test/common/pywrap -m pytest test/verify -n4

Note: the -n option is for specifying the level of parallelism and
requires the python3-pytest-xdist plugin to be installed.

We adjust our vulture config a bit to teach it about pytest plugins.