testlib: tweak label calculation for pixel tests
The current calculation in label() is trying to drop the module name,
keeping the class and function name.  Up until now, the module name has
always contained one component (in fact, it's always '__main__') so the
code has worked by dropping the first component.

We can do this a lot easier: we already have other code that's using
self._testMethodName to calculate a label, so do the same thing here.