test: remove mypy crash workaround
This got fixed in allisonkarlitskaya/systemd_ctypes#58 and merged as
part of bc72007094204f24174057c7904f45dec8efab6d.

Enabling the caching makes the linting no longer take the majority of
time of a test run.

The comment about wanting to exclude systemd_ctypes is also no longer
valid: it now lives inside of src/cockpit/_vendor/ anyway, and is a lot
further along with typing hints at this point.