pybridge: Fix ssh-agent cleanup for setgid
Some distributions install ssh-agent as setgid, to protect against
ptracing and malicious environment variables. Unfortunately this also
clears PDEATHSIG, causing a leak of the agent. (This does not affect the
one from cockpit-session, as that is launched via PAM).

Replace this with running ssh-agent for a shell process which will exit
on EOF on stdin when the parent bridge process dies.

This fixes cleaning up ssh-agent in Cockpit client and
testCockpitDesktop. However, bump the timeout, as it usually takes
around 8 seconds for the process to actually go away (it apparently
regularly polls its parent).