test: Add cockpit-client browser integration test
The flatpak Client's webkit browser is really hard to test, as it is so
inaccessible. Introduce test/verify/check-client which reproduces the
cockpit-{ws,client-ssh,beiboot} plumbing that cockpit-client does, and
exercise a lot of login scenarios to a machine with and without
cockpit-bridge.

This finds several minor bugs, mark them with FIXME and disable the
check or do a workaround.