test: Don't unmount restored directories lazily
If there is a busy file, this leads to the cleanup silently succeeding,
and the subsequent test then failing with "directory already exists" or
similar errors from trying to recursively mount/restore.

This was introduced into the initial restore_dir() implementation in
commit 456c8001820fb (later generalized in commit ba707dacd70721 without
particular justification).

With this change, the cleanup of the affected test now fails directly,
which is easier to debug.

The main exception is restoring /home/admin -- many tests do that
without stopping all admin processes first. This happens in
nonDestructiveSetup(), but that runs after restoring the home dir. This
is tricky to fix, so for now add an exception to continue unmounting
that lazily.