test: Split off TestServices.testFilter
It's rather complex and long, and does not have a lot of dependencies to
the other bits in testBasic, other than the presence of
test{,-fail}.service. Split it out, to clean up testBasic a bit.

This also is pure UI functionality, it does not need to run twice for
session and user.

Rename the pixel test references accordingly.