systemd: Fix journalctl follow invocation race
When the filter gets changed while the initial non-following journalctl
runs, that runs updateQuery with new options. That will first stop all
previous journalctls. However, that will still call the `.done()`
handler for that initial journalctl query, and happily go on with
starting the subsequent journalctl in following mode. This results in
two journalctls running at the same time, leading to duplicated entries.

To fix that, add a `stopped` flag to the returned journalctl object, and
don't start the following journalctl if it's set.

Fixes #18976