pyproject.toml: copy systemd_ctypes tox setup
Copy the same sort of tox setup that systemd_ctypes is using, with some
minor adjustments to testing dependencies and the like.

The biggest change here is that instead of running each linter tool as a
separate command, we run them all together via test/static-code.  The
main reason for that is that we have a lot of stand-alone executable
Python scripts (chmod +x, and not ending in .py) that we also want to
lint, and the linters don't know how to find those on their own.

As in systemd_ctypes, typing `tox p` now gives a fairly nice experience.

Note also: it's now possible to do a linter run with tools installed
from PyPI using something like:

    tox -e py311-lint

It's also possible to run pytest with older version of Python:

    pytest -e py36-pytest

It's also possible to run all of the venv-based environments like:

    tox p -m venv