test_transport.py: port to pytest
This was our last remaining user of unittest.TestCase in test/pytest/.
The unittest.IsolatedAsyncioTestCase API was added in Python 3.8, which
clears one more hurdle for running unit tests against older Python
versions.

We continue to use unittest.mock instead of pytest-mock: it does what we
need, causes us no problems, and is available in the standard library
(instead of being an extra dependency that we'd need to install).