test: Fix unintentional ignoring of console errors
Commit cae883461c354 broke the checking of browser errors, as with an
unset `$TEST_ALLOW_JOURNAL_MESSAGES` (which is the case in our CI) it
was adding the empty string as allowed pattern, which always matches.

Ignore the React errors that happen with our current code,  and add some
explanations. We have so many of them that we can't fix all of them in
one go, but we have to put a stop to introducing new ones. Some of them
are test specific.

Co-Authored-By: Martin Pitt <mpitt@redhat.com>