metrics: Disable on-demand cockpit-pcp installation with beibooted bridge
Installing cockpit-pcp on a system without any cockpit packages
currently wreaks havoc: It pulls in cockpit-bridge as a dependency, and
thus on the next Cockpit log in it will switch from "beibooted bridge"
to "local bridge". But there are no actual Cockpit packages installed.

As a stopgap, disable cockpit-pcp on-demand installation with beiboot,
and just show a hint. Detecting whether the bridge is beibooted is not actually
that easy; add a hack that checks the running bridge's cmdline for the time
being.

This will keep the "beiboot" and "distro packages" worlds cleanly separated
until we design this more properly, see https://issues.redhat.com/browse/COCKPIT-942
After that, the above hack can be dropped again.

This is the only place where this happens: All other PackageKit
invocations are either on-demand installs that don't directly affect
cockpit (like realmd, vdo, dnf-automatic, or kpatch),  or are on the
Software Updates page (which is an explicitly supported use case).

Fixes #19143