Improve JSON deconstruction static typing
Add a family of typed dictionary accessor functions to our jsonutil
file, with names like get_bool(), get_int(), get_strv() and so on.

Those will throw JsonError if the typing is wrong or a value is missing.

At the core of these functions is a new typechecked() function which
performs a checked cast to a given value type, plus a _get() function
which accepts a cast function as a argument.  Everything is correctly
statically typed through liberal use of type variables.

Use these throughout the code, allowing us to remove many assertions and
type:ignore comments.  This is a big chunk of the cases present, but
there are still some corners that need taking care of.

Catch the JsonError in various places for appropriate error reporting.

Having _get() take a cast function as an argument allows for another
interesting possibility: converting certain JSON types to other object
types.

Building on that, this patch also changes our approach to parsing
manifest files: we now do a lot more of the parsing up-front, storing
the results into properly-typed Python objects (BridgeConfig, Condition,
etc) so that we can detect errors and throw them at the time we parse
the file.  We also do requirements version checking up-front, allowing
for better error reporting.

We also do a similar (but much more trivial) change for parsing the
rows/cols fields of window size requests for pty-enabled stream
channels.