test: Update TestNetworkingBasic pixel test with shorter MAC
On the most recent image refresh [1] the rendered MAC string becomes
shorter, which makes the string from the reference stick out of the
ignored area, and thus cause a pixel diff. Use the new reference with
the shorter string, so that it works with both the old and the new
image.

[1] https://github.com/cockpit-project/bots/pull/5189