networking: don't pass null to input elements
changing the `value` attribute of React input elements to null changes
it from controlled to uncontrolled, which leads to React logging errors
in the browser console.

This fixes errors of the form:

Warning: A component is changing a controlled input to be uncontrolled. This is likely caused by the value changing from a defined to undefined, which should not happen...