test: Wait after layout change for some metrics pixels
Previously, we were only waiting after switching to mobile, but the
layout changes seem to be problematic in general for the metrics
tests.

So let's just add a general mechanism for waiting a bit after a layout
change and use it for those tests.

This will hopefully allow us to not skip "mobile" and "rtl" with the
second pixel test.