test: Fix restore_dir() unit restart race
firewalld writes its configuration files on shutdown. Previously, the
test cleanup first restored the files and *then* restarted firewalld,
which sometimes leaked changes from the test into the persistent config.
That led to subsequent tests timing out to load the login page, as the
`cockpit` service was missing in firewalld.

This is a potential race condition for all service restarts, so fix it
in a general way: Introduce a new `restart_unit` argument to
`restore_dir()`, which first stops the unit, then restores the files,
and then starts the unit again if it was running before. Convert all
affected tests to the new API.

Cherry-picked from main commit ce07ea7372c