ws: do shell splitting on auth command
Currently the auth command is interpreted as a literal string pointing
to the binary.  That means commands like `python3 -m cockpit.beiboot`
are not possible, but also means things like trailing whitespace are
taken to be part of the filename.

Use the standard shell splitting provided to us by GLib to give a nicer
interface here.