cockpit.remote: minor adjustments to open handling
Use our type-checking json accessors to inspect open messages when
routing to remote peers.

Slightly modify the handling of the password field: unlike all of the
other fields which are unpacked at the routing-rule level (since they
are required for making routing decisions) leave the password to be
unpacked by the SshPeer itself, which we now pass the open message to.

This change is trivial, but it will be helpful in coming commits to
avoid an explosion in arguments as we pull more information out of that
message.

Cherry-picked from main commit 22f89b2a4f8ead2dc5.