networking: fix page crash when wg properties go missing
Though the original issue #19384 used wg-quick to configure wireguard, I
was unable to reproduce the issue with it, unless I created a brand new
connection which seems unnecessary as the tests already had a profile
configured. Using `sed` to edit wg keyfile and reproduce should be
equivalent.

Other wireguard properties `endpoint` and `listen-port` are either
optional or have a default value, so the /network page shouldn't break
if they're not set. Missing or invalid `public-key` is equivalent to no
peer, so it shouldn't break /network page as well.

fixes #19384