selinux: Allow cockpit_ws_t to read /sys
On Fedora 39 we get this rejection:

  avc:  denied  { read } for comm="cockpit-askpass" name="possible" dev="sysfs"
  scontext=system_u:system_r:cockpit_ws_t:s0 tcontext=system_u:object_r:sysfs_t:s0

This also happens for "cockpit-beiboot". Our code doesn't directly
access sysfs (at least not /devices/system/cpu/possible), but glibc
does, and Python somehow inherits this.

There is no harm in letting it read sysfs, and it may even be necessary
for hardware based authentication schemas, so allow it.