test: Follow ASCII chart instead of browser's keycodes
Until now, we were mapping values of characters to their browser's
keycodes values [1]. This however is incorrect, since to get a value of
character, we use Python's 'ord' function, which returns character's
Unicode value.

In some cases, these values are different. For example, a value for
Enter has keycode '13' in browser [1], but Python's 'ord' will give us
value of '10' for a line feed character. [2]

[1] https://developer.mozilla.org/en-US/docs/Web/API/KeyboardEvent/keyCode
[2] https://python-reference.readthedocs.io/en/latest/docs/str/ASCII.html