kdump: RHEL/Fedora now defaults to setting crashkernel
In Fedora-39 and all supported RHEL versions (8 and 9) crashkernel is
now set in the default cmdline. RHEL also defaults in enabling
kdump.service which Fedora does not.

This simplifies our testing as we don't have to re-generate initrd and
reboot. Plus this means we test the distribution default which is a lot
more useful then an arbitrary setting we came up with.