storage: Always resize filesystems and LUKS to the maximum possible
This is always the right thing to do, since Cockpit will first resize
the block device (which is a logical volume or partition) and then
resize the content to match.  This will cover the case where the block
device resizing ended up with something different from what we asked
for.