storage: Put the minimum of a slider at the far left
Instead of always putting zero at the far left.

The previous code was written for a older version of the slider that
could be restricted, and would not allow the user to slide below the
minimum. Patternfly sliders can not be restricted in this way and will
allow the user to slide the knob into every position. Thus, we need to
make the far left the minimum to avoid confusion.

We also need to adapt the step size, so that the cursor keys actually
move the slider noticeable.