fmf: Fix prioritization of our main-builds COPR
Commit a6dd74a59c was not sufficient: The problem wasn't the tag
repository, but that current Fedora rawhide has a 303-2 rebuild of
cockpit which is formally newer than our 303-1.<git stamp> revisions
from our main-builds COPR.

If the env has a main-builds COPR (for reverse dependency tests), we
always want to install the latest version from that. So give it the
highest priority (uninutitively, lower priority wins), and use
`distro-sync` to up- or downgrade the packages as necessary.