router: explicitly track endpoints
The router currently keeps a mapping of open channels to the endpoints
responsible for them.  This presents two problems:

 - when we close down an endpoint, we need to iterate all open channels
   in order to determine which channels belong to that endpoint
 - it's possible to have active endpoints associated with the router
   which the router has no idea about

Move to a more explicit model where we add a second mapping: endpoints
to their set of open channels.  This makes endpoint shutdown easier and
adds the advantage that an endpoint with no channels can still be
tracked by the router (with an empty channel list).

The second point of this will be useful in future commits when the
router (and not the routing rules) become responsible for ensuring that
all endpoints are correctly shutdown.

We need to adjust the beiboot code a bit to avoid a catch22 whereby the
routing rule needs to exist before the router is initialized, but the
Peer itself can only now exist after the router is initialized.