test: Reduce error message expectation for unreachable tang server
stratis 3.6.0 changed its implementation from calling the cryptsetup CLI
to the API. This made the error message for an unreachable tang server
rather useless:

> Cryptsetup error: IO error occurred: Operation not permitted (os error 1)"

That is reported in https://bugzilla.redhat.com/show_bug.cgi?id=2246920
but fixing this will take a while. In the meantime, don't skip the rest
of the test via the naughty, but just adjust to current reality.

After that we can drop https://github.com/cockpit-project/bots/issues/5469