test: fix more test_spawn_broken_pipe races
On old Python versions, we can get the process-exited callback from the
child watcher at the time of registration, without a return to the
mainloop.  That means that for very fast-exiting processes, we might
never get a chance to write to them to observe a EPIPE error.

Add a synchronization point so that we can prevent the process from
exiting until after we've finished spawning it (and registering the
watch).

Closes #19586