test: Drop sssd fiddling in testClientCertAuthentication
Restarting sssd that often causes state corruption, as it often cannot
initialize in 5s. It's also too much fiddling with the OS -- joining a
domain should make the users available automatically, otherwise this is
a bug.

This works fine with IPA, and doesn't regess AD either.
testUnqualifiedUsers() already does it that way, too.