storaged: add btrfs volume integration
Btrfs is a filesystem which can span over multiple devices a bit
similar to LVM but not 100%. On a standard Fedora server image we have
one partition which holds a btrfs filesystem with two subvolumes for /
and /home. In Cockpit we represent this as:

-> btrfs subvolume
    -> btrfs volume (the filesystem)
        -> btrfs device (backing storage)
	   -> block device

A btrfs volume can have one or more devices and supports multiple raid
configurations. In Cockpit we show a btrfs volume with the devices it is
uses a backing and the space usage.

In UDisks a "volume" does not exist but only an
org.freedesktop.UDisks2.Filesystem.BTRFS object which per btrfs device,
setting the label or retrieving the subvolumes on one of these objects
whcih are part of the same volume affects the whole group.

This initial commit lacks support for Btrfs subvolumes.

The UDisks API lacks a few data points which would be useful for in
Cockpit such as the space usage per btrfs device, showing the health
status of a btrfs filesystem when one device is missing, balancing of
btrfs devices. The space use per btrfs device has been implemented using
btrfs-progs in Cockpit, health warnings are not yet shown in Cockpit as
it for now purely focuses on a single btrfs device volume.