storaged: btrfs: don't break on a btrfs mount without subvol
We always assume fstab_subvol to be truthy while when fstab contains a
btrfs mount without a subvol option it is not and we should just show
the root subvolume.

In an ideal case we would call GetDefaultSubvolume when the filesystem
is mounted and then show the subvolume name instead. I'll leave this for
a follow up.