ws: Support Content-Length in the fsread1 channel
Cockpit navigator and SOS report offer a file download option,
downloading a file in the browser shows the download speed but no
estimation of how long it takes. This is due to the lack of the
Content-Length header which could be added via the external channel
options however the `fsread1` channel already has all the information.

The `fsread1` channel now emits a `size-hint` in the ready message which
cockpit-ws picks up and adds a Content-Length header. The header is only
included for the binary channel mode.

Closes: #19861