bridge: clean up kill message forwarding
Since be9ed5938cceea5748fa1dc3656f862217d04f97 we have the ability to
send verbatim JSON objects as control messages, which we do when
forwarding channel control messages to peers.  Until now, we haven't
applied the same handling for 'kill' messages because we haven't had
access to the message object in order to be able to forward it.

This means that we've been unintentionally breaking protocol: the
current way we construct the control message from kwargs, we always send
up sending both "host" and "group", even if one of them is `null`, which
is a violation of protocol (these should either be strings, or absent).

Let's wire the original message through and send it verbatim: this will
help us in the next commit when we start enforcing type safety on kill
message parameters.