test: add device is in use test for formatting btrfs
When we have a multi device setup we want to not allow formatting of
btrfs devices as this breaks multi disk setups. We had a test for this
but it was only in testBasic, since we do allow formatting of a single
btrfs filesystem device now the test was wrongly removed and should have
been moved to testMultiDevice.